refactor(resolving-deps-resolver): split resolve_dependency_tree into focused modules
resolve_dependency_tree.rs carried the public options/error types, the shared
workspace context, the per-importer tree context, the recursive walk, node
reuse and children-ownership bookkeeping, lockfile-snapshot reuse checks,
catalog specifier resolution, and manifest extraction in one 4,037-line file.

Keep resolve_dependency_tree.rs as the facade holding the options, error, and
notification types plus the two entry points, and move the implementation into
resolve_dependency_tree/: workspace_ctx.rs (WorkspaceTreeCtx, the shared
per-pkgIdWithPatchHash dedup maps, and the children-ownership claim/handover),
tree_ctx.rs (TreeCtx and the per-edge ResolveOptions derivations), walk.rs (the
fresh-resolve walk — seed, children walk, per-wanted resolve cache, level
folds), reuse.rs (ReuseSource, the update scope, direct-dep change bookkeeping,
subtree-reuse predicates, and the snapshot-driven walk), manifest.rs (what the
walk reads off a resolved package: pkg id, child specs, peers, leaf
classification, deprecation), and catalogs.rs (catalog: specifier resolution).

TreeCtx and WorkspaceTreeCtx stay one struct each; their inherent impls are
divided across the modules with pub(super) visibility, so every cross-module
item stays private to resolve_dependency_tree and the crate's public API is
unchanged. The pub(super) surface was derived mechanically: every marker was
stripped and the compiler re-added only the ones the crate cannot compile
without. The unit tests that drive one module's internals move next to it, with
the shared result builder lifted into a test_support module; the end-to-end
tree-walk tests stay at the facade.

Pure code movement: no behavior, allocation, cloning, dispatch, or traversal
change. After normalizing away imports, module declarations, visibility
markers, and reference-style doc links, the multiset of source lines before and
after differs only by the eleven signatures rustfmt re-wrapped because
pub(super) pushed them past the width.

Closes pnpm/pnpm#13554.
