+++ pnpm/crates/resolving-deps-resolver/src/resolve_dependency_tree.rs
@@ -1,28 +1,57 @@
-use async_recursion::async_recursion;
-use chrono::{DateTime, Utc};
+//! Walks a project's dependency graph, resolving every edge and
+//! recording one occurrence node per edge.
+//!
+//! This file is the facade: the entry points
+//! ([`fn@resolve_dependency_tree`], [`extend_tree`]) plus the options,
+//! errors, and notification payloads they carry. The walk itself is
+//! split across [`workspace_ctx`] (the workspace-shared dedup maps and
+//! children ownership), [`tree_ctx`] (the per-importer context),
+//! [`walk`] (the fresh-resolve walk), [`reuse`] (prior-lockfile reuse),
+//! [`manifest`] (what the walk reads off a resolved package), and
+//! [`catalogs`].
+
 use derive_more::{Display, Error};
 use futures_util::future;
 use miette::Diagnostic;
-use pacquet_catalogs_resolver::{
-    CatalogResolutionError, CatalogResolutionResult, WantedDependency as CatalogWantedDependency,
-    resolve_from_catalog,
-};
+use pacquet_catalogs_resolver::CatalogResolutionError;
 use pacquet_catalogs_types::Catalogs;
 use pacquet_hooks::PnpmfileHooks;
-use pacquet_package_manifest::{DependencyGroup, PackageManifest, engines_runtime_dependencies};
-use pacquet_patching::{PatchGroupRecord, PatchKeyConflictError, get_patch_info};
+use pacquet_package_manifest::{DependencyGroup, PackageManifest};
+use pacquet_patching::{PatchGroupRecord, PatchKeyConflictError};
 use pacquet_resolving_resolver_base::{
-    CurrentPkg, NoMatchingVersionError, PreferredVersionsOverlay, RegistryResponseError,
-    ResolveError, ResolveOptions, Resolver, WantedDependency,
+    NoMatchingVersionError, PreferredVersionsOverlay, RegistryResponseError, ResolveOptions,
+    Resolver, WantedDependency,
 };
 use pipe_trait::Pipe;
 use rustc_hash::{FxHashMap as HashMap, FxHashSet as HashSet};
 use serde_json::Value;
-use std::{
-    borrow::Cow,
-    collections::BTreeMap,
-    path::{Path, PathBuf},
-    sync::{Arc, Mutex, MutexGuard},
+use std::sync::{Arc, Mutex, MutexGuard};
+
+use crate::{
+    parent_pkg_aliases::ParentPkgAliases,
+    resolved_tree::{DirectDep, ResolvedTree},
+};
+
+mod catalogs;
+mod manifest;
+mod reuse;
+mod tree_ctx;
+mod walk;
+mod workspace_ctx;
+
+#[cfg(test)]
+mod test_support;
+
+pub use tree_ctx::TreeCtx;
+pub use workspace_ctx::WorkspaceTreeCtx;
+
+pub(crate) use catalogs::resolve_catalog_specifiers;
+pub(crate) use reuse::{record_changed_direct_deps, unwrap_package_name};
+
+use reuse::{ReuseSource, record_direct_dep_versions};
+use walk::{
+    NodeSeed, level_aliases, level_versions, resolve_node_seed, walk_node_children,
+    warm_children_resolutions,
 };
 
 /// Acquire a [`Mutex`] guard, recovering from poisoning the same way
@@ -35,20 +64,6 @@ use std::{
 fn lock_recoverable<Inner>(mutex: &Mutex<Inner>) -> MutexGuard<'_, Inner> {
     mutex.lock().unwrap_or_else(std::sync::PoisonError::into_inner)
 }
-
-use crate::{
-    lockfile_reuse::{
-        current_pkg_from_lockfile, prior_child_key, reusable_importer_dep, synthesize_reused_result,
-    },
-    node_id::NodeId,
-    parent_pkg_aliases::{ParentPkgAliases, peer_shadowed_dependencies},
-    resolved_tree::{
-        AncestorIds, DependenciesTreeNode, DirectDep, PeerDep, ResolvedPackage, ResolvedTree,
-    },
-};
-use pacquet_lockfile::{
-    PkgName, PkgNameVerPeer, ProjectSnapshot, ResolvedDependencyMap, SnapshotDepRef, SnapshotEntry,
-};
 
 /// Which dependencies `pacquet update` excludes from lockfile-resolution
 /// reuse. An excluded package re-resolves to highest-in-range, and its
@@ -99,66 +114,6 @@ impl UpdateDepth {
             None => 0,
             Some(max_depth) => depth.min(max_depth.saturating_add(1)),
         }
-    }
-}
-
-/// The `pacquet update` scope a node is judged against: which names the
-/// update targets, and how deep it reaches.
-#[derive(Debug, Clone, Copy)]
-struct UpdateScope<'a> {
-    reuse: &'a UpdateReuseScope,
-    max_depth: UpdateDepth,
-}
-
-/// How the current [`fn@resolve_node`] call may reuse the prior
-/// lockfile's resolution instead of re-resolving from the registry.
-///
-/// Threaded down the recursion to drive the `resolvedDependencies` /
-/// `parentPkg.updated` reuse mechanism: an updated parent discards its
-/// child-refs so the subtree re-resolves, while a non-updated parent
-/// keeps them alive.
-#[derive(Clone)]
-enum ReuseSource {
-    /// A direct dependency of importer `importer_id`. Reuse matches the
-    /// manifest specifier against the importer's recorded resolution via
-    /// semver-satisfies ([`reusable_importer_dep`]).
-    Importer { importer_id: String },
-    /// A transitive dependency whose resolved snapshot key the parent's
-    /// snapshot already pins. `Some` reuses that key directly (no semver
-    /// check — the parent version pins it); `None` means an updated
-    /// ancestor discarded its child-refs, forcing this subtree to
-    /// re-resolve. The parent may itself be reused (snapshot walk) or
-    /// freshly resolved onto its previously recorded version — both
-    /// keep the prior child refs alive, mirroring pnpm's
-    /// `parentPkg.updated ? undefined : resolvedDependencies`.
-    Transitive { key: Option<PkgNameVerPeer> },
-    /// Reuse disabled for this node (no prior lockfile).
-    Off,
-}
-
-impl ReuseSource {
-    /// The prior lockfile snapshot key recorded for this edge, if any —
-    /// the basis of both subtree reuse and the `currentPkg` payload.
-    /// The `Importer` arm applies the semver-satisfies gate
-    /// ([`reusable_importer_dep`]): a lockfile reference is only reused
-    /// when the recorded version still satisfies the wanted spec.
-    fn prior_key(&self, ctx: &TreeCtx, wanted: &WantedDependency) -> Option<PkgNameVerPeer> {
-        let lockfile = ctx.workspace.wanted_lockfile.as_ref()?;
-        match self {
-            ReuseSource::Importer { importer_id } => reusable_importer_dep(
-                lockfile,
-                importer_id,
-                wanted.alias.as_deref()?,
-                wanted.bare_specifier.as_deref()?,
-            ),
-            ReuseSource::Transitive { key } => key.clone(),
-            ReuseSource::Off => None,
-        }
-    }
-
-    /// Whether this edge may reuse the prior lockfile's subtree.
-    fn allows_reuse(&self) -> bool {
-        matches!(self, ReuseSource::Importer { .. } | ReuseSource::Transitive { .. })
     }
 }
 
@@ -388,6 +343,9 @@ impl From<PatchKeyConflictError> for ResolveDependencyTreeError {
 /// `await` inside, so a sync mutex is the right tool — tokio's async
 /// mutex adds per-acquire overhead that the resolve hot path was
 /// paying once per visit per ctx field.
+///
+/// [`ResolvedPackage`]: crate::ResolvedPackage
+/// [`DependenciesTreeNode`]: crate::DependenciesTreeNode
 pub async fn resolve_dependency_tree<DependencyGroupList, Chain>(
     resolver: &Chain,
     manifest: &PackageManifest,
@@ -531,142 +489,6 @@ where
     resolve_catalog_specifiers(wanted, catalogs)
 }
 
-/// Cache key for [`WorkspaceTreeCtx`]'s `resolved_by_wanted` map.
-///
-/// The npm-shaped slice pacquet exposes today calls
-/// [`Resolver::resolve`] with four [`WantedDependency`] fields
-/// populated — `alias`, `bare_specifier`, `optional`, and `injected` (see
-/// the `WantedDependency` literals in [`extend_tree`] and the recursive
-/// arm of [`fn@resolve_node`]). Anything else stays at `Default::default()`,
-/// so a tuple over those four fields uniquely identifies a wanted
-/// dep across revisits.
-///
-/// `optional` is part of the key because the npm resolver's
-/// `pick_package` toggles between the abbreviated and full packument
-/// based on `wanted.optional` — caching by `(alias, bare_specifier)`
-/// alone would let an optional caller satisfy itself with a
-/// non-optional caller's abbreviated result, losing the
-/// `libc`/`cpu`/`os` filter inputs that mode supplies.
-///
-/// `injected` is part of the key because the workspace branch of the
-/// npm resolver returns a `file:<path>` resolution when the dep is
-/// injected and a `link:<path>` resolution otherwise (see
-/// `resolve_from_local_package`). Two importers asking for the same
-/// workspace dep with different `dependenciesMeta[*].injected` flags
-/// must take different cache slots.
-///
-/// `pick_lowest_version` and `published_by` are part of the key because
-/// `resolutionMode` makes the version pick depend on them: under
-/// `time-based` / `lowest-direct` a direct dependency is resolved
-/// lowest while a transitive one is resolved highest, and under
-/// `time-based` transitive deps carry a publish-date cutoff a direct
-/// dep does not. The same wanted spec (`react@^18`) can therefore
-/// resolve to a different version as a direct vs. transitive dep, so
-/// the two occurrences must take different cache slots. In `highest`
-/// mode (the default) every occurrence shares the same pair, so the
-/// dedup is unchanged.
-///
-/// `project_dir` is part of the key for any specifier that can produce
-/// a project-relative resolution. This includes explicit local
-/// specifiers (`link:` / `file:` / `workspace:`) and normal semver
-/// specifiers in workspace mode, because `linkWorkspacePackages` can
-/// replace the registry pick with a workspace package. A non-injected
-/// workspace dep resolves through `resolve_from_local_package` to a
-/// `link:<path>` whose `<path>` is computed *relative to the
-/// consuming importer's directory*. Without `project_dir` in the key,
-/// the first importer to resolve `(@scope/lib, ^1.0.0)` would
-/// seed the workspace-wide cache with its own relative path and every
-/// other importer would reuse it verbatim — e.g. a root resolving to
-/// `link:packages/lib` would hand `packages/app` the same string,
-/// which from `packages/app` points at the non-existent
-/// `packages/app/packages/lib`.
-///
-/// The final two fields isolate importers with active update policies
-/// and record whether this wanted dependency is an explicit update target.
-/// Ordinary keep-all importers use no importer scope and retain the existing
-/// cross-importer cache sharing.
-type WantedKey = (
-    Option<String>,
-    Option<String>,
-    Option<bool>,
-    Option<bool>,
-    bool,
-    Option<DateTime<Utc>>,
-    Option<PathBuf>,
-    Option<PkgNameVerPeer>,
-    Vec<(String, Vec<String>)>,
-    Option<String>,
-    bool,
-);
-
-type SubtreeReuseKey = (Option<String>, PkgNameVerPeer, i32);
-
-/// Whether a wanted dep's resolution is computed relative to the
-/// consuming importer's directory rather than being
-/// importer-independent. True for the `link:` / `file:` / `workspace:`
-/// protocols, whose resolved path `resolve_from_local_package` derives
-/// from `project_dir`. Such resolutions must not be shared across
-/// importers in [`WantedKey`].
-fn project_relative_cache_scope(
-    wanted: &WantedDependency,
-    opts: &ResolveOptions,
-) -> Option<PathBuf> {
-    (wanted.bare_specifier.as_deref().is_some_and(|spec| {
-        spec.starts_with("link:") || spec.starts_with("file:") || spec.starts_with("workspace:")
-    }) || (opts.always_try_workspace_packages && opts.workspace_packages.is_some()))
-    .then(|| opts.project_dir.clone())
-}
-
-/// The directory the `file:` dependencies declared by a resolved
-/// package resolve against — pnpm's `parentPkg.rootDir`. A package
-/// copied from a local directory declares them relative to its own
-/// directory; every other resolution has no directory of its own, so
-/// its `file:` children stay on the consuming importer's.
-///
-/// Only `file:`-shaped directory resolutions qualify. They record
-/// their directory relative to the lockfile root (absolute under
-/// `preserveAbsolutePaths`), whereas a `link:`-shaped one records it
-/// relative to the consuming importer (workspace links) or absolute
-/// (the local resolver). Nothing asks for a linked node's directory:
-/// a linked project resolves its own dependencies as a separate
-/// importer, so the walk never descends into one.
-fn declaring_manifest_dir(
-    ctx: &TreeCtx,
-    result: &pacquet_resolving_resolver_base::ResolveResult,
-) -> Option<Arc<Path>> {
-    let pacquet_lockfile::LockfileResolution::Directory(resolution) = &result.resolution else {
-        return None;
-    };
-    if !result.id.as_str().starts_with("file:") {
-        return None;
-    }
-    let directory = Path::new(&resolution.directory);
-    let absolute = if directory.is_absolute() {
-        pacquet_fs::lexical_normalize(directory)
-    } else {
-        pacquet_fs::lexical_normalize(&ctx.lockfile_dir.join(directory))
-    };
-    Some(Arc::from(absolute))
-}
-
-/// Point a `file:` specifier at the directory of the manifest that
-/// declares it ([`declaring_manifest_dir`]) instead of the consuming
-/// importer's. Every other specifier keeps the caller's options.
-fn opts_relative_to_declaring_manifest<'a>(
-    opts: &'a ResolveOptions,
-    wanted: &WantedDependency,
-    parent_dir: Option<&Path>,
-) -> Cow<'a, ResolveOptions> {
-    match parent_dir {
-        Some(parent_dir)
-            if wanted.bare_specifier.as_deref().is_some_and(|spec| spec.starts_with("file:")) =>
-        {
-            Cow::Owned(ResolveOptions { project_dir: parent_dir.to_path_buf(), ..opts.clone() })
-        }
-        _ => Cow::Borrowed(opts),
-    }
-}
-
 /// One spec carried through [`extend_tree`] and the importer-side
 /// orchestrator: `(alias, range, optional, injected)`. `injected`
 /// reflects the importer manifest's `dependenciesMeta[alias].injected`
@@ -678,1005 +500,6 @@ fn opts_relative_to_declaring_manifest<'a>(
 /// `false` — peers picked up via auto-install don't carry per-dep
 /// meta from any manifest.
 pub(crate) type WantedSpec = (String, String, bool, bool);
-
-/// An importer's resolved direct-dependency versions, keyed by package
-/// name. See [`WorkspaceTreeCtx::direct_dep_versions`].
-type DirectDepVersions = HashMap<String, Vec<node_semver::Version>>;
-
-/// One entry in [`WorkspaceTreeCtx`]'s `children_specs_by_id` map —
-/// `(child_alias, child_range, child_optional)` triples extracted from
-/// a resolved package's manifest's `dependencies` plus
-/// `optionalDependencies` sections.
-type ChildSpec = (String, String, bool);
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct ChildrenOwner {
-    update_active: bool,
-    depth: i32,
-    importer_order: usize,
-    parent_path: Vec<String>,
-    importer_id: String,
-}
-
-impl ChildrenOwner {
-    fn wins_over(&self, other: &Self) -> bool {
-        if self.update_active != other.update_active {
-            return self.update_active;
-        }
-        (&self.depth, &self.importer_order, &self.parent_path)
-            < (&other.depth, &other.importer_order, &other.parent_path)
-    }
-}
-
-/// What the current children owner of a `pkgIdWithPatchHash` decided
-/// for it. Both halves are per-occurrence inputs the walk has to settle
-/// on one answer for, so they are claimed together, under one lock.
-#[derive(Debug, Clone)]
-struct ChildrenOwnerEntry {
-    owner: ChildrenOwner,
-    /// The owner occurrence's peer-shadowed `dependencies` (see
-    /// [`peer_shadowed_dependencies`]). Which names are shadowed
-    /// depends on the parent scope, which differs per occurrence;
-    /// pnpm lets the first occurrence to resolve decide, which is
-    /// arrival-ordered. The deterministic children owner decides here
-    /// instead, so the same graph always yields the same lockfile.
-    peer_shadowed: Arc<HashSet<String>>,
-}
-
-/// Workspace-shared maps. Every per-importer [`TreeCtx`] in a
-/// multi-importer install holds an `Arc<WorkspaceTreeCtx>` so the
-/// resolver's per-`pkgIdWithPatchHash` dedup (`packages`,
-/// `children_specs_by_id`, `children_by_id`, `resolved_by_wanted`) and
-/// the peer-walker's seed sets (`all_peer_dep_names`,
-/// `applied_patches`, `policy_violations`) carry across importers. One
-/// shared context is handed to every importer's hoist loop.
-///
-/// `dependencies_tree` (`NodeId → DependenciesTreeNode`) is keyed by
-/// per-occurrence `NodeIds`, which are unique even across importers, so
-/// every importer's walk contributes entries to one combined tree
-/// without colliding.
-pub struct WorkspaceTreeCtx {
-    /// Bumped whenever an [`fn@extend_tree`] call may mutate the shared
-    /// maps. The peer-hoist discovery engine compares it against the
-    /// revision of its last sync to skip re-syncing an unchanged
-    /// context.
-    revision: std::sync::atomic::AtomicU64,
-    /// Bumped whenever a children-ownership change rewrites an existing
-    /// occurrence node's children in `dependencies_tree` (see
-    /// [`fn@make_non_owner_nodes_lazy`]). Such a rewrite invalidates
-    /// walk state derived from the previous children, so the discovery
-    /// engine rebuilds its view instead of merging when this advanced
-    /// since its last sync.
-    children_rewrites: std::sync::atomic::AtomicU64,
-    packages: Mutex<HashMap<String, ResolvedPackage>>,
-    /// Every `name → version` this run resolved (one fold per new
-    /// `packages` entry carrying a `name_ver`, plus every workspace
-    /// package version a directory resolution surfaced), shaped as the
-    /// plain [`pacquet_resolving_resolver_base::PreferredVersions`]
-    /// entries the peer-hoist pickers bias toward. Maintained once,
-    /// workspace-wide: importers query the buckets they need through
-    /// [`TreeCtx::preferred_versions_for_names`] instead of each
-    /// replaying the whole run history into a private copy.
-    preferred_versions_from_run: Mutex<pacquet_resolving_resolver_base::PreferredVersions>,
-    dependencies_tree: Mutex<HashMap<NodeId, DependenciesTreeNode>>,
-    all_peer_dep_names: Mutex<HashSet<String>>,
-    policy_violations: Mutex<Vec<pacquet_resolving_resolver_base::ResolutionPolicyViolation>>,
-    applied_patches: Mutex<HashSet<String>>,
-    resolved_by_wanted:
-        Mutex<HashMap<WantedKey, Arc<pacquet_resolving_resolver_base::ResolveResult>>>,
-    children_specs_by_id: Mutex<HashMap<String, Arc<Vec<ChildSpec>>>>,
-    children_by_id: Mutex<HashMap<String, Arc<Vec<crate::resolved_tree::ChildEdge>>>>,
-    children_owner_by_id: Mutex<HashMap<String, ChildrenOwnerEntry>>,
-    node_parent_ids_by_id: Mutex<HashMap<NodeId, Arc<Vec<String>>>>,
-    /// Reverse index over `dependencies_tree`: every occurrence node
-    /// recorded for a `pkgIdWithPatchHash`. Keeps
-    /// [`fn@make_non_owner_nodes_lazy`] proportional to the package's
-    /// own occurrences — scanning the whole tree per recorded package
-    /// made lockfile-reuse walks quadratic in workspace size.
-    nodes_by_pkg_id: Mutex<HashMap<String, Vec<NodeId>>>,
-    manifest_hook: Option<ManifestHook>,
-    /// [`ManifestHook`] applied *after* [`Self::pnpmfile_hook`], where
-    /// `manifest_hook` runs before it. pnpm's `createReadPackageHook`
-    /// composes `packageExtensions → readPackage hooks → overrides`, so
-    /// overrides land here: a hook that replaces the manifest (e.g. an
-    /// embedder substituting a workspace project's raw manifest) must not
-    /// erase the overrides.
-    overrides_hook: Option<ManifestHook>,
-    /// The previous `pnpm-lock.yaml` the install started from, when one
-    /// exists. Consulted by [`resolve_node`] to reuse an already-resolved
-    /// dependency + its transitive subtree instead of re-resolving from
-    /// the registry (see `pnpm/plans/LOCKFILE_RESOLUTION_REUSE.md`).
-    /// `None` on a first install or when reuse is disabled.
-    wanted_lockfile: Option<Arc<pacquet_lockfile::Lockfile>>,
-    /// Lockfile-reuse suppression for `pacquet update`. `update`
-    /// re-resolves its target deps to highest-in-range, so a reused
-    /// resolution would defeat the bump. See [`UpdateReuseScope`].
-    update_reuse_scope: UpdateReuseScope,
-    /// Importer overrides used by filtered workspace updates. IDs absent from
-    /// this map keep the workspace default above.
-    update_reuse_scopes_by_importer: BTreeMap<String, UpdateReuseScope>,
-    /// `pacquet update --depth`: how deep the suppression above reaches.
-    update_depth: UpdateDepth,
-    /// Memoises [`fn@subtree_fully_reusable`] per update scope and snapshot
-    /// key. Keep-all importers share one scope; update-active importers use
-    /// isolated scopes so one importer's reuse answer cannot leak to another.
-    /// `true` means the package and its entire transitive subtree can be
-    /// synthesized from the prior lockfile.
-    subtree_reusable: Mutex<HashMap<SubtreeReuseKey, bool>>,
-    pnpmfile_hook: Option<Arc<dyn PnpmfileHooks>>,
-    /// `context.log(...)` sink for the `pnpmfile_hook`'s `readPackage`
-    /// calls, pre-bound to the install's reporter, project prefix, and
-    /// pnpmfile path. `None` leaves hook logging a no-op. See
-    /// [`WorkspaceTreeCtx::with_read_package_log`].
-    read_package_log: Option<pacquet_hooks::LogFn>,
-    /// Sink for skipped-optional-dependency notifications. `None`
-    /// keeps the skip behavior but drops the notification. See
-    /// [`SkippedOptionalLogFn`].
-    skipped_optional_log: Option<SkippedOptionalLogFn>,
-    /// The `pnpm.allowedDeprecatedVersions` map. See
-    /// [`crate::WorkspaceResolveOptions::allowed_deprecated_versions`].
-    allowed_deprecated_versions: BTreeMap<String, String>,
-    /// Sink for deprecation notifications. `None` keeps the
-    /// deprecation check but drops the notification. See
-    /// [`DeprecationLogFn`].
-    deprecation_log: Option<DeprecationLogFn>,
-    /// The install's `autoInstallPeers` setting. It widens which of a
-    /// resolved package's `dependencies` its own `peerDependencies`
-    /// shadow — see [`peer_shadowed_dependencies`].
-    auto_install_peers: bool,
-    /// Resolved registry map (`"default"` + per-scope) used to
-    /// materialize a prior `Registry` lockfile resolution back into its
-    /// tarball URL for the `currentPkg` payload. Empty when the entry
-    /// point doesn't thread registries (then `currentPkg` is withheld
-    /// for `Registry`-shaped entries rather than sent without a URL).
-    registries: std::collections::HashMap<String, String>,
-    /// `pkg id → importer id` of the importer whose occurrence owns
-    /// that package's shared children context. Ownership is chosen by
-    /// update-active status followed by `(depth, importer order, parent path)`:
-    /// a package's subtree is recorded once per id, and a non-owner occurrence
-    /// reuses the owner occurrence's children and missing-peer report. Consumed via
-    /// [`crate::HoistMissingScope`].
-    first_importer_by_pkg: Mutex<HashMap<String, String>>,
-    /// Per package: the missing-peer names reported by the *initial*
-    /// peer walk of the current children-owner generation, plus the
-    /// owner that recorded them (`None` while only a non-owner's
-    /// provisional walk has been seen). The record is written once per
-    /// generation: later hoist waves of the same owner never refresh it,
-    /// so a peer the owner only satisfied by hoisting stays visible to
-    /// every other importer's hoist. Consumed via
-    /// [`crate::HoistMissingScope`].
-    first_walk_missing_by_pkg: Mutex<HashMap<String, OwnerMissingRecord>>,
-    /// Per importer: direct-dep aliases whose manifest specifier differs
-    /// from the prior lockfile (new deps included). Gates the stale-pin
-    /// refresh's reuse-decline; only a changed direct dep can re-resolve
-    /// away from a transitive occurrence's pin. Keyed by importer: this
-    /// crate resolves importers sequentially (no workspace-wide
-    /// directs-before-transitives barrier), so a shared map would refresh
-    /// one importer's edges from another's direct deps order-dependently.
-    /// (pnpm has that barrier and so converges cross-importer; pacquet
-    /// stays per-importer to stay deterministic.)
-    changed_direct_deps: Mutex<HashMap<String, HashSet<PkgName>>>,
-    /// Per importer: the parsed resolved versions of its direct
-    /// dependencies, recorded once the direct-dep level finishes resolving.
-    /// The direct-dep versions folded into the children's
-    /// `preferredVersions`; consulted by [`fn@higher_direct_dep_version`].
-    /// `Arc` so the hot child walk snapshots the importer's map with one
-    /// lock + refcount bump instead of locking per edge.
-    direct_dep_versions: Mutex<HashMap<String, Arc<DirectDepVersions>>>,
-}
-
-/// One [`WorkspaceTreeCtx::first_walk_missing_by_pkg`] entry: the
-/// missing-peer names plus the owner generation that recorded them
-/// (`None` for a non-owner's provisional report).
-struct OwnerMissingRecord {
-    recorded_by: Option<ChildrenOwner>,
-    names: HashSet<String>,
-}
-
-impl Default for WorkspaceTreeCtx {
-    fn default() -> Self {
-        WorkspaceTreeCtx {
-            revision: std::sync::atomic::AtomicU64::new(0),
-            children_rewrites: std::sync::atomic::AtomicU64::new(0),
-            packages: Mutex::new(HashMap::default()),
-            preferred_versions_from_run: Mutex::new(
-                pacquet_resolving_resolver_base::PreferredVersions::new(),
-            ),
-            dependencies_tree: Mutex::new(HashMap::default()),
-            all_peer_dep_names: Mutex::new(HashSet::default()),
-            policy_violations: Mutex::new(Vec::new()),
-            applied_patches: Mutex::new(HashSet::default()),
-            resolved_by_wanted: Mutex::new(HashMap::default()),
-            children_specs_by_id: Mutex::new(HashMap::default()),
-            children_by_id: Mutex::new(HashMap::default()),
-            children_owner_by_id: Mutex::new(HashMap::default()),
-            node_parent_ids_by_id: Mutex::new(HashMap::default()),
-            nodes_by_pkg_id: Mutex::new(HashMap::default()),
-            manifest_hook: None,
-            overrides_hook: None,
-            wanted_lockfile: None,
-            update_reuse_scope: UpdateReuseScope::All,
-            update_reuse_scopes_by_importer: BTreeMap::new(),
-            update_depth: UpdateDepth::UNLIMITED,
-            subtree_reusable: Mutex::new(HashMap::default()),
-            pnpmfile_hook: None,
-            read_package_log: None,
-            skipped_optional_log: None,
-            allowed_deprecated_versions: BTreeMap::new(),
-            deprecation_log: None,
-            auto_install_peers: false,
-            registries: std::collections::HashMap::new(),
-            first_importer_by_pkg: Mutex::new(HashMap::default()),
-            first_walk_missing_by_pkg: Mutex::new(HashMap::default()),
-            changed_direct_deps: Mutex::new(HashMap::default()),
-            direct_dep_versions: Mutex::new(HashMap::default()),
-        }
-    }
-}
-
-impl WorkspaceTreeCtx {
-    /// Snapshot the workspace context into a [`ResolvedTree`] without
-    /// consuming `self`. `direct` carries the combined direct-dep
-    /// envelopes the caller built up across importers; multi-importer
-    /// orchestration usually leaves this empty and threads per-importer
-    /// direct deps separately into [`fn@crate::resolve_peers_workspace`].
-    pub fn snapshot(&self, direct: Vec<DirectDep>) -> ResolvedTree {
-        ResolvedTree {
-            direct,
-            packages: lock_recoverable(&self.packages).clone(),
-            dependencies_tree: lock_recoverable(&self.dependencies_tree).clone(),
-            all_peer_dep_names: lock_recoverable(&self.all_peer_dep_names).clone(),
-            policy_violations: lock_recoverable(&self.policy_violations).clone(),
-            applied_patches: lock_recoverable(&self.applied_patches).clone(),
-            children_by_id: lock_recoverable(&self.children_by_id).clone(),
-        }
-    }
-
-    /// Snapshot only the part of the occurrence tree reachable from one
-    /// importer's direct dependencies.
-    ///
-    /// A workspace resolve keeps every importer's occurrence nodes in
-    /// this shared context, so a full [`Self::snapshot`] taken for one
-    /// importer retains every other importer's nodes too. The
-    /// reachable-only snapshot keeps per-importer consumers (the
-    /// workspace-root hoistable-deps scan) proportional to that
-    /// importer's own subtree.
-    ///
-    /// Realized edges provide the occurrence-node closure. Lazy edges are
-    /// materialized by the peer walker from `children_by_id`, so their
-    /// package closure is collected separately and included here too.
-    #[must_use]
-    pub fn snapshot_reachable_from(&self, direct: Vec<DirectDep>) -> ResolvedTree {
-        let dependencies_tree = lock_recoverable(&self.dependencies_tree);
-        let mut reachable_node_ids = HashSet::default();
-        let mut reachable_pkg_ids = HashSet::default();
-        let mut pending_node_ids: Vec<NodeId> =
-            direct.iter().map(|dep| dep.node_id.clone()).collect();
-        while let Some(node_id) = pending_node_ids.pop() {
-            if !reachable_node_ids.insert(node_id.clone()) {
-                continue;
-            }
-            let Some(node) = dependencies_tree.get(&node_id) else {
-                continue;
-            };
-            reachable_pkg_ids.insert(node.resolved_package_id.clone());
-            if let crate::resolved_tree::TreeChildren::Realized(children) = &node.children {
-                pending_node_ids.extend(children.values().cloned());
-            }
-        }
-        let reachable_dependencies_tree: HashMap<_, _> = reachable_node_ids
-            .iter()
-            .filter_map(|node_id| {
-                dependencies_tree.get(node_id).cloned().map(|node| (node_id.clone(), node))
-            })
-            .collect();
-        // Release before taking the next guard so this function never
-        // holds two of the context's locks at once — `snapshot` takes
-        // `packages` before `dependencies_tree`, so overlapping here
-        // would create a reversed acquisition order.
-        drop(dependencies_tree);
-        let dependencies_tree = reachable_dependencies_tree;
-
-        let all_children = lock_recoverable(&self.children_by_id);
-        let mut pending_pkg_ids: Vec<String> = reachable_pkg_ids.iter().cloned().collect();
-        let mut children_by_id = HashMap::default();
-        while let Some(pkg_id) = pending_pkg_ids.pop() {
-            let Some(children) = all_children.get(&pkg_id) else {
-                continue;
-            };
-            children_by_id.insert(pkg_id, Arc::clone(children));
-            for child in children.iter() {
-                if reachable_pkg_ids.insert(child.pkg_id.clone()) {
-                    pending_pkg_ids.push(child.pkg_id.clone());
-                }
-            }
-        }
-        drop(all_children);
-
-        let packages = lock_recoverable(&self.packages);
-        let packages = reachable_pkg_ids
-            .into_iter()
-            .filter_map(|pkg_id| packages.get(&pkg_id).cloned().map(|pkg| (pkg_id, pkg)))
-            .collect();
-
-        ResolvedTree {
-            direct,
-            packages,
-            dependencies_tree,
-            all_peer_dep_names: lock_recoverable(&self.all_peer_dep_names).clone(),
-            policy_violations: lock_recoverable(&self.policy_violations).clone(),
-            applied_patches: lock_recoverable(&self.applied_patches).clone(),
-            children_by_id,
-        }
-    }
-
-    /// Attach a `readPackageHook` applied to every resolved manifest
-    /// before it enters the wanted-dep cache. See [`ManifestHook`] for
-    /// the signature.
-    #[must_use]
-    pub fn with_manifest_hook(mut self, manifest_hook: Option<ManifestHook>) -> Self {
-        self.manifest_hook = manifest_hook;
-        self
-    }
-
-    /// Attach the post-pnpmfile [`ManifestHook`] (overrides). See the
-    /// `overrides_hook` field for the ordering contract.
-    #[must_use]
-    pub fn with_overrides_hook(mut self, overrides_hook: Option<ManifestHook>) -> Self {
-        self.overrides_hook = overrides_hook;
-        self
-    }
-
-    /// Attach the prior `pnpm-lock.yaml` so `resolve_node` can reuse
-    /// already-resolved dependencies instead of re-resolving them. See
-    /// the `wanted_lockfile` field.
-    #[must_use]
-    pub fn with_wanted_lockfile(
-        mut self,
-        wanted_lockfile: Option<Arc<pacquet_lockfile::Lockfile>>,
-    ) -> Self {
-        self.wanted_lockfile = wanted_lockfile;
-        self
-    }
-
-    /// The prior `pnpm-lock.yaml` to reuse resolutions from, if any.
-    pub fn wanted_lockfile(&self) -> Option<&Arc<pacquet_lockfile::Lockfile>> {
-        self.wanted_lockfile.as_ref()
-    }
-
-    /// Snapshot of `pkg id → children-owner importer id`. See the field doc.
-    #[must_use]
-    pub fn first_importer_by_pkg(&self) -> HashMap<String, String> {
-        lock_recoverable(&self.first_importer_by_pkg).clone()
-    }
-
-    /// Record a walk's per-package missing-peer names. The owning
-    /// importer's report is written once per ownership generation —
-    /// its own later hoist waves never refresh it — and replaces any
-    /// provisional report a non-owner's earlier walk left behind. See
-    /// the `first_walk_missing_by_pkg` field doc.
-    pub fn record_first_walk_missing(
-        &self,
-        importer_id: &str,
-        missing_by_pkg: &HashMap<String, HashSet<String>>,
-    ) {
-        // Lock order: `children_owner_by_id` before
-        // `first_walk_missing_by_pkg`, the only place both are held.
-        let owners = lock_recoverable(&self.children_owner_by_id);
-        let mut record = lock_recoverable(&self.first_walk_missing_by_pkg);
-        for (pkg_id, ChildrenOwnerEntry { owner, .. }) in owners.iter() {
-            if owner.importer_id != importer_id {
-                continue;
-            }
-            let recorded_by_current_owner =
-                record.get(pkg_id).is_some_and(|entry| entry.recorded_by.as_ref() == Some(owner));
-            if !recorded_by_current_owner {
-                record.insert(
-                    pkg_id.clone(),
-                    OwnerMissingRecord {
-                        recorded_by: Some(owner.clone()),
-                        names: missing_by_pkg.get(pkg_id).cloned().unwrap_or_default(),
-                    },
-                );
-            }
-        }
-        for (pkg_id, names) in missing_by_pkg {
-            if owners.get(pkg_id).is_none_or(|entry| entry.owner.importer_id != importer_id) {
-                record.entry(pkg_id.clone()).or_insert_with(|| OwnerMissingRecord {
-                    recorded_by: None,
-                    names: names.clone(),
-                });
-            }
-        }
-    }
-
-    /// Snapshot of the per-package owner-context missing-peer names.
-    /// See the `first_walk_missing_by_pkg` field doc.
-    #[must_use]
-    pub fn first_walk_missing_by_pkg(&self) -> HashMap<String, HashSet<String>> {
-        lock_recoverable(&self.first_walk_missing_by_pkg)
-            .iter()
-            .map(|(pkg_id, entry)| (pkg_id.clone(), entry.names.clone()))
-            .collect()
-    }
-
-    /// Fold a freshly-inserted package's `(name, version)` into
-    /// [`Self::preferred_versions_from_run`]. Call once per new
-    /// `packages` entry.
-    fn record_resolved_version(&self, result: &pacquet_resolving_resolver_base::ResolveResult) {
-        if let Some(name_ver) = result.name_ver.as_ref() {
-            self.record_run_version(name_ver.name.to_string(), name_ver.suffix.to_string());
-        }
-    }
-
-    /// See [`Self::preferred_versions_from_run`]. Seed entries win over
-    /// run entries, and the first fold of a `(name, version)` pair wins
-    /// over later ones — the same `or_insert` semantics the per-importer
-    /// fold applied.
-    fn record_run_version(&self, name: String, version: String) {
-        lock_recoverable(&self.preferred_versions_from_run)
-            .entry(name)
-            .or_default()
-            .entry(version)
-            .or_insert(pacquet_resolving_resolver_base::VersionSelectorEntry::Plain(
-                pacquet_resolving_resolver_base::VersionSelectorType::Version,
-            ));
-    }
-
-    /// See the `revision` field doc.
-    pub(crate) fn revision(&self) -> u64 {
-        self.revision.load(std::sync::atomic::Ordering::Relaxed)
-    }
-
-    pub(crate) fn bump_revision(&self) {
-        self.revision.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-    }
-
-    /// See the `children_rewrites` field doc.
-    pub(crate) fn children_rewrites(&self) -> u64 {
-        self.children_rewrites.load(std::sync::atomic::Ordering::Relaxed)
-    }
-
-    pub(crate) fn record_children_rewrite(&self) {
-        self.children_rewrites.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-    }
-
-    /// Fold the context's growth since the last sync into `tree`, the
-    /// peer-hoist discovery engine's persistent view of the workspace.
-    ///
-    /// The shared maps grow monotonically during the hoist rounds, so
-    /// the sync inserts entries `tree` doesn't have yet and lowers node
-    /// depths that shrank. The exceptions are a children-owner change,
-    /// which can rewrite an existing package's recorded child list or
-    /// peer-dependency split: those invalidate walk results already
-    /// derived from the old values, so the sync reports them as
-    /// unmergeable (`false`) and the engine rebuilds its view from
-    /// scratch. A replaced `children_by_id` `Arc` with equal contents
-    /// (an ownership handover that re-recorded the same children) is
-    /// re-pointed without invalidating.
-    pub(crate) fn sync_discovery_tree(&self, tree: &mut ResolvedTree) -> bool {
-        use std::collections::hash_map::Entry;
-        {
-            let children_by_id = lock_recoverable(&self.children_by_id);
-            for (pkg_id, spec) in children_by_id.iter() {
-                match tree.children_by_id.entry(pkg_id.clone()) {
-                    Entry::Vacant(entry) => {
-                        entry.insert(Arc::clone(spec));
-                    }
-                    Entry::Occupied(mut entry) => {
-                        if Arc::ptr_eq(entry.get(), spec) {
-                            continue;
-                        }
-                        if **entry.get() != **spec {
-                            return false;
-                        }
-                        entry.insert(Arc::clone(spec));
-                    }
-                }
-            }
-        }
-        {
-            let packages = lock_recoverable(&self.packages);
-            for (pkg_id, pkg) in packages.iter() {
-                match tree.packages.entry(pkg_id.clone()) {
-                    Entry::Vacant(entry) => {
-                        entry.insert(pkg.clone());
-                    }
-                    Entry::Occupied(entry) => {
-                        if entry.get().peer_dependencies != pkg.peer_dependencies {
-                            return false;
-                        }
-                    }
-                }
-            }
-        }
-        {
-            let dependencies_tree = lock_recoverable(&self.dependencies_tree);
-            for (node_id, node) in dependencies_tree.iter() {
-                match tree.dependencies_tree.entry(node_id.clone()) {
-                    Entry::Vacant(entry) => {
-                        entry.insert(node.clone());
-                    }
-                    Entry::Occupied(mut entry) => {
-                        if entry.get().depth > node.depth {
-                            entry.get_mut().depth = node.depth;
-                        }
-                    }
-                }
-            }
-        }
-        tree.all_peer_dep_names.extend(lock_recoverable(&self.all_peer_dep_names).iter().cloned());
-        true
-    }
-
-    /// `NodeId → pkgIdWithPatchHash` for the given peer-provider nodes,
-    /// keeping only nodes the shared context resolved eagerly (a node
-    /// the peer walker realized lazily has no context entry and is
-    /// dropped, matching the hoist loop's providers-that-existed-before-
-    /// the-pass filter).
-    pub(crate) fn provider_pkg_ids<'node_ids>(
-        &self,
-        node_ids: impl Iterator<Item = &'node_ids NodeId>,
-    ) -> HashMap<NodeId, String> {
-        let dependencies_tree = lock_recoverable(&self.dependencies_tree);
-        let packages = lock_recoverable(&self.packages);
-        node_ids
-            .filter_map(|node_id| {
-                let pkg_id = &dependencies_tree.get(node_id)?.resolved_package_id;
-                packages.contains_key(pkg_id).then(|| (node_id.clone(), pkg_id.clone()))
-            })
-            .collect()
-    }
-
-    pub(crate) fn set_direct_locked_peer_names(
-        &self,
-        direct: &[DirectDep],
-        names_by_alias: &HashMap<String, Arc<HashSet<String>>>,
-    ) {
-        let mut tree = lock_recoverable(&self.dependencies_tree);
-        for dep in direct {
-            let Some(names) = names_by_alias.get(&dep.alias) else {
-                continue;
-            };
-            if let Some(node) = tree.get_mut(&dep.node_id) {
-                node.locked_peer_names = Some(Arc::clone(names));
-            }
-        }
-    }
-
-    /// Set which dependencies `pacquet update` excludes from reuse. See
-    /// [`UpdateReuseScope`].
-    #[must_use]
-    pub fn with_update_reuse_scope(mut self, scope: UpdateReuseScope) -> Self {
-        self.update_reuse_scope = scope;
-        self
-    }
-
-    #[must_use]
-    pub fn with_update_reuse_scopes_by_importer(
-        mut self,
-        scopes: BTreeMap<String, UpdateReuseScope>,
-    ) -> Self {
-        self.update_reuse_scopes_by_importer = scopes;
-        self
-    }
-
-    #[must_use]
-    pub fn with_update_depth(mut self, update_depth: UpdateDepth) -> Self {
-        self.update_depth = update_depth;
-        self
-    }
-
-    fn update_reuse_scope_for(&self, importer_id: &str) -> &UpdateReuseScope {
-        if matches!(self.update_reuse_scope, UpdateReuseScope::None) {
-            return &self.update_reuse_scope;
-        }
-        self.update_reuse_scopes_by_importer.get(importer_id).unwrap_or(&self.update_reuse_scope)
-    }
-
-    #[must_use]
-    pub fn with_pnpmfile_hook(mut self, pnpmfile_hook: Option<Arc<dyn PnpmfileHooks>>) -> Self {
-        self.pnpmfile_hook = pnpmfile_hook;
-        self
-    }
-
-    /// Attach the `context.log(...)` sink the `pnpmfile_hook`'s
-    /// `readPackage` calls forward to. The install layer pre-binds the
-    /// reporter, project prefix, and pnpmfile path into the closure so the
-    /// resolver stays reporter-agnostic.
-    #[must_use]
-    pub fn with_read_package_log(mut self, read_package_log: Option<pacquet_hooks::LogFn>) -> Self {
-        self.read_package_log = read_package_log;
-        self
-    }
-
-    /// Attach the sink skipped-optional-dependency notifications are
-    /// forwarded to. See [`SkippedOptionalLogFn`].
-    #[must_use]
-    pub fn with_skipped_optional_log(
-        mut self,
-        skipped_optional_log: Option<SkippedOptionalLogFn>,
-    ) -> Self {
-        self.skipped_optional_log = skipped_optional_log;
-        self
-    }
-
-    /// Attach the `pnpm.allowedDeprecatedVersions` map. See
-    /// [`crate::WorkspaceResolveOptions::allowed_deprecated_versions`].
-    #[must_use]
-    pub fn with_allowed_deprecated_versions(
-        mut self,
-        allowed_deprecated_versions: BTreeMap<String, String>,
-    ) -> Self {
-        self.allowed_deprecated_versions = allowed_deprecated_versions;
-        self
-    }
-
-    /// Attach the sink deprecation notifications are forwarded to.
-    /// See [`DeprecationLogFn`].
-    #[must_use]
-    pub fn with_deprecation_log(mut self, deprecation_log: Option<DeprecationLogFn>) -> Self {
-        self.deprecation_log = deprecation_log;
-        self
-    }
-
-    /// Set the install's `autoInstallPeers` flag. See the field doc.
-    #[must_use]
-    pub fn with_auto_install_peers(mut self, auto_install_peers: bool) -> Self {
-        self.auto_install_peers = auto_install_peers;
-        self
-    }
-
-    /// Attach the resolved registry map. See the `registries` field.
-    #[must_use]
-    pub fn with_registries(
-        mut self,
-        registries: std::collections::HashMap<String, String>,
-    ) -> Self {
-        self.registries = registries;
-        self
-    }
-
-    /// Take ownership of `self` and emit the final [`ResolvedTree`].
-    /// Pacquet's single-importer path consumes the context via
-    /// [`TreeCtx::into_resolved_tree`], which routes through here once
-    /// the last `Arc<WorkspaceTreeCtx>` reference is the [`TreeCtx`]'s
-    /// own.
-    pub fn into_resolved_tree(self, direct: Vec<DirectDep>) -> ResolvedTree {
-        ResolvedTree {
-            direct,
-            packages: self.packages.into_inner().unwrap_or_else(std::sync::PoisonError::into_inner),
-            dependencies_tree: self
-                .dependencies_tree
-                .into_inner()
-                .unwrap_or_else(std::sync::PoisonError::into_inner),
-            all_peer_dep_names: self
-                .all_peer_dep_names
-                .into_inner()
-                .unwrap_or_else(std::sync::PoisonError::into_inner),
-            policy_violations: self
-                .policy_violations
-                .into_inner()
-                .unwrap_or_else(std::sync::PoisonError::into_inner),
-            applied_patches: self
-                .applied_patches
-                .into_inner()
-                .unwrap_or_else(std::sync::PoisonError::into_inner),
-            children_by_id: self
-                .children_by_id
-                .into_inner()
-                .unwrap_or_else(std::sync::PoisonError::into_inner),
-        }
-    }
-}
-
-/// Mutable workspace for an in-flight tree walk. The orchestrator
-/// (`resolve_importer`) holds one of these across hoist iterations and
-/// extends it via [`extend_tree`] so newly-hoisted peer dependencies
-/// reuse the existing per-id dedup map instead of restarting the walk.
-///
-/// The shared per-`pkgIdWithPatchHash` dedup maps live on
-/// [`WorkspaceTreeCtx`] behind an `Arc`. In single-importer mode this
-/// `Arc` is sole-owned by [`TreeCtx`]; in multi-importer mode
-/// `Arc::clone(&workspace)` is handed to every per-importer
-/// [`TreeCtx`] so importer N's tree walk reuses importer M's resolved
-/// envelopes via the shared maps.
-pub struct TreeCtx {
-    base_opts: ResolveOptions,
-    /// Absolute root used to make ownerless snapshot `link:` ids stable
-    /// across importers at different depths.
-    lockfile_dir: PathBuf,
-    /// [`ResolveOptions`] handed to the resolver for importer-level
-    /// (direct) dependencies — `depth == 0`. Differs from `base_opts`
-    /// only in `pick_lowest_version`, which is set under
-    /// `resolutionMode: time-based` / `lowest-direct`. Built once per
-    /// importer by [`Self::with_resolution_mode`].
-    direct_opts: ResolveOptions,
-    /// [`ResolveOptions`] handed to the resolver for transitive
-    /// dependencies — `depth > 0`. Always picks highest; carries the
-    /// `time-based` publish-date cutoff in `published_by`. Built once
-    /// per importer by [`Self::with_resolution_mode`].
-    subdep_opts: ResolveOptions,
-    /// Workspace catalogs used to resolve `catalog:` children of injected
-    /// workspace packages. Other transitive dependencies keep catalog
-    /// resolution disabled.
-    catalogs: Catalogs,
-    workspace: Arc<WorkspaceTreeCtx>,
-    /// Configured `patchedDependencies` (already grouped by name).
-    /// Shared by `Arc` so the lookup table doesn't get cloned per
-    /// recursive call. `None` when no patches are configured for this
-    /// install.
-    patched_dependencies: Option<Arc<PatchGroupRecord>>,
-    /// The importer this per-importer context walks for. Recorded into
-    /// [`WorkspaceTreeCtx`]'s `first_importer_by_pkg` when one of its
-    /// occurrences owns a package's shared children context.
-    importer_id: String,
-    importer_order: usize,
-}
-
-impl TreeCtx {
-    /// Construct a single-importer context with a fresh
-    /// [`WorkspaceTreeCtx`]. The multi-importer orchestrator uses
-    /// [`Self::with_workspace`] instead so per-importer contexts share
-    /// the same workspace ctx.
-    #[must_use]
-    pub fn new(base_opts: ResolveOptions) -> Self {
-        let lockfile_dir = if base_opts.lockfile_dir.as_os_str().is_empty() {
-            base_opts.project_dir.clone()
-        } else {
-            base_opts.lockfile_dir.clone()
-        };
-        TreeCtx {
-            direct_opts: base_opts.clone(),
-            subdep_opts: base_opts.clone(),
-            base_opts,
-            lockfile_dir: pacquet_fs::lexical_normalize(&lockfile_dir),
-            catalogs: Catalogs::new(),
-            workspace: Arc::new(WorkspaceTreeCtx::default()),
-            patched_dependencies: None,
-            importer_id: pacquet_lockfile::Lockfile::ROOT_IMPORTER_KEY.to_string(),
-            importer_order: 0,
-        }
-    }
-
-    /// Construct a per-importer context that shares its dedup maps
-    /// with `workspace`. The caller is responsible for keeping
-    /// `workspace` alive across importers (typically via
-    /// `Arc::clone(&workspace)`).
-    pub fn with_workspace(workspace: Arc<WorkspaceTreeCtx>, base_opts: ResolveOptions) -> Self {
-        let lockfile_dir = if base_opts.lockfile_dir.as_os_str().is_empty() {
-            base_opts.project_dir.clone()
-        } else {
-            base_opts.lockfile_dir.clone()
-        };
-        TreeCtx {
-            direct_opts: base_opts.clone(),
-            subdep_opts: base_opts.clone(),
-            base_opts,
-            lockfile_dir: pacquet_fs::lexical_normalize(&lockfile_dir),
-            catalogs: Catalogs::new(),
-            workspace,
-            patched_dependencies: None,
-            importer_id: pacquet_lockfile::Lockfile::ROOT_IMPORTER_KEY.to_string(),
-            importer_order: 0,
-        }
-    }
-
-    #[must_use]
-    pub(crate) fn with_lockfile_dir(mut self, lockfile_dir: &Path) -> Self {
-        self.lockfile_dir = pacquet_fs::lexical_normalize(lockfile_dir);
-        self
-    }
-
-    /// Derive the depth-specific resolve options from `resolutionMode`.
-    ///
-    /// - `pick_lowest_direct` — resolve direct dependencies to their
-    ///   lowest satisfying version (`time-based` / `lowest-direct`).
-    /// - `subdep_published_by` — the publish-date cutoff applied to
-    ///   transitive dependencies. Under `time-based` this is the
-    ///   workspace-wide cutoff computed from the resolved direct deps
-    ///   (clamped by `minimumReleaseAge`); otherwise it is just
-    ///   `base_opts.published_by` (the `minimumReleaseAge` cutoff),
-    ///   leaving subdep resolution unchanged.
-    ///
-    /// Splits the importer-dep pick from the subdep pick (always
-    /// highest, constrained by the computed `publishedBy`).
-    #[must_use]
-    pub fn with_resolution_mode(
-        mut self,
-        pick_lowest_direct: bool,
-        subdep_published_by: Option<DateTime<Utc>>,
-    ) -> Self {
-        self.direct_opts.pick_lowest_version = pick_lowest_direct;
-        self.subdep_opts.pick_lowest_version = false;
-        self.subdep_opts.published_by = subdep_published_by;
-        self
-    }
-
-    #[must_use]
-    pub fn with_catalogs(mut self, catalogs: Catalogs) -> Self {
-        self.catalogs = catalogs;
-        self
-    }
-
-    /// The [`ResolveOptions`] to hand the resolver for a node at the
-    /// given `depth`: importer-level deps (`depth == 0`) use
-    /// [`Self::direct_opts`]; everything below uses
-    /// [`Self::subdep_opts`].
-    fn opts_for_depth(&self, depth: i32) -> &ResolveOptions {
-        if depth == 0 { &self.direct_opts } else { &self.subdep_opts }
-    }
-
-    /// Borrow the shared workspace ctx so callers can hand the same
-    /// `Arc::clone` to the next per-importer [`TreeCtx`].
-    #[must_use]
-    pub fn workspace(&self) -> &Arc<WorkspaceTreeCtx> {
-        &self.workspace
-    }
-
-    fn update_reuse_scope(&self) -> &UpdateReuseScope {
-        self.workspace.update_reuse_scope_for(&self.importer_id)
-    }
-
-    fn update_scope(&self) -> UpdateScope<'_> {
-        UpdateScope { reuse: self.update_reuse_scope(), max_depth: self.workspace.update_depth }
-    }
-
-    fn update_cache_scope(&self) -> Option<String> {
-        (!matches!(self.update_reuse_scope(), UpdateReuseScope::All))
-            .then(|| self.importer_id.clone())
-    }
-
-    /// Set the importer this context walks for. See [`TreeCtx`]'s
-    /// `importer_id` field.
-    #[must_use]
-    pub fn with_importer_id(mut self, importer_id: &str) -> Self {
-        self.importer_id = importer_id.to_string();
-        self
-    }
-
-    /// Set this importer's position in the workspace input order.
-    /// Child-subtree ownership uses it after depth, matching pnpm's
-    /// deterministic `(depth, importer order, parent path)` tie-break.
-    #[must_use]
-    pub fn with_importer_order(mut self, importer_order: usize) -> Self {
-        self.importer_order = importer_order;
-        self
-    }
-
-    /// Attach the install's `patchedDependencies` map. When `Some`,
-    /// the per-node walker looks every resolved `name@version` up via
-    /// [`get_patch_info`] and appends `(patch_hash=<hash>)` to the
-    /// `pkgIdWithPatchHash` on a match.
-    #[must_use]
-    pub fn with_patched_dependencies(
-        mut self,
-        patched_dependencies: Option<Arc<PatchGroupRecord>>,
-    ) -> Self {
-        self.patched_dependencies = patched_dependencies;
-        self
-    }
-
-    /// Attach a `readPackageHook` to the underlying [`WorkspaceTreeCtx`].
-    /// `manifest_hook` is workspace-wide (one hook per install), so this
-    /// passthrough relies on the workspace ctx being sole-owned —
-    /// `TreeCtx::new` always satisfies that, and the multi-importer
-    /// orchestrator [`fn@crate::resolve_workspace`] hands the hook in via
-    /// [`WorkspaceTreeCtx::with_manifest_hook`] before sharing the
-    /// `Arc`. Panics if the workspace ctx has already been cloned —
-    /// callers must set the hook before sharing the context.
-    #[must_use]
-    pub fn with_manifest_hook(mut self, manifest_hook: Option<ManifestHook>) -> Self {
-        Arc::get_mut(&mut self.workspace)
-            .expect("with_manifest_hook called after the workspace ctx was shared via Arc::clone")
-            .manifest_hook = manifest_hook;
-        self
-    }
-
-    /// Attach the post-pnpmfile [`ManifestHook`] (overrides) to the
-    /// underlying [`WorkspaceTreeCtx`]; same sole-ownership contract as
-    /// [`Self::with_manifest_hook`].
-    #[must_use]
-    pub fn with_overrides_hook(mut self, overrides_hook: Option<ManifestHook>) -> Self {
-        Arc::get_mut(&mut self.workspace)
-            .expect("with_overrides_hook called after the workspace ctx was shared via Arc::clone")
-            .overrides_hook = overrides_hook;
-        self
-    }
-
-    #[must_use]
-    pub fn with_pnpmfile_hook(mut self, pnpmfile_hook: Option<Arc<dyn PnpmfileHooks>>) -> Self {
-        Arc::get_mut(&mut self.workspace)
-            .expect("with_pnpmfile_hook called after the workspace ctx was shared via Arc::clone")
-            .pnpmfile_hook = pnpmfile_hook;
-        self
-    }
-
-    /// Attach the `context.log(...)` sink the `pnpmfile_hook`'s
-    /// `readPackage` calls forward to. Like [`Self::with_pnpmfile_hook`],
-    /// this targets the underlying [`WorkspaceTreeCtx`] and panics if it
-    /// has already been shared via `Arc::clone`.
-    #[must_use]
-    pub fn with_read_package_log(mut self, read_package_log: Option<pacquet_hooks::LogFn>) -> Self {
-        Arc::get_mut(&mut self.workspace)
-            .expect(
-                "with_read_package_log called after the workspace ctx was shared via Arc::clone",
-            )
-            .read_package_log = read_package_log;
-        self
-    }
-
-    /// Set the install's `autoInstallPeers` flag on the underlying
-    /// [`WorkspaceTreeCtx`]. Like [`Self::with_pnpmfile_hook`], panics if
-    /// it has already been shared via `Arc::clone`.
-    #[must_use]
-    pub fn with_auto_install_peers(mut self, auto_install_peers: bool) -> Self {
-        Arc::get_mut(&mut self.workspace)
-            .expect(
-                "with_auto_install_peers called after the workspace ctx was shared via Arc::clone",
-            )
-            .auto_install_peers = auto_install_peers;
-        self
-    }
-
-    /// Take ownership of `self` and emit the final [`ResolvedTree`]
-    /// the peer-resolution stage consumes. The orchestrator passes its
-    /// cumulative [`DirectDep`] list (initial walk + each hoist
-    /// iteration's contributions) as `direct`.
-    ///
-    /// When the [`WorkspaceTreeCtx`] is sole-owned by this context
-    /// (single-importer install) the inner mutex contents move
-    /// directly into the [`ResolvedTree`]; otherwise the maps are
-    /// cloned out via [`WorkspaceTreeCtx::snapshot`].
-    #[must_use]
-    pub fn into_resolved_tree(self, direct: Vec<DirectDep>) -> ResolvedTree {
-        match Arc::try_unwrap(self.workspace) {
-            Ok(ws) => ws.into_resolved_tree(direct),
-            Err(arc) => arc.snapshot(direct),
-        }
-    }
-
-    /// Build a snapshot of the current tree state without consuming
-    /// `self`. The orchestrator's hoist loop snapshots after each
-    /// [`extend_tree`] call to run [`fn@crate::resolve_peers`] over the
-    /// growing tree and find missing peers to hoist next.
-    #[must_use]
-    pub fn snapshot(&self, direct: Vec<DirectDep>) -> ResolvedTree {
-        self.workspace.snapshot(direct)
-    }
-
-    /// Build an importer-scoped snapshot for a peer-hoist pass.
-    #[must_use]
-    pub fn snapshot_reachable_from(&self, direct: Vec<DirectDep>) -> ResolvedTree {
-        self.workspace.snapshot_reachable_from(direct)
-    }
-
-    /// The preferred-version buckets for `names`: the caller's seed
-    /// entries merged with every version this run has resolved so far
-    /// (seed entries win per selector). The peer-hoist pickers look up
-    /// only their missing-peer names, so this materializes a handful of
-    /// buckets instead of a per-importer copy of the whole run history.
-    pub(crate) fn preferred_versions_for_names<'name>(
-        &self,
-        seed: &pacquet_resolving_resolver_base::PreferredVersions,
-        names: impl Iterator<Item = &'name str>,
-    ) -> pacquet_resolving_resolver_base::PreferredVersions {
-        let run = lock_recoverable(&self.workspace.preferred_versions_from_run);
-        let mut out = pacquet_resolving_resolver_base::PreferredVersions::new();
-        for name in names {
-            let mut bucket = seed.get(name).cloned().unwrap_or_default();
-            if let Some(run_bucket) = run.get(name) {
-                for (selector, entry) in run_bucket {
-                    bucket.entry(selector.clone()).or_insert_with(|| entry.clone());
-                }
-            }
-            if !bucket.is_empty() {
-                out.insert(name.to_string(), bucket);
-            }
-        }
-        out
-    }
-}
 
 /// Walk an additional set of `(alias, range)` pairs as new direct
 /// dependencies of the importer, extending `ctx` in place. Returns the
@@ -1690,6 +513,8 @@ impl TreeCtx {
 /// importer can't appear in its own ancestor chain), but the walker
 /// may still return `None` for any spec the cycle break gated out;
 /// those are filtered here.
+///
+/// [`ResolvedPackage`]: crate::ResolvedPackage
 pub async fn extend_tree<Chain>(
     ctx: &TreeCtx,
     resolver: &Chain,
@@ -1785,2252 +610,6 @@ where
         .pipe(future::try_join_all)
         .await?;
     Ok(results.into_iter().flatten().collect())
-}
-
-/// Resolve one `(alias, range)` edge end-to-end with no
-/// preferred-versions overlay: [`fn@resolve_node_seed`] then
-/// [`fn@walk_node_children`]. Used where per-level preference folding
-/// does not apply — the lockfile-reuse subtree walk, whose versions
-/// are exact pins.
-///
-/// The node's children resolve in the same `parent_pkg_aliases` scope
-/// as the node itself, without this level's own aliases folded in: a
-/// reused subtree takes its children from the lockfile snapshot rather
-/// than from a manifest, so no shadowed dependency can be dropped
-/// inside it, and the narrower scope only ever means one omission
-/// fewer where an edge falls back to a fresh resolve.
-#[expect(
-    clippy::too_many_arguments,
-    reason = "internal walker helper threading per-node context through the recursion"
-)]
-#[async_recursion]
-async fn resolve_node<Chain>(
-    ctx: &TreeCtx,
-    resolver: &Chain,
-    wanted: WantedDependency,
-    ancestor_ids: &Arc<Vec<String>>,
-    depth: i32,
-    parent_optional: bool,
-    reuse: ReuseSource,
-    parent_pkg_aliases: &Arc<ParentPkgAliases>,
-) -> Result<Option<DirectDep>, ResolveDependencyTreeError>
-where
-    Chain: Resolver + ?Sized,
-{
-    let base_overlay = ctx.base_opts.preferred_versions_overlay.clone();
-    match resolve_node_seed(
-        ctx,
-        resolver,
-        wanted,
-        ancestor_ids,
-        depth,
-        parent_optional,
-        reuse,
-        base_overlay.clone(),
-        None,
-        parent_pkg_aliases,
-    )
-    .await?
-    {
-        NodeSeed::Done(dep) => Ok(dep),
-        NodeSeed::Pending(pending) => {
-            walk_node_children(
-                ctx,
-                resolver,
-                *pending,
-                base_overlay,
-                Arc::clone(parent_pkg_aliases),
-            )
-            .await
-        }
-    }
-}
-
-/// Outcome of [`fn@resolve_node_seed`]: either the edge completed
-/// without a children walk (lockfile reuse, cycle break), or the
-/// package resolved and its children walk is still pending — the
-/// caller runs it via [`fn@walk_node_children`] once every sibling
-/// seed settled, so the children's resolution sees the whole level's
-/// versions in its preferred-versions overlay.
-enum NodeSeed {
-    Done(Option<DirectDep>),
-    Pending(Box<PendingNode>),
-}
-
-/// A resolved-but-not-walked node: everything
-/// [`fn@walk_node_children`] needs to recurse into the children.
-struct PendingNode {
-    result: Arc<pacquet_resolving_resolver_base::ResolveResult>,
-    id: String,
-    alias: String,
-    node_id: NodeId,
-    is_link: bool,
-    parent_ancestors: Arc<Vec<String>>,
-    next_ancestors: Arc<Vec<String>>,
-    /// The deterministic children-ownership claim taken at seed time;
-    /// the walk phase re-checks it before recording the children, so
-    /// a better-placed occurrence seeded after this one still wins.
-    /// It also carries the peer-shadowed dependency names the walk
-    /// phase filters this package's children by.
-    children_owner: ChildrenOwnerClaim,
-    depth: i32,
-    current_is_optional: bool,
-    /// The edge's recorded snapshot key in the prior lockfile, if
-    /// any — threads each child's prior ref through the walk phase
-    /// via `ReuseSource::Transitive`.
-    prior_key: Option<PkgNameVerPeer>,
-}
-
-/// Resolve one `(alias, range)` edge and register the resolved package
-/// in the dedup map if absent, run for a whole sibling level before any
-/// child subtree starts.
-///
-/// `pick_overlay` carries the per-level preferred-version additions
-/// (the parent level's resolved versions) consulted by the npm
-/// resolver's version pick; it participates in the per-wanted dedup
-/// cache key so the same range can legitimately pick different
-/// versions under different levels, layering each level's resolved
-/// versions onto the preferred-versions fold.
-///
-/// `ancestor_ids` is the chain of `pkgIdWithPatchHash` values from the
-/// root importer down to the current node's parent. When the resolved
-/// id appears in the chain, this call is a cycle re-entry: the edge is
-/// dropped entirely (returns `Done(None)`) so the parent's `children`
-/// map omits the cycled child. Without this, two nodes for the same id
-/// race each other into `graph.insert`, and an empty-children entry for
-/// the cycled occurrence can overwrite the real one.
-///
-/// `parent_dir` is the directory of the manifest that declares this
-/// edge, when that manifest is a package resolved from a local
-/// directory — see [`declaring_manifest_dir`].
-///
-/// `parent_pkg_aliases` is the scope this edge's level resolves in; it
-/// decides which of the resolved package's `dependencies` its own
-/// `peerDependencies` shadow (see [`peer_shadowed_dependencies`]).
-#[expect(
-    clippy::too_many_arguments,
-    reason = "internal walker helper threading per-node context through the recursion"
-)]
-#[async_recursion]
-async fn resolve_node_seed<Chain>(
-    ctx: &TreeCtx,
-    resolver: &Chain,
-    wanted: WantedDependency,
-    ancestor_ids: &Arc<Vec<String>>,
-    depth: i32,
-    parent_optional: bool,
-    reuse: ReuseSource,
-    pick_overlay: Option<Arc<PreferredVersionsOverlay>>,
-    parent_dir: Option<&Path>,
-    parent_pkg_aliases: &Arc<ParentPkgAliases>,
-) -> Result<NodeSeed, ResolveDependencyTreeError>
-where
-    Chain: Resolver + ?Sized,
-{
-    let current_is_optional = wanted.optional.unwrap_or(false) || parent_optional;
-
-    // The edge's recorded snapshot key in the prior lockfile, if any.
-    // Feeds both subtree reuse (below) and — when the edge re-resolves
-    // anyway — the `currentPkg` payload custom resolvers receive.
-    let prior_key = reuse.prior_key(ctx, &wanted);
-
-    // **Lockfile-resolution reuse.** When the prior lockfile already
-    // resolved this edge (and the recorded version still satisfies the
-    // manifest range, for a direct dep), synthesize the resolution from
-    // the lockfile and walk its transitive subtree from the snapshot
-    // graph instead of re-resolving from the registry.
-    // `synthesize_reused_result` is conservative: any shape it can't
-    // faithfully reproduce (non-registry resolutions, missing metadata)
-    // yields `None` here and the node falls through to a fresh resolve.
-    //
-    // Stale-pin refresh: a node depending on a changed direct dep is
-    // resolved fresh rather than reused, so its children walk against
-    // their manifest ranges where `walk_node_children` can redirect a
-    // stale pin onto the higher direct-dep version (reusing the subtree
-    // would keep the pin, leaving the lockfile non-convergent).
-    if reuse.allows_reuse()
-        && !node_depends_on_changed_direct_dep(ctx, prior_key.as_ref())
-        && let Some(reused) = try_reuse_node(ctx, &wanted, prior_key.as_ref(), depth)
-    {
-        return resolve_reused_node(
-            ctx,
-            resolver,
-            wanted,
-            ancestor_ids,
-            depth,
-            current_is_optional,
-            reused,
-            parent_pkg_aliases,
-        )
-        .await
-        .map(NodeSeed::Done);
-    }
-
-    // Memoise the per-wanted resolve. The first caller for a given
-    // `(alias, bare_specifier, optional)` runs the resolver chain and
-    // stores the `Arc<ResolveResult>` on `ctx.resolved_by_wanted`;
-    // every later caller for the same wanted dep clones the `Arc` and
-    // skips the chain entirely. Concurrent first-callers can both miss
-    // the cache and run `resolver.resolve` in parallel — the resolver's
-    // own per-cache-key semaphore (`pick_package::fetch_locker`)
-    // already coalesces those into a single network fetch, so the
-    // doubled work is bounded to in-memory packument lookups + semver
-    // matching, and the second to finish loses the `insert` race
-    // harmlessly (the entry holds an `Arc` to an equivalent
-    // `ResolveResult`).
-    // `resolutionMode` makes the version pick depend on whether this is
-    // a direct (`depth == 0`) or transitive dep, so the cache key and
-    // the resolver call both key off the depth-specific options.
-    let opts = ctx.opts_for_depth(depth);
-    // The prior lockfile entry rides along as `currentPkg`, handed to
-    // the resolver. Only custom resolvers read it today; the clone of
-    // the shared per-depth options is paid only when a prior entry
-    // exists for a freshly resolving edge.
-    let current_pkg = prior_key.as_ref().and_then(|key| {
-        let lockfile = ctx.workspace.wanted_lockfile.as_ref()?;
-        current_pkg_from_lockfile(lockfile, key, &ctx.workspace.registries)
-    });
-    let opts_with_current_pkg;
-    let opts = match current_pkg {
-        Some(current_pkg) => {
-            opts_with_current_pkg =
-                ResolveOptions { current_pkg: Some(current_pkg), ..opts.clone() };
-            &opts_with_current_pkg
-        }
-        None => opts,
-    };
-    let opts_for_file_dep = opts_relative_to_declaring_manifest(opts, &wanted, parent_dir);
-    let opts = opts_for_file_dep.as_ref();
-    // Project-relative resolutions (`link:`/`file:`/`workspace:`) are
-    // keyed by the consuming importer so one importer's relative path
-    // is never reused by another. See [`WantedKey`]. The prior key
-    // joins so two edges that share a specifier but recorded different
-    // versions never share a `currentPkg`-dependent result.
-    let project_scope = project_relative_cache_scope(&wanted, opts);
-    // The overlay's view for this edge joins the cache key: the same
-    // range can legitimately pick different versions under levels
-    // that resolved different siblings. The view keeps each candidate
-    // name (alias, `npm:` inner target, folded `jsr:` name) paired
-    // with its versions — the picker consults the overlay per name,
-    // so a flat union of versions could collide two overlays that
-    // distribute the same versions across different names. Empty for
-    // almost every edge, so the dedup keeps working where it matters.
-    let overlay_versions: Vec<(String, Vec<String>)> = pick_overlay
-        .as_ref()
-        .map(|overlay| {
-            let mut view: Vec<(String, Vec<String>)> = overlay_lookup_names(&wanted)
-                .into_iter()
-                .filter_map(|name| {
-                    let mut versions: Vec<String> =
-                        overlay.versions_for(&name).into_iter().map(str::to_string).collect();
-                    if versions.is_empty() {
-                        return None;
-                    }
-                    versions.sort_unstable();
-                    versions.dedup();
-                    Some((name, versions))
-                })
-                .collect();
-            view.sort_unstable();
-            view
-        })
-        .unwrap_or_default();
-    let update_target = is_update_target(ctx.update_scope(), &wanted, depth);
-    let cache_key: WantedKey = (
-        wanted.alias.clone(),
-        wanted.bare_specifier.clone(),
-        wanted.optional,
-        wanted.injected,
-        opts.pick_lowest_version,
-        opts.published_by,
-        project_scope,
-        prior_key.clone(),
-        overlay_versions.clone(),
-        ctx.update_cache_scope(),
-        update_target,
-    );
-    let result =
-        match resolve_wanted_cached(ctx, resolver, &wanted, opts, pick_overlay.as_ref(), cache_key)
-            .await
-        {
-            Ok(result) => result,
-            // A resolution failure on an optional edge drops the edge
-            // instead of failing the install — unless the wanted lockfile
-            // already holds a satisfying entry, where the silent skip would
-            // erase the locked entries (see
-            // `wanted_lockfile_contains_satisfying_entry`). Hook errors
-            // keep aborting even for optional edges.
-            Err(
-                err @ (ResolveDependencyTreeError::Resolve(_)
-                | ResolveDependencyTreeError::NoMatchingVersion(_)
-                | ResolveDependencyTreeError::RegistryResponse(_)
-                | ResolveDependencyTreeError::SpecNotSupported { .. }),
-            ) if wanted.optional.unwrap_or(false) => {
-                if wanted_lockfile_contains_satisfying_entry(
-                    ctx.workspace.wanted_lockfile.as_deref(),
-                    &wanted,
-                ) {
-                    return Err(ResolveDependencyTreeError::LockedOptionalResolutionFailure(
-                        Box::new(err),
-                    ));
-                }
-                if let Some(log) = ctx.workspace.skipped_optional_log.as_ref() {
-                    log(SkippedOptionalDependency {
-                        details: err.to_string(),
-                        name: wanted.alias.clone(),
-                        version: wanted
-                            .alias
-                            .is_some()
-                            .then(|| wanted.bare_specifier.clone())
-                            .flatten(),
-                        bare_specifier: wanted.bare_specifier.clone().unwrap_or_default(),
-                        parents: pkgs_info_from_ids(ctx, ancestor_ids),
-                        prefix: opts.project_dir.display().to_string(),
-                    });
-                }
-                return Ok(NodeSeed::Done(None));
-            }
-            Err(err) => return Err(err),
-        };
-
-    if let Some(violation) = result.policy_violation.clone() {
-        lock_recoverable(&ctx.workspace.policy_violations).push(violation);
-    }
-
-    if ctx.base_opts.block_exotic_subdeps
-        && depth > 0
-        && is_exotic_resolved_via(&result.resolved_via)
-    {
-        return Err(ResolveDependencyTreeError::ExoticSubdep {
-            specifier: wanted
-                .alias
-                .clone()
-                .or_else(|| wanted.bare_specifier.clone())
-                .unwrap_or_default(),
-            resolved_via: result.resolved_via.clone(),
-        });
-    }
-
-    if result.name_ver.is_none()
-        && wanted.bare_specifier.as_deref().is_some_and(|specifier| {
-            specifier.starts_with("workspace:") && !specifier.starts_with("workspace:.")
-        })
-        && let Some(manifest) = result.manifest.as_deref()
-        && let (Some(name), Some(version)) = (
-            manifest.get("name").and_then(Value::as_str),
-            manifest.get("version").and_then(Value::as_str),
-        )
-    {
-        ctx.workspace.record_run_version(name.to_string(), version.to_string());
-    }
-
-    let id = build_pkg_id_with_patch_hash(ctx, &result).await?;
-
-    // Cycle break — see the doc comment above. A direct self-edge and
-    // the second lap of a longer cycle are dropped; the first re-entry
-    // is kept so the cycle-closing edge reaches the lockfile snapshot.
-    if ancestor_ids.last().is_some_and(|parent| {
-        *parent == id || parent_ids_contain_sequence(ancestor_ids, parent, &id)
-    }) {
-        return Ok(NodeSeed::Done(None));
-    }
-
-    let alias = node_alias(&wanted, &result, &id);
-
-    // Build (or look up) the ResolvedPackage envelope. The first
-    // visitor populates it; later visitors AND-fold the `optional`
-    // flag so a single non-optional path flips it back to `false`.
-    // Child traversal is claimed first, and a later
-    // deterministically-better occurrence replaces the shared
-    // `children_by_id` entry — plus, since the two are two halves of
-    // one manifest reading, the envelope's peer dependencies.
-    // Leaves (no deps / optional deps / peers / peerDependenciesMeta)
-    // reuse the package id as their `NodeId`, collapsing every parent
-    // edge onto one tree node. Non-leaves still get a fresh per-
-    // occurrence id so the peer resolver can attach different peer
-    // suffixes per call site.
-    // The leaf flag is computed before the dedup insert so it can be
-    // persisted on [`ResolvedPackage::is_leaf`] for the lazy realisation
-    // path to read back.
-    // Workspace-link nodes get empty children (the linked project
-    // resolves its own deps as a separate importer), `depth = -1` flags
-    // the node for the peer-resolution short-circuit, and the
-    // [`ResolvedPackage`] carries no peer dependencies (peer matching is
-    // the linked importer's responsibility, not the parent's). The node
-    // id is collapsed to a leaf so every reference to the same workspace
-    // path shares one [`NodeId`].
-    let is_link = id.starts_with("link:");
-    let is_leaf = is_link || pkg_is_leaf(&result);
-    let node_id = if is_leaf { NodeId::leaf(&id) } else { NodeId::next() };
-
-    // The claim is taken while holding the `packages` lock so claims and
-    // envelope writes stay in the same order: an occurrence that lost the
-    // claim in between would otherwise leave its peer set on an envelope
-    // whose children the winning occurrence walked.
-    let mut packages = lock_recoverable(&ctx.workspace.packages);
-    let children_owner = claim_children_owner(
-        ctx,
-        &id,
-        depth,
-        ancestor_ids,
-        peer_shadowed_dependencies(
-            result.manifest.as_deref(),
-            parent_pkg_aliases,
-            ctx.workspace.auto_install_peers,
-        ),
-    );
-    let peer_dependencies = if is_link {
-        BTreeMap::new()
-    } else {
-        extract_peer_dependencies(&result, &children_owner.peer_shadowed)
-    };
-    let package_is_new = if let Some(existing) = packages.get_mut(&id) {
-        existing.optional = existing.optional && current_is_optional;
-        // A fresh claim can shadow a different set of dependencies than
-        // the occurrence that populated the envelope did, which moves
-        // names between the package's children and its peers.
-        if children_owner.owns_children && existing.peer_dependencies != peer_dependencies {
-            register_peer_dep_names(ctx, &peer_dependencies);
-            existing.peer_dependencies = peer_dependencies;
-        }
-        false
-    } else {
-        register_peer_dep_names(ctx, &peer_dependencies);
-        packages.insert(
-            id.clone(),
-            ResolvedPackage {
-                id: id.clone(),
-                result: Arc::clone(&result),
-                peer_dependencies,
-                optional: current_is_optional,
-                is_leaf,
-            },
-        );
-        true
-    };
-    drop(packages);
-
-    if package_is_new {
-        ctx.workspace.record_resolved_version(&result);
-        emit_deprecation_if_needed(ctx, &result, &id, depth);
-    }
-
-    let next_ancestors: Vec<String> =
-        ancestor_ids.iter().cloned().chain(std::iter::once(id.clone())).collect();
-    let next_ancestors = Arc::new(next_ancestors);
-
-    Ok(NodeSeed::Pending(Box::new(PendingNode {
-        result,
-        id,
-        alias,
-        node_id,
-        is_link,
-        parent_ancestors: Arc::clone(ancestor_ids),
-        next_ancestors,
-        children_owner,
-        depth,
-        current_is_optional,
-        prior_key,
-    })))
-}
-
-/// The name the edge installs under: the key its parent manifest
-/// declares, else the name the resolver resolved it under, else the
-/// resolved package's own name.
-// The manifest key has to come first: an importer entry is only
-// recorded for an alias the manifest declares, so an edge keyed by a
-// resolver alias that differs from its manifest key is dropped from the
-// lockfile importer.
-fn node_alias(
-    wanted: &WantedDependency,
-    result: &pacquet_resolving_resolver_base::ResolveResult,
-    id: &str,
-) -> String {
-    wanted
-        .alias
-        .clone()
-        .filter(|alias| !alias.is_empty())
-        .or_else(|| result.alias.clone())
-        .or_else(|| result.name_ver.as_ref().map(|name_ver| name_ver.name.to_string()))
-        .unwrap_or_else(|| id.to_string())
-}
-
-/// Walk a seeded node's children. `children_overlay` is the preferred-versions
-/// overlay covering this node's own level (built by the caller from
-/// every sibling seed); the grandchildren's overlay layers this
-/// node's resolved children on top, extending the per-level fold.
-/// `children_pkg_aliases` follows the same shape: the caller folds this
-/// node's whole level into the scope, and the grandchildren's scope
-/// layers this node's resolved children on top.
-///
-/// Only the deterministic children owner walks this package's
-/// manifest children. Other occurrences stay lazy and expand from
-/// `children_by_id`, applying their own `parent_ids` cycle break.
-#[async_recursion]
-async fn walk_node_children<Chain>(
-    ctx: &TreeCtx,
-    resolver: &Chain,
-    pending: PendingNode,
-    children_overlay: Option<Arc<PreferredVersionsOverlay>>,
-    children_pkg_aliases: Arc<ParentPkgAliases>,
-) -> Result<Option<DirectDep>, ResolveDependencyTreeError>
-where
-    Chain: Resolver + ?Sized,
-{
-    let PendingNode {
-        result,
-        id,
-        alias,
-        node_id,
-        is_link,
-        parent_ancestors,
-        next_ancestors,
-        children_owner,
-        depth,
-        current_is_optional,
-        prior_key,
-    } = pending;
-    let children = if is_link {
-        // Linked nodes don't walk their manifest's deps — see the
-        // `is_link` comment block above. They get an empty `Realized`
-        // map: a linked node has no children of its own here.
-        crate::resolved_tree::TreeChildren::Realized(BTreeMap::new())
-    } else if !children_owner.owns_children {
-        crate::resolved_tree::TreeChildren::Lazy {
-            parent_ids: AncestorIds::from(Arc::clone(&parent_ancestors)),
-        }
-    } else {
-        // Look up cached children specs first; only walk the manifest on
-        // a miss. The cache value is held by `Arc` so revisits clone the
-        // refcount instead of the inner `Vec<(String, String, bool)>`.
-        let child_specs = {
-            let cache = lock_recoverable(&ctx.workspace.children_specs_by_id);
-            cache.get(&id).map(Arc::clone)
-        };
-        let child_specs = if let Some(specs) = child_specs {
-            specs
-        } else {
-            let specs = Arc::new(extract_children(&result)?);
-            lock_recoverable(&ctx.workspace.children_specs_by_id)
-                .entry(id.clone())
-                .or_insert_with(|| Arc::clone(&specs));
-            specs
-        };
-        // The specs are cached unfiltered because which of them the
-        // package's own `peerDependencies` shadow is a property of the
-        // owner occurrence, not of the manifest.
-        let peer_shadowed = &children_owner.peer_shadowed;
-        let child_specs: Cow<'_, [ChildSpec]> = if peer_shadowed.is_empty() {
-            Cow::Borrowed(child_specs.as_slice())
-        } else {
-            Cow::Owned(
-                child_specs
-                    .iter()
-                    .filter(|(name, _, optional)| *optional || !peer_shadowed.contains(name))
-                    .cloned()
-                    .collect(),
-            )
-        };
-        let child_specs =
-            if result.resolved_via == "workspace" && result.id.as_str().starts_with("file:") {
-                Cow::Owned(
-                    child_specs
-                        .iter()
-                        .map(|(name, range, optional)| {
-                            resolve_catalog_specifier(name.clone(), range.clone(), &ctx.catalogs)
-                                .map(|(name, range)| (name, range, *optional))
-                        })
-                        .collect::<Result<Vec<_>, _>>()?,
-                )
-            } else {
-                child_specs
-            };
-        // An *updated* parent (one that landed on a different version
-        // than the lockfile recorded, or a new dep) discards its
-        // `resolvedDependencies` child refs, forcing its subtree to
-        // re-resolve. A parent that freshly resolved but landed back on
-        // its previously recorded version keeps the prior child refs
-        // alive (pnpm's non-`parentPkg.updated` arm), and each child
-        // edge re-enters the reuse gate with its recorded key — so a
-        // still-satisfied subtree is reused rather than re-resolved.
-        // Re-resolving those children would re-pick open ranges (`*`)
-        // at their newest versions and churn the lockfile.
-        let prior_children_snapshot = prior_key
-            .as_ref()
-            .filter(|key| landed_on_prior_entry(key, &id))
-            .and_then(|key| ctx.workspace.wanted_lockfile.as_ref()?.snapshots.as_ref()?.get(key));
-        // Phase 1: resolve every child package before any grandchild
-        // walk starts, so the level's resolved versions can feed the
-        // grandchildren's preferred-versions overlay — the
-        // postponed-resolution barrier.
-        // Snapshot this importer's direct-dep versions once for the whole
-        // child fanout instead of locking per edge.
-        let direct_versions = lock_recoverable(&ctx.workspace.direct_dep_versions)
-            .get(&ctx.importer_id)
-            .map(Arc::clone);
-        let declaring_dir = declaring_manifest_dir(ctx, &result);
-        let child_seeds = child_specs
-            .iter()
-            .map(|(child_name, child_range, child_optional)| {
-                let mut child_wanted = WantedDependency {
-                    alias: Some(child_name.clone()),
-                    bare_specifier: Some(child_range.clone()),
-                    optional: Some(*child_optional),
-                    ..WantedDependency::default()
-                };
-                let mut child_prior = prior_children_snapshot
-                    .and_then(|snapshot| prior_child_key(snapshot, child_name, child_range));
-                // Stale-pin refresh: force the edge onto a higher in-range
-                // direct-dep version instead of reusing the pin, so the
-                // pinned version is never resolved or fetched.
-                let forced_version = child_prior
-                    .as_ref()
-                    .and_then(|key| key.suffix.version_semver().cloned())
-                    .zip(child_range.parse::<node_semver::Range>().ok())
-                    .and_then(|(pinned, range)| {
-                        higher_direct_dep_version(
-                            direct_versions.as_deref(),
-                            child_name,
-                            &pinned,
-                            &range,
-                        )
-                    });
-                if let Some(higher) = forced_version {
-                    child_wanted.bare_specifier = Some(higher.to_string());
-                    child_prior = None;
-                }
-                let next_ancestors = Arc::clone(&next_ancestors);
-                let pick_overlay = children_overlay.clone();
-                let declaring_dir = declaring_dir.clone();
-                let parent_pkg_aliases = &children_pkg_aliases;
-                async move {
-                    let seed = resolve_node_seed(
-                        ctx,
-                        resolver,
-                        child_wanted,
-                        &next_ancestors,
-                        depth + 1,
-                        current_is_optional,
-                        ReuseSource::Transitive { key: child_prior },
-                        pick_overlay,
-                        declaring_dir.as_deref(),
-                        parent_pkg_aliases,
-                    )
-                    .await?;
-                    warm_children_resolutions(ctx, resolver, &seed).await;
-                    Ok::<NodeSeed, ResolveDependencyTreeError>(seed)
-                }
-            })
-            .pipe(future::try_join_all)
-            .await?;
-        let grandchild_overlay = PreferredVersionsOverlay::layer(
-            children_overlay.clone(),
-            level_versions(ctx, &child_seeds),
-        );
-        let grandchild_pkg_aliases = children_pkg_aliases.extend(level_aliases(&child_seeds));
-        // Phase 2: walk each child's own children with the extended
-        // overlay.
-        let child_results = child_seeds
-            .into_iter()
-            .map(|seed| {
-                let overlay = grandchild_overlay.clone();
-                let pkg_aliases = Arc::clone(&grandchild_pkg_aliases);
-                async move {
-                    match seed {
-                        NodeSeed::Done(dep) => Ok(dep),
-                        NodeSeed::Pending(pending) => {
-                            walk_node_children(ctx, resolver, *pending, overlay, pkg_aliases).await
-                        }
-                    }
-                }
-            })
-            .pipe(future::try_join_all)
-            .await?;
-        if is_current_children_owner(ctx, &id, &children_owner.owner) {
-            // Build the realized `(alias → NodeId)` map for THIS
-            // occurrence and the per-pkg `children_by_id` entry future
-            // revisits will reuse. `children_by_id` records the resolved
-            // child pkg ids (not NodeIds) plus the `optional` flag so
-            // lazy realisation can thread `current_is_optional` correctly.
-            let mut realized: BTreeMap<String, NodeId> = BTreeMap::new();
-            let mut by_id: Vec<crate::resolved_tree::ChildEdge> = Vec::new();
-            let optional_by_alias: HashMap<&str, bool> =
-                child_specs.iter().map(|(name, _, optional)| (name.as_str(), *optional)).collect();
-            for dep in child_results.into_iter().flatten() {
-                let optional = optional_by_alias.get(dep.alias.as_str()).copied().unwrap_or(false);
-                by_id.push(crate::resolved_tree::ChildEdge {
-                    alias: dep.alias.clone(),
-                    pkg_id: dep.id.clone(),
-                    optional,
-                });
-                realized.insert(dep.alias, dep.node_id);
-            }
-            lock_recoverable(&ctx.workspace.children_by_id).insert(id.clone(), Arc::new(by_id));
-            crate::resolved_tree::TreeChildren::Realized(realized)
-        } else {
-            crate::resolved_tree::TreeChildren::Lazy {
-                parent_ids: AncestorIds::from(Arc::clone(&parent_ancestors)),
-            }
-        }
-    };
-
-    // Repeat-visit leaves collapse onto one tree node; keep the
-    // shallowest depth seen so downstream consumers that read
-    // `tree_node.depth` (the peer pass folds it onto the graph node's
-    // `depth`) take the minimum across visits.
-    // Per-occurrence counter ids are unique by construction, so the
-    // `and_modify` arm is dead for non-leaves.
-    // Linked nodes carry `depth = -1` so the peer-resolution pass
-    // short-circuits them in `resolve_node`.
-    let node_depth = if is_link { -1 } else { depth };
-    remember_node_parent_ids(ctx, &node_id, parent_ancestors);
-    insert_tree_node(ctx, node_id.clone(), &id, children, node_depth);
-    if children_owner.owns_children
-        && !children_owner.children_context_unchanged
-        && is_current_children_owner(ctx, &id, &children_owner.owner)
-    {
-        make_non_owner_nodes_lazy(ctx, &id, &node_id);
-    }
-
-    Ok(Some(DirectDep { alias, node_id, id }))
-}
-
-/// Whether the `parent → child` edge closes a dependency cycle's
-/// *second* lap. The first re-entry of a cycle is kept (so the
-/// cycle-closing dependency edge appears in the tree and the lockfile
-/// snapshot); only the repeat of the full `parent … child` sequence is
-/// dropped.
-pub(crate) fn parent_ids_contain_sequence(
-    pkg_ids: &[String],
-    pkg_id1: &str,
-    pkg_id2: &str,
-) -> bool {
-    let Some(pkg1_index) = pkg_ids.iter().position(|id| id == pkg_id1) else {
-        return false;
-    };
-    if pkg1_index == pkg_ids.len() - 1 {
-        return false;
-    }
-    let Some(pkg2_index) = pkg_ids.iter().rposition(|id| id == pkg_id2) else {
-        return false;
-    };
-    pkg1_index < pkg2_index && pkg2_index != pkg_ids.len() - 1
-}
-
-/// Whether a freshly resolved node landed back on its previously
-/// recorded lockfile entry — the non-`updated` arm, which keeps the
-/// prior child refs alive.
-fn landed_on_prior_entry(prior_key: &PkgNameVerPeer, resolved_pkg_id: &str) -> bool {
-    prior_key.without_peer().to_string() == pacquet_deps_path::remove_suffix(resolved_pkg_id)
-}
-
-/// The package names the npm picker may consult the preferred-versions
-/// overlay under for one wanted edge: the alias itself, plus the real
-/// package name from [`real_package_name_of`] when it differs (the
-/// inner target of an `npm:` alias or the folded `@jsr/...` name of a
-/// `jsr:` specifier) — mirroring the name derivation in the npm
-/// resolver's `parse_bare_specifier`, which keys its overlay merge by
-/// the resolved `spec.name` rather than the outer alias.
-fn overlay_lookup_names(wanted: &WantedDependency) -> Vec<String> {
-    let mut names: Vec<String> = Vec::new();
-    if let Some(alias) = wanted.alias.as_deref()
-        && !alias.is_empty()
-    {
-        names.push(alias.to_string());
-    }
-    if let Some(real_name) = real_package_name_of(wanted)
-        && !names.iter().any(|name| name == real_name.as_ref())
-    {
-        names.push(real_name.into_owned());
-    }
-    names
-}
-
-/// Look the wanted edge up in the per-wanted dedup cache or run the
-/// resolver chain and the manifest-hook pipeline, caching the
-/// `Arc<ResolveResult>` under `cache_key`. Concurrent first-callers
-/// can both miss and resolve in parallel — the resolver's own
-/// per-cache-key fetch locker coalesces the network work, and the
-/// second `or_insert` loses the race harmlessly.
-async fn resolve_wanted_cached<Chain>(
-    ctx: &TreeCtx,
-    resolver: &Chain,
-    wanted: &WantedDependency,
-    opts: &ResolveOptions,
-    pick_overlay: Option<&Arc<PreferredVersionsOverlay>>,
-    cache_key: WantedKey,
-) -> Result<Arc<pacquet_resolving_resolver_base::ResolveResult>, ResolveDependencyTreeError>
-where
-    Chain: Resolver + ?Sized,
-{
-    let cached =
-        lock_recoverable(&ctx.workspace.resolved_by_wanted).get(&cache_key).map(Arc::clone);
-    if let Some(result) = cached {
-        return Ok(result);
-    }
-    // Combine two per-package opts adjustments into one clone. The
-    // `update_requested` flag is scoped per wanted-dependency — true only
-    // when the package's real name (parsed from `bare_specifier` for
-    // npm-aliases, folded from the jsr specifier for jsr deps) is in the
-    // update target list — so the picker's held-back-update warning fires
-    // only for the packages the user actually asked to update.
-    let needs_overlay = !cache_key.8.is_empty();
-    let update_target = cache_key.10;
-    let needs_update = update_target != opts.update_requested;
-    let owned_opts;
-    let opts = if needs_overlay || needs_update {
-        let mut owned = opts.clone();
-        if needs_overlay {
-            owned.preferred_versions_overlay = pick_overlay.map(Arc::clone);
-        }
-        if needs_update {
-            owned.update_requested = update_target;
-        }
-        owned_opts = owned;
-        &owned_opts
-    } else {
-        opts
-    };
-    let mut result = resolver.resolve(wanted, opts).await.map_err(map_resolve_error)?;
-    let Some(result_inner) = result.as_mut() else {
-        return Err(ResolveDependencyTreeError::SpecNotSupported {
-            specifier: render_specifier(wanted),
-        });
-    };
-    if result_inner.manifest.is_none() {
-        result_inner.manifest =
-            Some(Arc::new(fallback_manifest(wanted, opts.current_pkg.as_ref())));
-    }
-    // Apply the configured `readPackageHook` (today:
-    // `packageExtensions`) to the manifest fragment before
-    // anything downstream sees it. The hook clones the inner `Value`
-    // only when it modifies it, so unrelated manifests keep sharing the
-    // resolver's cached `Arc`.
-    if let Some(hook) = ctx.workspace.manifest_hook.as_ref()
-        && let Some(manifest) = result_inner.manifest.take()
-    {
-        result_inner.manifest = Some(hook(manifest));
-    }
-
-    if let Some(pnpmfile_hook) = ctx.workspace.pnpmfile_hook.as_ref()
-        && let Some(manifest) = result_inner.manifest.take()
-    {
-        let log = ctx.workspace.read_package_log.clone().unwrap_or_else(|| Arc::new(|_| {}));
-        // Directory resolutions carry their lockfile-root-relative dir so the
-        // hook can tell a workspace project's dependency instance apart from a
-        // registry manifest — see `HookContext::dir`.
-        let dir = match &result_inner.resolution {
-            pacquet_lockfile::LockfileResolution::Directory(directory_resolution) => {
-                Some(directory_resolution.directory.clone())
-            }
-            _ => None,
-        };
-        let hook_ctx = pacquet_hooks::HookContext { log, dir };
-
-        let updated = pnpmfile_hook
-            .read_package((*manifest).clone(), hook_ctx)
-            .await
-            .map_err(ResolveDependencyTreeError::PnpmfileHook)?;
-        result_inner.manifest = Some(updated);
-    }
-
-    // Overrides run last so a pnpmfile hook that replaced the manifest
-    // cannot erase them — see `WorkspaceTreeCtx::overrides_hook`.
-    if let Some(hook) = ctx.workspace.overrides_hook.as_ref()
-        && let Some(manifest) = result_inner.manifest.take()
-    {
-        result_inner.manifest = Some(hook(manifest));
-    }
-
-    let result = result.expect("Some-guarded above");
-    // Wrap in `Arc` once so the cache, the per-id
-    // `ResolvedPackage` envelope, and the later peer-resolved
-    // graph node share one heap-allocated `ResolveResult`
-    // instead of cloning every `String` field per occurrence.
-    let result = Arc::new(result);
-    lock_recoverable(&ctx.workspace.resolved_by_wanted)
-        .entry(cache_key)
-        .or_insert_with(|| Arc::clone(&result));
-    Ok(result)
-}
-
-/// Stand in for the manifest a resolver didn't supply (pnpm's
-/// `getManifestFromResponse`).
-///
-/// Every consumer downstream of the resolver chain reads the package's
-/// identity off its manifest — most of all
-/// [`build_pkg_id_with_patch_hash`], which has no name to prefix the dep
-/// path with when there is none, leaving a bare `file:<path>` / URL that
-/// keys no `packages:` row. A package with no `package.json` of its own
-/// (an archive that ships none, a directory pin restored from the
-/// lockfile) still has to install, so it borrows the identity the
-/// lockfile recorded for it, or failing that its alias and `0.0.0` — the
-/// version pnpm writes into `packages:` for such a package.
-fn fallback_manifest(
-    wanted: &WantedDependency,
-    current_pkg: Option<&CurrentPkg>,
-) -> pacquet_resolving_resolver_base::DependencyManifest {
-    if let Some(current) = current_pkg
-        && let Some(name) = current.name.as_deref().filter(|name| !name.is_empty())
-        && let Some(version) = current.version.as_deref().filter(|version| !version.is_empty())
-    {
-        return serde_json::json!({ "name": name, "version": version });
-    }
-    let name = match wanted.alias.as_deref().filter(|alias| !alias.is_empty()) {
-        Some(alias) => alias,
-        // A specifier's last path segment is the closest thing to a name
-        // an unaliased dep carries: `file:./no-manifest-1.0.0.tgz` and
-        // `https://host/no-manifest-1.0.0.tgz` both name the archive.
-        None => wanted
-            .bare_specifier
-            .as_deref()
-            .unwrap_or_default()
-            .rsplit('/')
-            .next()
-            .unwrap_or_default(),
-    };
-    serde_json::json!({ "name": name, "version": "0.0.0" })
-}
-
-/// Wrap a resolver-chain failure, keeping the pnpm error code of the ones
-/// that carry one. The chain hands back a type-erased
-/// [`ResolveError`], which drops the `miette::Diagnostic` facet, so the codes
-/// that are part of pnpm's public contract are recovered by downcast; every
-/// other failure keeps the generic envelope.
-fn map_resolve_error(err: ResolveError) -> ResolveDependencyTreeError {
-    let err = match err.downcast::<NoMatchingVersionError>() {
-        Ok(no_matching_version) => {
-            return ResolveDependencyTreeError::NoMatchingVersion(*no_matching_version);
-        }
-        Err(err) => err,
-    };
-    match err.downcast::<RegistryResponseError>() {
-        Ok(response) => ResolveDependencyTreeError::RegistryResponse(*response),
-        Err(err) => ResolveDependencyTreeError::Resolve(err.to_string()),
-    }
-}
-
-/// Speculatively warm a freshly-seeded node's children resolutions so
-/// their packuments download while the sibling level's barrier waits
-/// for its slowest member. Results are discarded — the real picks run
-/// in the walk phase with the level's preferred-versions overlay and
-/// hit the warm metadata caches — and errors are swallowed: a
-/// speculative fetch must never fail the install (the real resolve
-/// will surface it). Recovers the cross-level pipelining the
-/// postponed-resolution barrier otherwise serializes; pure overlap,
-/// no behavioral effect.
-async fn warm_children_resolutions<Chain>(ctx: &TreeCtx, resolver: &Chain, seed: &NodeSeed)
-where
-    Chain: Resolver + ?Sized,
-{
-    // A configured pnpmfile hook is externally observable per call
-    // (`readPackage` IPC, `context.log`, custom resolvers), so
-    // speculative resolutions must not fire it; the pure in-memory
-    // manifest hook (packageExtensions / overrides) is idempotent and
-    // cache-deduped, indistinguishable from a first-caller win in the
-    // pre-existing concurrent-miss race.
-    if ctx.workspace.pnpmfile_hook.is_some() {
-        return;
-    }
-    let NodeSeed::Pending(pending) = seed else { return };
-    if pending.is_link || !pending.children_owner.owns_children {
-        return;
-    }
-    let Ok(specs) = extract_children(&pending.result) else { return };
-    let opts = ctx.opts_for_depth(pending.depth + 1);
-    let declaring_dir = declaring_manifest_dir(ctx, &pending.result);
-    specs
-        .iter()
-        .filter(|(name, _, optional)| {
-            *optional || !pending.children_owner.peer_shadowed.contains(name)
-        })
-        .map(|(name, range, optional)| {
-            let wanted = WantedDependency {
-                alias: Some(name.clone()),
-                bare_specifier: Some(range.clone()),
-                optional: Some(*optional),
-                ..WantedDependency::default()
-            };
-            let opts = opts_relative_to_declaring_manifest(opts, &wanted, declaring_dir.as_deref());
-            async move {
-                let opts = opts.as_ref();
-                // Warm through the same per-wanted dedup cache, under
-                // the empty-overlay-view key: when the real pick's
-                // view is empty too (the overwhelmingly common case)
-                // it reuses this entry outright; otherwise it misses
-                // into its own bucket and re-picks from the warm
-                // metadata caches.
-                let project_scope = project_relative_cache_scope(&wanted, opts);
-                let cache_key: WantedKey = (
-                    wanted.alias.clone(),
-                    wanted.bare_specifier.clone(),
-                    wanted.optional,
-                    wanted.injected,
-                    opts.pick_lowest_version,
-                    opts.published_by,
-                    project_scope,
-                    // No prior-lockfile key: a warm entry must only be
-                    // reused by edges that carry no currentPkg either.
-                    None,
-                    Vec::new(),
-                    ctx.update_cache_scope(),
-                    is_update_target(ctx.update_scope(), &wanted, pending.depth + 1),
-                );
-                let _ = resolve_wanted_cached(ctx, resolver, &wanted, opts, None, cache_key).await;
-            }
-        })
-        .pipe(future::join_all)
-        .await;
-}
-
-/// The install aliases one resolved level contributes to its
-/// children's [`ParentPkgAliases`] scope. An edge the walk dropped (a
-/// cycle re-entry, a skipped optional) contributes nothing, matching
-/// pnpm's fold over the level's resolved addresses.
-fn level_aliases(seeds: &[NodeSeed]) -> HashSet<String> {
-    seeds
-        .iter()
-        .filter_map(|seed| match seed {
-            NodeSeed::Pending(pending) => Some(pending.alias.clone()),
-            NodeSeed::Done(Some(dep)) => Some(dep.alias.clone()),
-            NodeSeed::Done(None) => None,
-        })
-        .collect()
-}
-
-/// The `(name → versions)` additions one resolved level contributes
-/// to its children's preferred-versions overlay. Linked nodes carry no
-/// `name_ver` and contribute nothing — they're skipped in the fold.
-fn level_versions(ctx: &TreeCtx, seeds: &[NodeSeed]) -> BTreeMap<String, Vec<String>> {
-    let packages = lock_recoverable(&ctx.workspace.packages);
-    let mut level: BTreeMap<String, Vec<String>> = BTreeMap::new();
-    for seed in seeds {
-        let name_ver = match seed {
-            NodeSeed::Pending(pending) => pending.result.name_ver.as_ref(),
-            NodeSeed::Done(Some(dep)) => {
-                packages.get(&dep.id).and_then(|pkg| pkg.result.name_ver.as_ref())
-            }
-            NodeSeed::Done(None) => None,
-        };
-        let Some(name_ver) = name_ver else { continue };
-        let versions = level.entry(name_ver.name.to_string()).or_default();
-        let version = name_ver.suffix.to_string();
-        if !versions.contains(&version) {
-            versions.push(version);
-        }
-    }
-    level
-}
-
-/// Record the importer direct deps whose manifest specifier differs from
-/// the prior lockfile's recorded specifier (a new dep counts as changed).
-/// See [`WorkspaceTreeCtx::changed_direct_deps`].
-///
-/// Called for the importer's *manifest* wave only, never for
-/// auto-installed peers: hoisted peers have no importer-snapshot entry
-/// to compare against, so recording them would mark them "changed" on
-/// every install and permanently decline subtree reuse for anything
-/// depending on them.
-pub(crate) fn record_changed_direct_deps(
-    ctx: &TreeCtx,
-    importer_id: &str,
-    wanted: &[WantedSpec],
-) -> HashSet<PkgName> {
-    let lockfile = ctx.workspace.wanted_lockfile.as_deref();
-    let prior = lockfile.and_then(|lockfile| lockfile.importers.get(importer_id));
-    let mut changed = lock_recoverable(&ctx.workspace.changed_direct_deps);
-    let bucket = changed.entry(importer_id.to_string()).or_default();
-    for (alias, spec, _optional, _injected) in wanted {
-        let unchanged = prior
-            .and_then(|importer| importer_dep_specifier(importer, alias))
-            .is_some_and(|recorded| {
-                recorded == spec || catalog_specifier_unchanged(lockfile, recorded, alias, spec)
-            });
-        if !unchanged && let Ok(name) = alias.parse::<PkgName>() {
-            bucket.insert(name);
-        }
-    }
-    bucket.clone()
-}
-
-/// Whether a `catalog:`-recorded direct dep still resolves to the same
-/// underlying range. The wanted specs reaching
-/// [`fn@record_changed_direct_deps`] have their `catalog:` protocol
-/// already replaced by the catalog's range
-/// ([`fn@resolve_catalog_specifiers`]), while the importer snapshot
-/// records the literal `catalog:` / `catalog:<name>` form — comparing
-/// those directly would mark every catalog-managed dep as changed on
-/// every install, and the changed-direct-dep reuse gate would then
-/// re-resolve (and drift) every subtree depending on one. The edge is
-/// unchanged when the lockfile's `catalogs:` snapshot recorded the same
-/// range for this alias.
-fn catalog_specifier_unchanged(
-    lockfile: Option<&pacquet_lockfile::Lockfile>,
-    recorded: &str,
-    alias: &str,
-    resolved_spec: &str,
-) -> bool {
-    let Some(catalog_name) = recorded.strip_prefix("catalog:") else {
-        return false;
-    };
-    let catalog_name = if catalog_name.is_empty() { "default" } else { catalog_name };
-    lockfile
-        .and_then(|lockfile| lockfile.catalogs.as_ref())
-        .and_then(|catalogs| catalogs.get(catalog_name))
-        .and_then(|catalog| catalog.get(alias))
-        .is_some_and(|entry| entry.specifier == resolved_spec)
-}
-
-/// The recorded specifier for direct-dep `alias` across the importer's
-/// prod / dev / optional dependency maps in the prior lockfile.
-fn importer_dep_specifier<'a>(importer: &'a ProjectSnapshot, alias: &str) -> Option<&'a str> {
-    let name: PkgName = alias.parse().ok()?;
-    let lookup = |map: Option<&'a ResolvedDependencyMap>| map.and_then(|deps| deps.get(&name));
-    lookup(importer.dependencies.as_ref())
-        .or_else(|| lookup(importer.optional_dependencies.as_ref()))
-        .or_else(|| lookup(importer.dev_dependencies.as_ref()))
-        .map(|dep| dep.specifier.as_str())
-}
-
-/// Store the importer's resolved (parsed) direct-dep versions for the
-/// per-edge stale-pin refresh. See [`WorkspaceTreeCtx::direct_dep_versions`].
-fn record_direct_dep_versions(
-    ctx: &TreeCtx,
-    importer_id: &str,
-    level: &BTreeMap<String, Vec<String>>,
-) {
-    let mut versions = lock_recoverable(&ctx.workspace.direct_dep_versions);
-    let by_name = Arc::make_mut(versions.entry(importer_id.to_string()).or_default());
-    for (name, level_versions) in level {
-        let bucket = by_name.entry(name.clone()).or_default();
-        for version in level_versions {
-            let Ok(parsed) = version.parse::<node_semver::Version>() else { continue };
-            if !bucket.contains(&parsed) {
-                bucket.push(parsed);
-            }
-        }
-    }
-}
-
-/// True when `snapshot` depends on one of this importer's changed direct
-/// deps (see [`WorkspaceTreeCtx::changed_direct_deps`]).
-fn reused_parent_has_changed_direct_child(ctx: &TreeCtx, snapshot: &SnapshotEntry) -> bool {
-    // Copy the (small) changed set out and drop the lock before scanning.
-    let importer_changed = {
-        let changed = lock_recoverable(&ctx.workspace.changed_direct_deps);
-        match changed.get(&ctx.importer_id) {
-            Some(set) if !set.is_empty() => set.clone(),
-            _ => return false,
-        }
-    };
-    let depends_on = |map: Option<&std::collections::HashMap<PkgName, SnapshotDepRef>>| {
-        map.is_some_and(|deps| deps.keys().any(|name| importer_changed.contains(name)))
-    };
-    depends_on(snapshot.dependencies.as_ref())
-        || depends_on(snapshot.optional_dependencies.as_ref())
-}
-
-/// Reuse-decline gate: whether `prior_key`'s prior snapshot depends on a
-/// changed direct dep.
-fn node_depends_on_changed_direct_dep(ctx: &TreeCtx, prior_key: Option<&PkgNameVerPeer>) -> bool {
-    prior_key
-        .and_then(|key| ctx.workspace.wanted_lockfile.as_ref()?.snapshots.as_ref()?.get(key))
-        .is_some_and(|snapshot| reused_parent_has_changed_direct_child(ctx, snapshot))
-}
-
-/// The highest resolved direct-dependency version of `name` strictly
-/// above `pinned` that still satisfies `range`, or `None`. Anchored to
-/// direct deps (the deterministic, resolved-before-the-walk signal).
-/// `direct_versions` is the importer's snapshot, taken once per walk by
-/// [`fn@walk_node_children`].
-fn higher_direct_dep_version(
-    direct_versions: Option<&DirectDepVersions>,
-    name: &str,
-    pinned: &node_semver::Version,
-    range: &node_semver::Range,
-) -> Option<node_semver::Version> {
-    direct_versions?
-        .get(name)?
-        .iter()
-        // Plain semver satisfaction (not prerelease-inclusive): a
-        // prerelease direct dep only refreshes an edge whose range admits
-        // prereleases.
-        .filter(|&version| version > pinned && range.satisfies(version))
-        .max()
-        .cloned()
-}
-
-/// One reusable node: its prior-lockfile snapshot key plus the
-/// `ResolveResult` synthesized from the lockfile metadata.
-struct ReusedNode {
-    key: PkgNameVerPeer,
-    result: pacquet_resolving_resolver_base::ResolveResult,
-}
-
-struct ChildrenOwnerClaim {
-    owner: ChildrenOwner,
-    owns_children: bool,
-    /// The shadowed-dependency set in force for the package id — this
-    /// occurrence's when it won the claim, the standing owner's when it
-    /// lost. See [`ChildrenOwnerEntry::peer_shadowed`].
-    peer_shadowed: Arc<HashSet<String>>,
-    /// A winning claim displaced an owner whose shadowed-dependency set
-    /// equals this occurrence's. Children resolve identically under
-    /// both, so the other occurrences' realized children stay valid and
-    /// the winner skips the lazy-flip (and the engine-invalidating
-    /// rewrite signal) it would otherwise broadcast.
-    children_context_unchanged: bool,
-}
-
-/// `peer_shadowed` is this occurrence's own set; it is installed as the
-/// package id's set only when the occurrence wins the claim, so the
-/// losing occurrences of a concurrent claim all read back the winner's.
-fn claim_children_owner(
-    ctx: &TreeCtx,
-    pkg_id: &str,
-    depth: i32,
-    ancestor_ids: &[String],
-    peer_shadowed: HashSet<String>,
-) -> ChildrenOwnerClaim {
-    let owner = ChildrenOwner {
-        update_active: !matches!(ctx.update_reuse_scope(), UpdateReuseScope::All),
-        depth,
-        importer_order: ctx.importer_order,
-        parent_path: ancestor_ids.to_vec(),
-        importer_id: ctx.importer_id.clone(),
-    };
-    let (owns_children, peer_shadowed, children_context_unchanged) = {
-        let mut owners = lock_recoverable(&ctx.workspace.children_owner_by_id);
-        match owners.get(pkg_id) {
-            Some(existing) if !owner.wins_over(&existing.owner) => {
-                (false, Arc::clone(&existing.peer_shadowed), false)
-            }
-            existing => {
-                let children_context_unchanged =
-                    existing.is_some_and(|entry| *entry.peer_shadowed == peer_shadowed);
-                let peer_shadowed = Arc::new(peer_shadowed);
-                owners.insert(
-                    pkg_id.to_string(),
-                    ChildrenOwnerEntry {
-                        owner: owner.clone(),
-                        peer_shadowed: Arc::clone(&peer_shadowed),
-                    },
-                );
-                (true, peer_shadowed, children_context_unchanged)
-            }
-        }
-    };
-    if owns_children {
-        lock_recoverable(&ctx.workspace.first_importer_by_pkg)
-            .insert(pkg_id.to_string(), owner.importer_id.clone());
-    }
-    ChildrenOwnerClaim { owner, owns_children, peer_shadowed, children_context_unchanged }
-}
-
-/// Seed the peer-walker's `parentPkgs` filter with the names a
-/// resolved package declares as peers.
-fn register_peer_dep_names(ctx: &TreeCtx, peer_dependencies: &BTreeMap<String, PeerDep>) {
-    let mut all_peers = lock_recoverable(&ctx.workspace.all_peer_dep_names);
-    for name in peer_dependencies.keys() {
-        all_peers.insert(name.clone());
-    }
-}
-
-fn is_current_children_owner(ctx: &TreeCtx, pkg_id: &str, owner: &ChildrenOwner) -> bool {
-    lock_recoverable(&ctx.workspace.children_owner_by_id)
-        .get(pkg_id)
-        .is_some_and(|current| current.owner == *owner)
-}
-
-fn remember_node_parent_ids(ctx: &TreeCtx, node_id: &NodeId, parent_ids: Arc<Vec<String>>) {
-    lock_recoverable(&ctx.workspace.node_parent_ids_by_id).insert(node_id.clone(), parent_ids);
-}
-
-/// Record an occurrence node in the shared tree (lowering the depth of
-/// a revisited leaf) and, on first insertion, in the per-package
-/// reverse index [`fn@make_non_owner_nodes_lazy`] flips through.
-fn insert_tree_node(
-    ctx: &TreeCtx,
-    node_id: NodeId,
-    pkg_id: &str,
-    children: crate::resolved_tree::TreeChildren,
-    depth: i32,
-) {
-    let inserted = match lock_recoverable(&ctx.workspace.dependencies_tree).entry(node_id.clone()) {
-        std::collections::hash_map::Entry::Occupied(mut entry) => {
-            if entry.get().depth > depth {
-                entry.get_mut().depth = depth;
-            }
-            false
-        }
-        std::collections::hash_map::Entry::Vacant(entry) => {
-            entry.insert(DependenciesTreeNode::new(pkg_id.to_string(), children, depth, true));
-            true
-        }
-    };
-    if inserted {
-        lock_recoverable(&ctx.workspace.nodes_by_pkg_id)
-            .entry(pkg_id.to_string())
-            .or_default()
-            .push(node_id);
-    }
-}
-
-fn make_non_owner_nodes_lazy(ctx: &TreeCtx, pkg_id: &str, owner_node_id: &NodeId) {
-    let pkg_nodes = match lock_recoverable(&ctx.workspace.nodes_by_pkg_id).get(pkg_id) {
-        Some(nodes) => nodes.clone(),
-        None => return,
-    };
-    // Collect the parent chains first so the two locks are never held
-    // together.
-    let parent_ids_by_node: Vec<(NodeId, Arc<Vec<String>>)> = {
-        let parent_ids = lock_recoverable(&ctx.workspace.node_parent_ids_by_id);
-        pkg_nodes
-            .into_iter()
-            .filter(|node_id| node_id != owner_node_id)
-            .filter_map(|node_id| {
-                let ids = Arc::clone(parent_ids.get(&node_id)?);
-                Some((node_id, ids))
-            })
-            .collect()
-    };
-    let mut tree = lock_recoverable(&ctx.workspace.dependencies_tree);
-    let mut rewrote_any = false;
-    for (node_id, parent_ids) in parent_ids_by_node {
-        if let Some(node) = tree.get_mut(&node_id) {
-            node.children = crate::resolved_tree::TreeChildren::Lazy {
-                parent_ids: AncestorIds::from(parent_ids),
-            };
-            rewrote_any = true;
-        }
-    }
-    if rewrote_any {
-        ctx.workspace.record_children_rewrite();
-    }
-}
-
-/// Decide whether the current edge can reuse the prior lockfile's
-/// resolution. `prior_key` is the edge's recorded snapshot key (see
-/// [`ReuseSource::prior_key`]). Returns the synthesized node when the
-/// edge's whole transitive subtree is reusable; `None` (fresh resolve)
-/// otherwise.
-///
-/// Conservative on every axis: no prior lockfile, no recorded key, a
-/// `link:` / non-registry shape anywhere in the subtree, or a missing
-/// snapshot entry all yield `None`. See [`fn@subtree_fully_reusable`]
-/// for the recursive subtree check.
-fn try_reuse_node(
-    ctx: &TreeCtx,
-    wanted: &WantedDependency,
-    prior_key: Option<&PkgNameVerPeer>,
-    depth: i32,
-) -> Option<ReusedNode> {
-    let lockfile = ctx.workspace.wanted_lockfile.as_ref()?;
-    let scope = ctx.update_scope();
-    if matches!(scope.reuse, UpdateReuseScope::None) && scope.max_depth.reaches(depth) {
-        return None;
-    }
-    let alias = wanted.alias.as_deref()?;
-    let key = prior_key?;
-    if !subtree_fully_reusable(ctx, lockfile, key, depth) {
-        return None;
-    }
-    let result = synthesize_reused_result(lockfile, key, alias)?;
-    Some(ReusedNode { key: key.clone(), result })
-}
-
-/// `true` when a node named `name` at `depth` is a `pacquet update`
-/// target, and so excluded from reuse. Past the `--depth` ceiling the
-/// update no longer reaches, so every node keeps its locked resolution.
-fn update_excludes(scope: UpdateScope<'_>, name: &str, depth: i32) -> bool {
-    if !scope.max_depth.reaches(depth) {
-        return false;
-    }
-    match scope.reuse {
-        UpdateReuseScope::All => false,
-        // `None` is handled earlier in `try_reuse_node`; treat it the
-        // same here for completeness.
-        UpdateReuseScope::None => true,
-        UpdateReuseScope::Except(names) => names.contains(name),
-    }
-}
-
-/// Whether the wanted lockfile already holds a package entry that
-/// satisfies the wanted dependency. An optional dependency that fails
-/// to resolve is normally skipped, but when a locked resolution exists
-/// the failure is environmental (e.g. a registry mirror that hasn't
-/// synced the release yet) rather than a genuinely uninstallable
-/// package. Silently skipping in that case would erase the locked
-/// entries, making the lockfile differ across machines from identical
-/// inputs and leaving frozen installs on other hosts with nothing to
-/// link (<https://github.com/pnpm/pnpm/issues/12853>).
-///
-/// Only plain semver specifiers are checked; exotic specifiers (git,
-/// catalogs, tags, URLs) keep the skip-on-failure behavior.
-///
-/// The check is deliberately by package name and range rather than by
-/// the current edge's locked snapshot ref: a satisfying entry locked
-/// via any edge means the registry served this package in-range
-/// before, so failing loudly instead of skipping is the right outcome
-/// even when the failing edge itself was never locked. This matches
-/// the TypeScript resolver's `wantedLockfileContainsSatisfyingEntry`.
-fn wanted_lockfile_contains_satisfying_entry(
-    lockfile: Option<&pacquet_lockfile::Lockfile>,
-    wanted: &WantedDependency,
-) -> bool {
-    let Some(packages) = lockfile.and_then(|lockfile| lockfile.packages.as_ref()) else {
-        return false;
-    };
-    let Some(alias) = wanted.alias.as_deref().filter(|alias| !alias.is_empty()) else {
-        return false;
-    };
-    let (pkg_name, range) =
-        unwrap_package_name(alias, wanted.bare_specifier.as_deref().unwrap_or_default());
-    let Ok(range) = range.parse::<node_semver::Range>() else {
-        return false;
-    };
-    let Ok(pkg_name) = PkgName::parse(pkg_name) else {
-        return false;
-    };
-    packages.keys().any(|key| {
-        key.name == pkg_name
-            && key.suffix.version_semver().is_some_and(|version| range.satisfies(version))
-    })
-}
-
-/// Map an ancestor-id chain to the `parents` payload of a
-/// skipped-optional-dependency notification, resolving each
-/// `pkgIdWithPatchHash` through the shared packages map (the
-/// counterpart of pnpm's `getPkgsInfoFromIds`).
-fn pkgs_info_from_ids(
-    ctx: &TreeCtx,
-    ancestor_ids: &[String],
-) -> Vec<SkippedOptionalDependencyParent> {
-    let packages = lock_recoverable(&ctx.workspace.packages);
-    ancestor_ids
-        .iter()
-        .map(|id| {
-            let name_ver = packages.get(id).and_then(|pkg| pkg.result.name_ver.as_ref());
-            SkippedOptionalDependencyParent {
-                id: id.clone(),
-                name: name_ver.map(|name_ver| name_ver.name.to_string()).unwrap_or_default(),
-                version: name_ver.map(|name_ver| name_ver.suffix.to_string()).unwrap_or_default(),
-            }
-        })
-        .collect()
-}
-
-/// Normalize an `npm:` alias specifier into the real package name and
-/// the wanted range (`("is-positive", "^1.0.0")` for the edge
-/// `my-alias@npm:is-positive@^1.0.0`; the range is `"*"` for the
-/// spec-less `npm:is-positive` form). A specifier without the `npm:`
-/// prefix is returned as-is: the alias is the real package name.
-///
-/// Mirrors the TypeScript resolver's `unwrapPackageName`; unlike
-/// [`fn@real_package_name_of`] it also yields the range and does not
-/// special-case the `npm:<range>` form, so the locked-entry check
-/// matches its TypeScript counterpart byte for byte.
-pub(crate) fn unwrap_package_name<'a>(
-    alias: &'a str,
-    bare_specifier: &'a str,
-) -> (&'a str, &'a str) {
-    let Some(rest) = bare_specifier.strip_prefix("npm:") else {
-        return (alias, bare_specifier);
-    };
-    match rest.rfind('@') {
-        None | Some(0) => (rest, "*"),
-        Some(index) => (&rest[..index], &rest[index + 1..]),
-    }
-}
-
-/// Resolve the *real* package name a [`WantedDependency`] targets —
-/// i.e. the name update targeting matches against, not the local
-/// install alias. For a plain dep (`foo@^1`) or a workspace dep, this is just
-/// `wanted.alias`. For an npm-alias dep (`foo@npm:bar@^4`), this is
-/// `bar` — parsed out of `bare_specifier` because `wanted.alias` keeps
-/// the local install name `foo`. For a jsr dep (`foo@jsr:@bar/baz@^1`
-/// or just `jsr:@bar/baz`), this is the folded npm registry name
-/// (`@jsr/bar__baz`) that the picker and lockfile snapshots key on.
-/// [`overlay_lookup_names`] builds its candidate set from this name.
-///
-/// Returns `None` when no name can be recovered (no alias, malformed
-/// npm-alias target, unparsable jsr specifier). The caller treats
-/// `None` as "not a targeted update" since update targets are keyed
-/// by package name.
-fn real_package_name_of(wanted: &WantedDependency) -> Option<Cow<'_, str>> {
-    let bare = wanted.bare_specifier.as_deref()?;
-    if let Some(rest) = bare.strip_prefix("npm:") {
-        // `npm:<range>` form: the body is a semver range, not a name,
-        // so the install alias IS the real package name.
-        let alias_keeps_name = wanted
-            .alias
-            .as_deref()
-            .is_some_and(|alias| !alias.is_empty() && rest.parse::<node_semver::Range>().is_ok());
-        if !alias_keeps_name {
-            // `npm:<name>@<range>` or `npm:<name>`: parse the real name
-            // out of `bare_specifier` because `wanted.alias` keeps the
-            // local install name. Split at the last `@` to separate the
-            // real name from the version range. When there is no `@`,
-            // the whole `rest` is the name (default-tag form `npm:bar`).
-            let last_at =
-                rest.bytes().enumerate().rev().find_map(|(i, b)| (b == b'@').then_some(i));
-            let name = match last_at {
-                Some(idx) if idx >= 1 => &rest[..idx],
-                _ => rest,
-            };
-            return (!name.is_empty()).then_some(Cow::Borrowed(name));
-        }
-    }
-    if bare.starts_with("jsr:") {
-        // An unparsable `jsr:` specifier carries no recoverable real
-        // name — return `None` rather than falling back to the install
-        // alias, which could accidentally match an update target and
-        // mark the broken dep as a targeted update.
-        let spec = pacquet_resolving_jsr_specifier_parser::parse_jsr_specifier(
-            bare,
-            wanted.alias.as_deref(),
-        )
-        .ok()
-        .flatten()?;
-        return Some(Cow::Owned(spec.npm_pkg_name));
-    }
-    // Plain dep, workspace dep, or `npm:<range>` form: the install
-    // alias is the real name.
-    wanted.alias.as_deref().map(Cow::Borrowed)
-}
-
-/// Whether `wanted` is one of the packages the user asked to update,
-/// given the install's [`UpdateReuseScope`]. Feeds the per-resolve
-/// `ResolveOptions::update_requested` flag, which gates the npm
-/// picker's held-back-update warning.
-///
-/// Returns `false` for `All`/`None` scopes (no individually targeted
-/// packages), for a node past the `--depth` ceiling, and for `Except`
-/// scopes where the wanted package's real name is not in the target
-/// list. Returns `true` only when the real name is in the `Except` set.
-#[inline]
-fn is_update_target(scope: UpdateScope<'_>, wanted: &WantedDependency, depth: i32) -> bool {
-    if !scope.max_depth.reaches(depth) {
-        return false;
-    }
-    match scope.reuse {
-        UpdateReuseScope::All | UpdateReuseScope::None => false,
-        UpdateReuseScope::Except(_) => {
-            real_package_name_of(wanted).is_some_and(|n| update_excludes(scope, n.as_ref(), depth))
-        }
-    }
-}
-
-/// `true` when `key` and its entire transitive subtree can be
-/// synthesized from `lockfile` (every node a plain-semver registry
-/// package present in `packages:`, every snapshot child non-`link:`).
-/// Memoised on [`WorkspaceTreeCtx::subtree_reusable`] so each package is
-/// checked once.
-///
-/// A snapshot cycle is treated as **non**-reusable at the back-edge: the
-/// key is provisionally inserted as `false` before recursing, so a node
-/// reached through a still-in-progress ancestor resolves to `false` and
-/// any subtree containing a dependency cycle conservatively re-resolves.
-/// This avoids the unsound alternative — a provisional `true` could cache
-/// a cycle member as reusable based on an ancestor that later finalizes
-/// `false` (e.g. an update-excluded target reachable only through the
-/// cycle), wrongly reusing it. SCC-aware reuse of acyclic-equivalent
-/// cycles is possible but not worth the complexity for an uncommon case.
-fn subtree_fully_reusable(
-    ctx: &TreeCtx,
-    lockfile: &pacquet_lockfile::Lockfile,
-    key: &PkgNameVerPeer,
-    depth: i32,
-) -> bool {
-    let scope = ctx.update_scope();
-    let memo_key = (ctx.update_cache_scope(), key.clone(), scope.max_depth.memo_bucket(depth));
-    if let Some(&cached) = lock_recoverable(&ctx.workspace.subtree_reusable).get(&memo_key) {
-        return cached;
-    }
-    // Provisionally mark non-reusable so a cycle back to `key` resolves to
-    // `false` (re-resolve) instead of recursing forever — see the doc above
-    // for why `false` rather than `true`.
-    lock_recoverable(&ctx.workspace.subtree_reusable).insert(memo_key.clone(), false);
-    // A `pacquet update` target anywhere in the subtree forces the whole
-    // subtree to re-resolve so the bump's new transitive deps are picked
-    // up — update names match at every depth the update reaches.
-    let name = key.name.to_string();
-    let reusable = !update_excludes(scope, &name, depth)
-        && synthesize_reused_result(lockfile, key, &name).is_some()
-        && subtree_children_reusable(ctx, lockfile, key, depth);
-    lock_recoverable(&ctx.workspace.subtree_reusable).insert(memo_key, reusable);
-    reusable
-}
-
-/// Recurse [`fn@subtree_fully_reusable`] across `key`'s snapshot
-/// children. A `link:` child (no snapshot key) makes the subtree
-/// non-reusable: the linked importer resolves its own deps, which this
-/// reuse path doesn't model.
-fn subtree_children_reusable(
-    ctx: &TreeCtx,
-    lockfile: &pacquet_lockfile::Lockfile,
-    key: &PkgNameVerPeer,
-    depth: i32,
-) -> bool {
-    let Some(snapshot) = lockfile.snapshots.as_ref().and_then(|snaps| snaps.get(key)) else {
-        // No snapshot entry → the lockfile doesn't record this node's
-        // children, so the reuse walk can't reproduce its subtree.
-        // Force a fresh resolve rather than risk silently dropping
-        // transitive deps. A genuine leaf has an empty-but-*present*
-        // snapshot entry (`{}`); a missing one means an inconsistent
-        // lockfile, which `try_reuse_node`'s contract sends to a fresh
-        // resolve.
-        return false;
-    };
-    let dep_maps = [snapshot.dependencies.as_ref(), snapshot.optional_dependencies.as_ref()];
-    for dep_map in dep_maps.into_iter().flatten() {
-        for (child_name, dep_ref) in dep_map {
-            let Some(child_key) = dep_ref.resolve(child_name) else {
-                return false;
-            };
-            if !subtree_fully_reusable(ctx, lockfile, &child_key, depth + 1) {
-                return false;
-            }
-        }
-    }
-    true
-}
-
-/// Register a node whose resolution was reused from the prior lockfile,
-/// then walk its transitive children from the snapshot graph instead of
-/// re-resolving them. Mirrors the post-resolve half of
-/// [`fn@resolve_node`], specialized for a node whose subtree
-/// [`fn@try_reuse_node`] already confirmed reusable.
-#[async_recursion]
-#[expect(
-    clippy::too_many_arguments,
-    reason = "internal walker helper threading per-node context through the recursion"
-)]
-async fn resolve_reused_node<Chain>(
-    ctx: &TreeCtx,
-    resolver: &Chain,
-    wanted: WantedDependency,
-    ancestor_ids: &Arc<Vec<String>>,
-    depth: i32,
-    current_is_optional: bool,
-    reused: ReusedNode,
-    parent_pkg_aliases: &Arc<ParentPkgAliases>,
-) -> Result<Option<DirectDep>, ResolveDependencyTreeError>
-where
-    Chain: Resolver + ?Sized,
-{
-    let ReusedNode { key, result } = reused;
-    let result = Arc::new(result);
-
-    // The synthesized result must stay out of `resolved_by_wanted`: its
-    // manifest deliberately omits `dependencies` (a reused node's
-    // children come from the snapshot graph), so if the fresh-resolve
-    // path ever read it back — e.g. an identical edge denied reuse by
-    // the changed-direct-dep gate or by `subtree_fully_reusable`'s
-    // provisional-`false` cycle guard — `extract_children` /
-    // `pkg_is_leaf` would misread the package as dependency-less and
-    // record it as a leaf, emptying its lockfile snapshot.
-
-    let id = build_pkg_id_with_patch_hash(ctx, &result).await?;
-
-    // Cycle break — same as the fresh path.
-    if ancestor_ids.last().is_some_and(|parent| {
-        *parent == id || parent_ids_contain_sequence(ancestor_ids, parent, &id)
-    }) {
-        return Ok(None);
-    }
-
-    let alias = node_alias(&wanted, &result, &id);
-
-    // Leaf classification reads the snapshot graph (the source of truth
-    // for a reused node's children), not the synthesized manifest (whose
-    // `dependencies` are deliberately omitted). A node with no recorded
-    // children and no peers is a leaf, matching `pkg_is_leaf`.
-    let snapshot = ctx
-        .workspace
-        .wanted_lockfile
-        .as_ref()
-        .and_then(|lockfile| lockfile.snapshots.as_ref())
-        .and_then(|snaps| snaps.get(&key));
-    // A reused node's children come from the snapshot rather than from
-    // a manifest, so no `dependencies` entry of its synthesized
-    // manifest can be peer-shadowed.
-    let peer_dependencies = extract_peer_dependencies(&result, &HashSet::default());
-    let child_refs = snapshot_child_refs(snapshot, &peer_dependencies);
-    let is_leaf = child_refs.is_empty() && peer_dependencies.is_empty();
-    let node_id = if is_leaf { NodeId::leaf(&id) } else { NodeId::next() };
-
-    let package_is_new = {
-        let mut packages = lock_recoverable(&ctx.workspace.packages);
-        if let Some(existing) = packages.get_mut(&id) {
-            existing.optional = existing.optional && current_is_optional;
-            false
-        } else {
-            {
-                let mut all_peers = lock_recoverable(&ctx.workspace.all_peer_dep_names);
-                for name in peer_dependencies.keys() {
-                    all_peers.insert(name.clone());
-                }
-            }
-            packages.insert(
-                id.clone(),
-                ResolvedPackage {
-                    id: id.clone(),
-                    result: Arc::clone(&result),
-                    peer_dependencies,
-                    optional: current_is_optional,
-                    is_leaf,
-                },
-            );
-            true
-        }
-    };
-
-    if package_is_new {
-        ctx.workspace.record_resolved_version(&result);
-        emit_deprecation_if_needed(ctx, &result, &id, depth);
-    }
-
-    let next_ancestors: Vec<String> =
-        ancestor_ids.iter().cloned().chain(std::iter::once(id.clone())).collect();
-    let next_ancestors = Arc::new(next_ancestors);
-    let children_owner = claim_children_owner(ctx, &id, depth, ancestor_ids, HashSet::default());
-
-    let children = if children_owner.owns_children {
-        let child_results = child_refs
-            .iter()
-            .map(|(child_alias, child_key)| {
-                let child_wanted = WantedDependency {
-                    alias: Some(child_alias.clone()),
-                    // The snapshot pins the exact version; carry it as
-                    // the bare specifier so the per-wanted dedup cache
-                    // key is stable and a fresh fallback (if reuse were
-                    // ever disabled) would still target the right pin.
-                    bare_specifier: Some(child_key.suffix.without_peer().to_string()),
-                    ..WantedDependency::default()
-                };
-                let next_ancestors = Arc::clone(&next_ancestors);
-                let child_key = child_key.clone();
-                async move {
-                    resolve_node(
-                        ctx,
-                        resolver,
-                        child_wanted,
-                        &next_ancestors,
-                        depth + 1,
-                        current_is_optional,
-                        ReuseSource::Transitive { key: Some(child_key) },
-                        parent_pkg_aliases,
-                    )
-                    .await
-                }
-            })
-            .pipe(future::try_join_all)
-            .await?;
-        if is_current_children_owner(ctx, &id, &children_owner.owner) {
-            let mut realized: BTreeMap<String, NodeId> = BTreeMap::new();
-            let mut by_id: Vec<crate::resolved_tree::ChildEdge> = Vec::new();
-            let optional_by_alias: HashMap<&str, bool> = child_refs
-                .iter()
-                .map(|(alias, _)| (alias.as_str(), is_optional_child(snapshot, alias)))
-                .collect();
-            for dep in child_results.into_iter().flatten() {
-                let optional = optional_by_alias.get(dep.alias.as_str()).copied().unwrap_or(false);
-                by_id.push(crate::resolved_tree::ChildEdge {
-                    alias: dep.alias.clone(),
-                    pkg_id: dep.id.clone(),
-                    optional,
-                });
-                realized.insert(dep.alias, dep.node_id);
-            }
-            lock_recoverable(&ctx.workspace.children_by_id).insert(id.clone(), Arc::new(by_id));
-            crate::resolved_tree::TreeChildren::Realized(realized)
-        } else {
-            crate::resolved_tree::TreeChildren::Lazy {
-                parent_ids: AncestorIds::from(Arc::clone(ancestor_ids)),
-            }
-        }
-    } else {
-        crate::resolved_tree::TreeChildren::Lazy {
-            parent_ids: AncestorIds::from(Arc::clone(ancestor_ids)),
-        }
-    };
-
-    remember_node_parent_ids(ctx, &node_id, Arc::clone(ancestor_ids));
-    insert_tree_node(ctx, node_id.clone(), &id, children, depth);
-    if children_owner.owns_children
-        && !children_owner.children_context_unchanged
-        && is_current_children_owner(ctx, &id, &children_owner.owner)
-    {
-        make_non_owner_nodes_lazy(ctx, &id, &node_id);
-    }
-
-    Ok(Some(DirectDep { alias, node_id, id }))
-}
-
-/// `(install_alias, resolved_snapshot_key)` for every non-`link:` child
-/// recorded on `snapshot`'s `dependencies` + `optionalDependencies`,
-/// excluding resolved peers. Sorted by alias so the per-occurrence walk
-/// order is deterministic.
-///
-/// A snapshot's `dependencies` map lists not only the package's real
-/// dependencies but also every *resolved peer* — the node's own peers
-/// (`peer_dependencies`) and the peers its descendants required and
-/// resolved through this node (`transitivePeerDependencies`) — each
-/// pinned to the version it matched in the recorded context. Those are
-/// not real children: a fresh resolve walks only the package's manifest
-/// `dependencies` and re-derives peers separately against the parent
-/// context. Reuse must walk the manifest's deps too — so peer-named
-/// entries are dropped here, and the snapshot's `dependencies` is used
-/// only as the locked-ref lookup, not as the child set. Treating a
-/// resolved peer as a regular child makes the peer pass satisfy the peer
-/// from the node's own subtree instead of propagating it up, collapsing
-/// the peer-context suffix.
-fn snapshot_child_refs(
-    snapshot: Option<&SnapshotEntry>,
-    peer_dependencies: &BTreeMap<String, PeerDep>,
-) -> Vec<(String, PkgNameVerPeer)> {
-    let Some(snapshot) = snapshot else { return Vec::new() };
-    let transitive_peers: HashSet<&str> =
-        snapshot.transitive_peer_dependencies.iter().flatten().map(String::as_str).collect();
-    let mut out: Vec<(String, PkgNameVerPeer)> = Vec::new();
-    for dep_map in [snapshot.dependencies.as_ref(), snapshot.optional_dependencies.as_ref()]
-        .into_iter()
-        .flatten()
-    {
-        for (alias, dep_ref) in dep_map {
-            let alias_str = match &alias.scope {
-                Some(scope) => Cow::Owned(format!("@{scope}/{}", alias.bare)),
-                None => Cow::Borrowed(alias.bare.as_str()),
-            };
-            if peer_dependencies.contains_key(alias_str.as_ref())
-                || transitive_peers.contains(alias_str.as_ref())
-            {
-                continue;
-            }
-            if let Some(key) = dep_ref.resolve(alias) {
-                out.push((alias_str.into_owned(), key));
-            }
-        }
-    }
-    out.sort_by(|a, b| a.0.cmp(&b.0));
-    out
-}
-
-/// `true` when `alias` is recorded under `snapshot.optionalDependencies`
-/// (as opposed to `dependencies`). Threads the right `optional` flag onto
-/// the reused child's [`crate::resolved_tree::ChildEdge`].
-fn is_optional_child(snapshot: Option<&SnapshotEntry>, alias: &str) -> bool {
-    let Some(snapshot) = snapshot else { return false };
-    let Ok(name) = alias.parse::<pacquet_lockfile::PkgName>() else { return false };
-    snapshot.optional_dependencies.as_ref().is_some_and(|deps| deps.contains_key(&name))
-}
-
-/// Replace `catalog:` bare specifiers on direct dependencies with the
-/// version recorded in the catalogs map. Non-`catalog:` specifiers
-/// pass through unchanged.
-///
-/// Catalog resolution runs only on importer-level deps. A misconfigured
-/// entry surfaces immediately rather than masquerading as a
-/// `ERR_PNPM_SPEC_NOT_SUPPORTED_BY_ANY_RESOLVER`.
-pub(crate) fn resolve_catalog_specifiers(
-    specs: Vec<WantedSpec>,
-    catalogs: &Catalogs,
-) -> Result<Vec<WantedSpec>, ResolveDependencyTreeError> {
-    specs
-        .into_iter()
-        .map(|(name, range, optional, injected)| {
-            resolve_catalog_specifier(name, range, catalogs)
-                .map(|(name, range)| (name, range, optional, injected))
-        })
-        .collect()
-}
-
-fn resolve_catalog_specifier(
-    name: String,
-    range: String,
-    catalogs: &Catalogs,
-) -> Result<(String, String), ResolveDependencyTreeError> {
-    let wanted = CatalogWantedDependency { alias: name.clone(), bare_specifier: range.clone() };
-    match resolve_from_catalog(catalogs, &wanted) {
-        CatalogResolutionResult::Found(found) => Ok((name, found.resolution.specifier)),
-        CatalogResolutionResult::Unused => Ok((name, range)),
-        CatalogResolutionResult::Misconfiguration(misconfig) => {
-            Err(ResolveDependencyTreeError::CatalogMisconfiguration(misconfig.error))
-        }
-    }
-}
-
-/// Compute the `pkgIdWithPatchHash` for a freshly-resolved package:
-///
-/// 1. Prefix the resolver's `id` with `<name>@` when it doesn't already
-///    start that way. The npm-registry resolver always returns
-///    `name@version`; git / tarball / local resolvers return URL-shaped
-///    ids and need the prefix so the snapshot key starts with the
-///    package name.
-/// 2. Look the `(name, version)` pair up in `ctx.patched_dependencies`
-///    via [`get_patch_info`] (exact → unique range → wildcard).
-/// 3. On a match, append `(patch_hash=<hash>)` and record the matched
-///    key on `ctx.applied_patches` so the post-walk
-///    `ERR_PNPM_UNUSED_PATCH` check sees the hit.
-///
-/// Packages whose resolver didn't supply [`pacquet_resolving_resolver_base::ResolveResult::name_ver`]
-/// (git / tarball / local — they learn the name from the manifest at
-/// fetch time) skip the patch lookup. That matches the surface
-/// `patchedDependencies` covers today: keys are `name[@version]`, so a
-/// package without a resolve-time name can't match a configured entry
-/// anyway. The lookup is also skipped when no patches are configured.
-async fn build_pkg_id_with_patch_hash(
-    ctx: &TreeCtx,
-    result: &pacquet_resolving_resolver_base::ResolveResult,
-) -> Result<String, ResolveDependencyTreeError> {
-    let raw_id = result.id.as_str();
-    if let Some(target) = raw_id.strip_prefix("link:") {
-        let target = std::path::Path::new(target);
-        let absolute_target = if target.is_absolute() {
-            pacquet_fs::lexical_normalize(target)
-        } else {
-            pacquet_fs::lexical_normalize(&ctx.base_opts.project_dir.join(target))
-        };
-        let relative_target =
-            pathdiff::diff_paths(&absolute_target, &ctx.lockfile_dir).unwrap_or(absolute_target);
-        let relative_target = relative_target.display().to_string().replace('\\', "/");
-        let relative_target = if relative_target.is_empty() { "." } else { &relative_target };
-        return Ok(format!("link:{relative_target}"));
-    }
-    // Resolvers that learn the name from the fetched manifest (git,
-    // tarball, directory) leave `name_ver` unset. The `name` is read
-    // from the manifest itself and the id is prefixed regardless — so a
-    // `file:project-1` id becomes `project-1@file:project-1`. Skipping
-    // the prefix would leave `(` as the first paren-bearing character in
-    // the downstream depPath, which `PkgNameVerPeer`'s `@`-split parser
-    // can't recover from (it finds the `@` inside the peer suffix first).
-    let manifest_name = result
-        .manifest
-        .as_ref()
-        .and_then(|manifest| manifest.get("name"))
-        .and_then(serde_json::Value::as_str);
-    let (name, version) = match (result.name_ver.as_ref(), manifest_name) {
-        (Some(name_ver), _) => (name_ver.name.to_string(), name_ver.suffix.to_string()),
-        (None, Some(name)) => (name.to_string(), String::new()),
-        (None, None) => return Ok(raw_id.to_string()),
-    };
-    let prefixed = if raw_id.starts_with(&format!("{name}@")) {
-        raw_id.to_string()
-    } else {
-        format!("{name}@{raw_id}")
-    };
-    // `patched_dependencies` keys carry a `name@version` shape, so
-    // entries that came in without a `name_ver` (file: / git: /
-    // tarball: resolutions whose name we just learned from the
-    // manifest above) can't match unless the manifest also surfaced
-    // a version. Bail out when version is empty so the patch lookup
-    // doesn't run a `name@""` query.
-    if version.is_empty() {
-        return Ok(prefixed);
-    }
-    let Some(groups) = ctx.patched_dependencies.as_deref() else {
-        return Ok(prefixed);
-    };
-    let Some(patch) = get_patch_info(Some(groups), &name, &version)? else {
-        return Ok(prefixed);
-    };
-    lock_recoverable(&ctx.workspace.applied_patches).insert(patch.key.clone());
-    Ok(format!("{prefixed}(patch_hash={})", patch.hash))
-}
-
-/// Render `{alias}@{bare}` (either half dropped when absent) for the
-/// no-resolver error message.
-fn render_specifier(wanted: &WantedDependency) -> String {
-    let alias = wanted.alias.as_deref().unwrap_or("");
-    let bare = wanted.bare_specifier.as_deref().unwrap_or("");
-    match (alias.is_empty(), bare.is_empty()) {
-        (true, true) => String::new(),
-        (true, false) => bare.to_string(),
-        (false, true) => alias.to_string(),
-        (false, false) => format!("{alias}@{bare}"),
-    }
-}
-
-/// Extract `dependencies` + `optionalDependencies` from a resolved
-/// package's manifest. Peer dependencies are **not** walked as regular
-/// edges here — they're hoisted to the importer level by the
-/// [`fn@crate::resolve_importer`] orchestrator (which calls [`extend_tree`]
-/// with the hoist-picker's output) so a peer ends up shared across
-/// every consumer, not nested under each one.
-///
-/// Peers are still recorded on [`ResolvedPackage::peer_dependencies`]
-/// (via [`extract_peer_dependencies`]) so the peer-resolution stage
-/// can compute the correct depPath suffix once everything is walked.
-///
-/// Each entry carries an `optional` flag — `true` when the name appears
-/// in `optionalDependencies`. The walker propagates this through
-/// `current_is_optional` so [`ResolvedPackage::optional`] reflects
-/// whether every path to the node went through an optional edge.
-///
-/// A name listed in both maps yields one optional edge with the
-/// `dependencies` range — npm merges `optionalDependencies` into
-/// `dependencies` at publish time, so registry manifests routinely
-/// list the same name in both.
-///
-/// Names the manifest bundles are dropped: npm ships them inside the
-/// package's own tarball, so resolving them again would install a
-/// second copy the package never loads.
-fn extract_children(
-    result: &pacquet_resolving_resolver_base::ResolveResult,
-) -> Result<Vec<ChildSpec>, ResolveDependencyTreeError> {
-    let Some(manifest) = result.manifest.as_ref() else { return Ok(Vec::new()) };
-    let parent = render_parent(result);
-    let bundled = bundled_dependency_names(manifest);
-    let mut out = Vec::new();
-    collect_deps(manifest, "dependencies", false, &parent, &bundled, &mut out)?;
-    let mut optional = Vec::new();
-    collect_deps(manifest, "optionalDependencies", true, &parent, &bundled, &mut optional)?;
-    if !optional.is_empty() {
-        let dependency_positions: HashMap<String, usize> =
-            out.iter().enumerate().map(|(index, (name, ..))| (name.clone(), index)).collect();
-        for spec in optional {
-            match dependency_positions.get(&spec.0) {
-                Some(&index) => out[index].2 = true,
-                None => out.push(spec),
-            }
-        }
-    }
-    for (name, specifier) in engines_runtime_dependencies(manifest, "engines", "dependencies") {
-        out.push((name.to_string(), specifier, false));
-    }
-    Ok(out)
-}
-
-/// The dependency names a manifest declares as bundled, read from
-/// `bundledDependencies` with `bundleDependencies` as the fallback
-/// spelling. `true` stands for "every entry in `dependencies`".
-fn bundled_dependency_names(manifest: &Value) -> HashSet<&str> {
-    let bundled = ["bundledDependencies", "bundleDependencies"]
-        .into_iter()
-        .find_map(|key| manifest.get(key).filter(|value| !value.is_null()));
-    match bundled {
-        Some(Value::Bool(true)) => manifest
-            .get("dependencies")
-            .and_then(Value::as_object)
-            .map(|map| map.keys().map(String::as_str).collect())
-            .unwrap_or_default(),
-        Some(Value::Array(names)) => names.iter().filter_map(Value::as_str).collect(),
-        _ => HashSet::default(),
-    }
-}
-
-/// Dependency names are validated before the bundled filter runs, so a
-/// manifest with an unusable alias is rejected whether or not the
-/// package bundles that alias.
-fn collect_deps(
-    manifest: &Value,
-    key: &str,
-    optional: bool,
-    parent: &str,
-    bundled: &HashSet<&str>,
-    out: &mut Vec<ChildSpec>,
-) -> Result<(), ResolveDependencyTreeError> {
-    let Some(map) = manifest.get(key).and_then(Value::as_object) else { return Ok(()) };
-    for (name, range) in map {
-        if let Some(range_str) = range.as_str() {
-            if !crate::is_valid_dependency_alias(name) {
-                return Err(ResolveDependencyTreeError::InvalidDependencyName {
-                    parent: parent.to_string(),
-                    alias: name.clone(),
-                });
-            }
-            if bundled.contains(name.as_str()) {
-                continue;
-            }
-            out.push((name.clone(), range_str.to_string(), optional));
-        }
-    }
-    Ok(())
-}
-
-fn render_parent(result: &pacquet_resolving_resolver_base::ResolveResult) -> String {
-    if let Some(name_ver) = result.name_ver.as_ref() {
-        format!(r#"Package "{}@{}""#, name_ver.name, name_ver.suffix)
-    } else {
-        format!(r#"Package "{}""#, result.id)
-    }
-}
-
-/// Extract `peerDependencies` from a resolved package's manifest, with
-/// `peerDependenciesMeta[name].optional` folded onto each entry. The
-/// package's own name plus names also present in `dependencies` or
-/// `optionalDependencies` are skipped because those edges already
-/// supply the same package directly — except for the `peer_shadowed`
-/// names, whose own-dependency edge the walk drops in favor of the peer
-/// (see [`peer_shadowed_dependencies`]). A `peerDependenciesMeta` entry
-/// without a matching `peerDependencies` entry only counts when
-/// `optional: true` — it is treated as an optional `"*"` peer exactly
-/// like an explicitly declared one, and non-optional meta-only entries
-/// are ignored.
-fn extract_peer_dependencies(
-    result: &pacquet_resolving_resolver_base::ResolveResult,
-    peer_shadowed: &HashSet<String>,
-) -> BTreeMap<String, PeerDep> {
-    let Some(manifest) = result.manifest.as_ref() else { return BTreeMap::new() };
-    let mut peers: BTreeMap<String, PeerDep> = BTreeMap::new();
-
-    let dep_names = |key| {
-        manifest.get(key).and_then(Value::as_object).into_iter().flat_map(|map| map.keys().cloned())
-    };
-    // Only `dependencies` are shadowed, so an entry that is *also* an
-    // optional dependency still supplies the name itself.
-    let mut own_deps: HashSet<String> = dep_names("dependencies")
-        .filter(|name| !peer_shadowed.contains(name))
-        .chain(dep_names("optionalDependencies"))
-        .collect();
-    if let Some(name) = manifest.get("name").and_then(Value::as_str) {
-        own_deps.insert(name.to_string());
-    }
-
-    if let Some(map) = manifest.get("peerDependencies").and_then(Value::as_object) {
-        for (name, range) in map {
-            if own_deps.contains(name) {
-                continue;
-            }
-            if let Some(range_str) = range.as_str() {
-                peers.insert(
-                    name.clone(),
-                    PeerDep { version: range_str.to_string(), optional: false },
-                );
-            }
-        }
-    }
-
-    if let Some(meta) = manifest.get("peerDependenciesMeta").and_then(Value::as_object) {
-        for (name, info) in meta {
-            if own_deps.contains(name)
-                || info.get("optional").and_then(Value::as_bool) != Some(true)
-            {
-                continue;
-            }
-            peers
-                .entry(name.clone())
-                .and_modify(|entry| entry.optional = true)
-                .or_insert_with(|| PeerDep { version: "*".to_string(), optional: true });
-        }
-    }
-
-    peers
-}
-
-/// `true` when the package has no `dependencies`, `optionalDependencies`,
-/// `peerDependencies`, or `peerDependenciesMeta`.
-///
-/// Conservatively returns `false` when the manifest is missing —
-/// collapsing onto a leaf `NodeId` would claim knowledge of children
-/// there is none of. Resolutions reaching here always carry one, real
-/// or synthesized by [`fallback_manifest`].
-fn pkg_is_leaf(result: &pacquet_resolving_resolver_base::ResolveResult) -> bool {
-    let Some(manifest) = result.manifest.as_ref() else { return false };
-    is_empty_or_absent(manifest.get("dependencies"))
-        && is_empty_or_absent(manifest.get("optionalDependencies"))
-        && is_empty_or_absent(manifest.get("peerDependencies"))
-        && is_empty_or_absent(manifest.get("peerDependenciesMeta"))
-}
-
-fn is_empty_or_absent(value: Option<&Value>) -> bool {
-    value.and_then(Value::as_object).is_none_or(serde_json::Map::is_empty)
-}
-
-/// Emits a [`Deprecation`] when a newly-resolved package's manifest carries
-/// a non-empty `deprecated` field not covered by `allowedDeprecatedVersions`.
-fn emit_deprecation_if_needed(
-    ctx: &TreeCtx,
-    result: &pacquet_resolving_resolver_base::ResolveResult,
-    id: &str,
-    depth: i32,
-) {
-    let Some(deprecated) = extract_deprecated_from_manifest(result.manifest.as_deref()) else {
-        return;
-    };
-    if deprecated.is_empty() {
-        return;
-    }
-    let Some((pkg_name, pkg_version)) = deprecated_pkg_name_ver(result) else {
-        return;
-    };
-    if is_deprecation_allowed(&pkg_name, &pkg_version, &ctx.workspace.allowed_deprecated_versions) {
-        return;
-    }
-    let Some(log) = ctx.workspace.deprecation_log.as_ref() else {
-        return;
-    };
-    log(Deprecation {
-        pkg_name,
-        pkg_version,
-        pkg_id: id.to_string(),
-        prefix: ctx.base_opts.project_dir.display().to_string(),
-        deprecated,
-        depth,
-    });
-}
-
-/// A missing manifest, an absent `deprecated` field, and a non-string
-/// one all count as not deprecated.
-fn extract_deprecated_from_manifest(manifest: Option<&Value>) -> Option<String> {
-    manifest?.get("deprecated")?.as_str().map(str::to_string)
-}
-
-/// The name/version a `pnpm:deprecation` payload reports:
-/// [`ResolveResult::name_ver`] when the resolver filled it, otherwise
-/// the manifest's `name`/`version` — the canonical source for non-npm
-/// resolutions (see the `name_ver` field doc). `None`, suppressing the
-/// warning, when neither carries both fields.
-///
-/// [`ResolveResult::name_ver`]: pacquet_resolving_resolver_base::ResolveResult::name_ver
-fn deprecated_pkg_name_ver(
-    result: &pacquet_resolving_resolver_base::ResolveResult,
-) -> Option<(String, String)> {
-    if let Some(nv) = result.name_ver.as_ref() {
-        return Some((nv.name.to_string(), nv.suffix.to_string()));
-    }
-    let manifest = result.manifest.as_deref()?;
-    let name = manifest.get("name")?.as_str()?;
-    let version = manifest.get("version")?.as_str()?;
-    Some((name.to_string(), version.to_string()))
-}
-
-/// Whether `pkg_name@pkg_version` satisfies its entry in the
-/// `pnpm.allowedDeprecatedVersions` map, matching the TS check at
-/// `pnpm11/installing/deps-resolver/src/resolveDependencies.ts:2122`.
-fn is_deprecation_allowed(
-    pkg_name: &str,
-    pkg_version: &str,
-    allowed_deprecated_versions: &BTreeMap<String, String>,
-) -> bool {
-    let Some(range_str) = allowed_deprecated_versions.get(pkg_name) else {
-        return false;
-    };
-    let Ok(range) = range_str.parse::<node_semver::Range>() else {
-        return false;
-    };
-    let Ok(version) = pkg_version.parse::<node_semver::Version>() else {
-        return false;
-    };
-    range.satisfies(&version)
-}
-
-/// Provenance tags that count as non-exotic for `blockExoticSubdeps`.
-const NON_EXOTIC_RESOLVED_VIA: &[&str] = &[
-    "custom-resolver",
-    "github.com/denoland/deno",
-    "github.com/oven-sh/bun",
-    "jsr-registry",
-    "local-filesystem",
-    "named-registry",
-    "nodejs.org",
-    "npm-registry",
-    "workspace",
-];
-
-fn is_exotic_resolved_via(resolved_via: &str) -> bool {
-    !NON_EXOTIC_RESOLVED_VIA.contains(&resolved_via)
 }
 
 #[cfg(test)]
