+++ pnpm/crates/resolving-deps-resolver/src/resolve_dependency_tree/walk.rs
@@ -1,0 +1,1125 @@
+//! The fresh-resolve walk: seeding one edge's package
+//! ([`fn@resolve_node_seed`]), walking its children
+//! ([`fn@walk_node_children`]), and the per-wanted resolve cache both
+//! go through. The two phases are separate so a whole sibling level
+//! resolves before any of its subtrees start, letting the level's
+//! versions seed the children's preferred-versions overlay.
+
+use async_recursion::async_recursion;
+use futures_util::future;
+use pacquet_lockfile::PkgNameVerPeer;
+use pacquet_resolving_resolver_base::{
+    CurrentPkg, NoMatchingVersionError, PreferredVersionsOverlay, RegistryResponseError,
+    ResolveError, ResolveOptions, Resolver, WantedDependency,
+};
+use pipe_trait::Pipe;
+use rustc_hash::{FxHashMap as HashMap, FxHashSet as HashSet};
+use serde_json::Value;
+use std::{borrow::Cow, collections::BTreeMap, path::Path, sync::Arc};
+
+use crate::{
+    lockfile_reuse::{current_pkg_from_lockfile, prior_child_key},
+    node_id::NodeId,
+    parent_pkg_aliases::{ParentPkgAliases, peer_shadowed_dependencies},
+    resolved_tree::{AncestorIds, DirectDep, ResolvedPackage},
+};
+
+use super::{
+    ResolveDependencyTreeError, SkippedOptionalDependency, SkippedOptionalDependencyParent,
+    catalogs::resolve_catalog_specifier,
+    lock_recoverable,
+    manifest::{
+        build_pkg_id_with_patch_hash, emit_deprecation_if_needed, extract_children,
+        extract_peer_dependencies, is_exotic_resolved_via, pkg_is_leaf,
+    },
+    reuse::{
+        ReuseSource, higher_direct_dep_version, is_update_target,
+        node_depends_on_changed_direct_dep, real_package_name_of, resolve_reused_node,
+        try_reuse_node, wanted_lockfile_contains_satisfying_entry,
+    },
+    tree_ctx::{
+        TreeCtx, declaring_manifest_dir, opts_relative_to_declaring_manifest,
+        project_relative_cache_scope,
+    },
+    workspace_ctx::{
+        ChildSpec, ChildrenOwnerClaim, WantedKey, claim_children_owner, insert_tree_node,
+        is_current_children_owner, make_non_owner_nodes_lazy, register_peer_dep_names,
+        remember_node_parent_ids,
+    },
+};
+
+/// Resolve one `(alias, range)` edge end-to-end with no
+/// preferred-versions overlay: [`fn@resolve_node_seed`] then
+/// [`fn@walk_node_children`]. Used where per-level preference folding
+/// does not apply — the lockfile-reuse subtree walk, whose versions
+/// are exact pins.
+///
+/// The node's children resolve in the same `parent_pkg_aliases` scope
+/// as the node itself, without this level's own aliases folded in: a
+/// reused subtree takes its children from the lockfile snapshot rather
+/// than from a manifest, so no shadowed dependency can be dropped
+/// inside it, and the narrower scope only ever means one omission
+/// fewer where an edge falls back to a fresh resolve.
+#[expect(
+    clippy::too_many_arguments,
+    reason = "internal walker helper threading per-node context through the recursion"
+)]
+#[async_recursion]
+pub(super) async fn resolve_node<Chain>(
+    ctx: &TreeCtx,
+    resolver: &Chain,
+    wanted: WantedDependency,
+    ancestor_ids: &Arc<Vec<String>>,
+    depth: i32,
+    parent_optional: bool,
+    reuse: ReuseSource,
+    parent_pkg_aliases: &Arc<ParentPkgAliases>,
+) -> Result<Option<DirectDep>, ResolveDependencyTreeError>
+where
+    Chain: Resolver + ?Sized,
+{
+    let base_overlay = ctx.base_opts.preferred_versions_overlay.clone();
+    match resolve_node_seed(
+        ctx,
+        resolver,
+        wanted,
+        ancestor_ids,
+        depth,
+        parent_optional,
+        reuse,
+        base_overlay.clone(),
+        None,
+        parent_pkg_aliases,
+    )
+    .await?
+    {
+        NodeSeed::Done(dep) => Ok(dep),
+        NodeSeed::Pending(pending) => {
+            walk_node_children(
+                ctx,
+                resolver,
+                *pending,
+                base_overlay,
+                Arc::clone(parent_pkg_aliases),
+            )
+            .await
+        }
+    }
+}
+
+/// Outcome of [`fn@resolve_node_seed`]: either the edge completed
+/// without a children walk (lockfile reuse, cycle break), or the
+/// package resolved and its children walk is still pending — the
+/// caller runs it via [`fn@walk_node_children`] once every sibling
+/// seed settled, so the children's resolution sees the whole level's
+/// versions in its preferred-versions overlay.
+pub(super) enum NodeSeed {
+    Done(Option<DirectDep>),
+    Pending(Box<PendingNode>),
+}
+
+/// A resolved-but-not-walked node: everything
+/// [`fn@walk_node_children`] needs to recurse into the children.
+pub(super) struct PendingNode {
+    result: Arc<pacquet_resolving_resolver_base::ResolveResult>,
+    id: String,
+    alias: String,
+    node_id: NodeId,
+    is_link: bool,
+    parent_ancestors: Arc<Vec<String>>,
+    next_ancestors: Arc<Vec<String>>,
+    /// The deterministic children-ownership claim taken at seed time;
+    /// the walk phase re-checks it before recording the children, so
+    /// a better-placed occurrence seeded after this one still wins.
+    /// It also carries the peer-shadowed dependency names the walk
+    /// phase filters this package's children by.
+    children_owner: ChildrenOwnerClaim,
+    depth: i32,
+    current_is_optional: bool,
+    /// The edge's recorded snapshot key in the prior lockfile, if
+    /// any — threads each child's prior ref through the walk phase
+    /// via `ReuseSource::Transitive`.
+    pub(super) prior_key: Option<PkgNameVerPeer>,
+}
+
+/// Resolve one `(alias, range)` edge and register the resolved package
+/// in the dedup map if absent, run for a whole sibling level before any
+/// child subtree starts.
+///
+/// `pick_overlay` carries the per-level preferred-version additions
+/// (the parent level's resolved versions) consulted by the npm
+/// resolver's version pick; it participates in the per-wanted dedup
+/// cache key so the same range can legitimately pick different
+/// versions under different levels, layering each level's resolved
+/// versions onto the preferred-versions fold.
+///
+/// `ancestor_ids` is the chain of `pkgIdWithPatchHash` values from the
+/// root importer down to the current node's parent. When the resolved
+/// id appears in the chain, this call is a cycle re-entry: the edge is
+/// dropped entirely (returns `Done(None)`) so the parent's `children`
+/// map omits the cycled child. Without this, two nodes for the same id
+/// race each other into `graph.insert`, and an empty-children entry for
+/// the cycled occurrence can overwrite the real one.
+///
+/// `parent_dir` is the directory of the manifest that declares this
+/// edge, when that manifest is a package resolved from a local
+/// directory — see [`declaring_manifest_dir`].
+///
+/// `parent_pkg_aliases` is the scope this edge's level resolves in; it
+/// decides which of the resolved package's `dependencies` its own
+/// `peerDependencies` shadow (see [`peer_shadowed_dependencies`]).
+#[expect(
+    clippy::too_many_arguments,
+    reason = "internal walker helper threading per-node context through the recursion"
+)]
+#[async_recursion]
+pub(super) async fn resolve_node_seed<Chain>(
+    ctx: &TreeCtx,
+    resolver: &Chain,
+    wanted: WantedDependency,
+    ancestor_ids: &Arc<Vec<String>>,
+    depth: i32,
+    parent_optional: bool,
+    reuse: ReuseSource,
+    pick_overlay: Option<Arc<PreferredVersionsOverlay>>,
+    parent_dir: Option<&Path>,
+    parent_pkg_aliases: &Arc<ParentPkgAliases>,
+) -> Result<NodeSeed, ResolveDependencyTreeError>
+where
+    Chain: Resolver + ?Sized,
+{
+    let current_is_optional = wanted.optional.unwrap_or(false) || parent_optional;
+
+    // The edge's recorded snapshot key in the prior lockfile, if any.
+    // Feeds both subtree reuse (below) and — when the edge re-resolves
+    // anyway — the `currentPkg` payload custom resolvers receive.
+    let prior_key = reuse.prior_key(ctx, &wanted);
+
+    // **Lockfile-resolution reuse.** When the prior lockfile already
+    // resolved this edge (and the recorded version still satisfies the
+    // manifest range, for a direct dep), synthesize the resolution from
+    // the lockfile and walk its transitive subtree from the snapshot
+    // graph instead of re-resolving from the registry.
+    // `synthesize_reused_result` is conservative: any shape it can't
+    // faithfully reproduce (non-registry resolutions, missing metadata)
+    // yields `None` here and the node falls through to a fresh resolve.
+    //
+    // Stale-pin refresh: a node depending on a changed direct dep is
+    // resolved fresh rather than reused, so its children walk against
+    // their manifest ranges where `walk_node_children` can redirect a
+    // stale pin onto the higher direct-dep version (reusing the subtree
+    // would keep the pin, leaving the lockfile non-convergent).
+    if reuse.allows_reuse()
+        && !node_depends_on_changed_direct_dep(ctx, prior_key.as_ref())
+        && let Some(reused) = try_reuse_node(ctx, &wanted, prior_key.as_ref(), depth)
+    {
+        return resolve_reused_node(
+            ctx,
+            resolver,
+            wanted,
+            ancestor_ids,
+            depth,
+            current_is_optional,
+            reused,
+            parent_pkg_aliases,
+        )
+        .await
+        .map(NodeSeed::Done);
+    }
+
+    // Memoise the per-wanted resolve. The first caller for a given
+    // `(alias, bare_specifier, optional)` runs the resolver chain and
+    // stores the `Arc<ResolveResult>` on `ctx.resolved_by_wanted`;
+    // every later caller for the same wanted dep clones the `Arc` and
+    // skips the chain entirely. Concurrent first-callers can both miss
+    // the cache and run `resolver.resolve` in parallel — the resolver's
+    // own per-cache-key semaphore (`pick_package::fetch_locker`)
+    // already coalesces those into a single network fetch, so the
+    // doubled work is bounded to in-memory packument lookups + semver
+    // matching, and the second to finish loses the `insert` race
+    // harmlessly (the entry holds an `Arc` to an equivalent
+    // `ResolveResult`).
+    // `resolutionMode` makes the version pick depend on whether this is
+    // a direct (`depth == 0`) or transitive dep, so the cache key and
+    // the resolver call both key off the depth-specific options.
+    let opts = ctx.opts_for_depth(depth);
+    // The prior lockfile entry rides along as `currentPkg`, handed to
+    // the resolver. Only custom resolvers read it today; the clone of
+    // the shared per-depth options is paid only when a prior entry
+    // exists for a freshly resolving edge.
+    let current_pkg = prior_key.as_ref().and_then(|key| {
+        let lockfile = ctx.workspace.wanted_lockfile.as_ref()?;
+        current_pkg_from_lockfile(lockfile, key, &ctx.workspace.registries)
+    });
+    let opts_with_current_pkg;
+    let opts = match current_pkg {
+        Some(current_pkg) => {
+            opts_with_current_pkg =
+                ResolveOptions { current_pkg: Some(current_pkg), ..opts.clone() };
+            &opts_with_current_pkg
+        }
+        None => opts,
+    };
+    let opts_for_file_dep = opts_relative_to_declaring_manifest(opts, &wanted, parent_dir);
+    let opts = opts_for_file_dep.as_ref();
+    // Project-relative resolutions (`link:`/`file:`/`workspace:`) are
+    // keyed by the consuming importer so one importer's relative path
+    // is never reused by another. See [`WantedKey`]. The prior key
+    // joins so two edges that share a specifier but recorded different
+    // versions never share a `currentPkg`-dependent result.
+    let project_scope = project_relative_cache_scope(&wanted, opts);
+    // The overlay's view for this edge joins the cache key: the same
+    // range can legitimately pick different versions under levels
+    // that resolved different siblings. The view keeps each candidate
+    // name (alias, `npm:` inner target, folded `jsr:` name) paired
+    // with its versions — the picker consults the overlay per name,
+    // so a flat union of versions could collide two overlays that
+    // distribute the same versions across different names. Empty for
+    // almost every edge, so the dedup keeps working where it matters.
+    let overlay_versions: Vec<(String, Vec<String>)> = pick_overlay
+        .as_ref()
+        .map(|overlay| {
+            let mut view: Vec<(String, Vec<String>)> = overlay_lookup_names(&wanted)
+                .into_iter()
+                .filter_map(|name| {
+                    let mut versions: Vec<String> =
+                        overlay.versions_for(&name).into_iter().map(str::to_string).collect();
+                    if versions.is_empty() {
+                        return None;
+                    }
+                    versions.sort_unstable();
+                    versions.dedup();
+                    Some((name, versions))
+                })
+                .collect();
+            view.sort_unstable();
+            view
+        })
+        .unwrap_or_default();
+    let update_target = is_update_target(ctx.update_scope(), &wanted, depth);
+    let cache_key: WantedKey = (
+        wanted.alias.clone(),
+        wanted.bare_specifier.clone(),
+        wanted.optional,
+        wanted.injected,
+        opts.pick_lowest_version,
+        opts.published_by,
+        project_scope,
+        prior_key.clone(),
+        overlay_versions.clone(),
+        ctx.update_cache_scope(),
+        update_target,
+    );
+    let result =
+        match resolve_wanted_cached(ctx, resolver, &wanted, opts, pick_overlay.as_ref(), cache_key)
+            .await
+        {
+            Ok(result) => result,
+            // A resolution failure on an optional edge drops the edge
+            // instead of failing the install — unless the wanted lockfile
+            // already holds a satisfying entry, where the silent skip would
+            // erase the locked entries (see
+            // `wanted_lockfile_contains_satisfying_entry`). Hook errors
+            // keep aborting even for optional edges.
+            Err(
+                err @ (ResolveDependencyTreeError::Resolve(_)
+                | ResolveDependencyTreeError::NoMatchingVersion(_)
+                | ResolveDependencyTreeError::RegistryResponse(_)
+                | ResolveDependencyTreeError::SpecNotSupported { .. }),
+            ) if wanted.optional.unwrap_or(false) => {
+                if wanted_lockfile_contains_satisfying_entry(
+                    ctx.workspace.wanted_lockfile.as_deref(),
+                    &wanted,
+                ) {
+                    return Err(ResolveDependencyTreeError::LockedOptionalResolutionFailure(
+                        Box::new(err),
+                    ));
+                }
+                if let Some(log) = ctx.workspace.skipped_optional_log.as_ref() {
+                    log(SkippedOptionalDependency {
+                        details: err.to_string(),
+                        name: wanted.alias.clone(),
+                        version: wanted
+                            .alias
+                            .is_some()
+                            .then(|| wanted.bare_specifier.clone())
+                            .flatten(),
+                        bare_specifier: wanted.bare_specifier.clone().unwrap_or_default(),
+                        parents: pkgs_info_from_ids(ctx, ancestor_ids),
+                        prefix: opts.project_dir.display().to_string(),
+                    });
+                }
+                return Ok(NodeSeed::Done(None));
+            }
+            Err(err) => return Err(err),
+        };
+
+    if let Some(violation) = result.policy_violation.clone() {
+        lock_recoverable(&ctx.workspace.policy_violations).push(violation);
+    }
+
+    if ctx.base_opts.block_exotic_subdeps
+        && depth > 0
+        && is_exotic_resolved_via(&result.resolved_via)
+    {
+        return Err(ResolveDependencyTreeError::ExoticSubdep {
+            specifier: wanted
+                .alias
+                .clone()
+                .or_else(|| wanted.bare_specifier.clone())
+                .unwrap_or_default(),
+            resolved_via: result.resolved_via.clone(),
+        });
+    }
+
+    if result.name_ver.is_none()
+        && wanted.bare_specifier.as_deref().is_some_and(|specifier| {
+            specifier.starts_with("workspace:") && !specifier.starts_with("workspace:.")
+        })
+        && let Some(manifest) = result.manifest.as_deref()
+        && let (Some(name), Some(version)) = (
+            manifest.get("name").and_then(Value::as_str),
+            manifest.get("version").and_then(Value::as_str),
+        )
+    {
+        ctx.workspace.record_run_version(name.to_string(), version.to_string());
+    }
+
+    let id = build_pkg_id_with_patch_hash(ctx, &result).await?;
+
+    // Cycle break — see the doc comment above. A direct self-edge and
+    // the second lap of a longer cycle are dropped; the first re-entry
+    // is kept so the cycle-closing edge reaches the lockfile snapshot.
+    if ancestor_ids.last().is_some_and(|parent| {
+        *parent == id || parent_ids_contain_sequence(ancestor_ids, parent, &id)
+    }) {
+        return Ok(NodeSeed::Done(None));
+    }
+
+    let alias = node_alias(&wanted, &result, &id);
+
+    // Build (or look up) the ResolvedPackage envelope. The first
+    // visitor populates it; later visitors AND-fold the `optional`
+    // flag so a single non-optional path flips it back to `false`.
+    // Child traversal is claimed first, and a later
+    // deterministically-better occurrence replaces the shared
+    // `children_by_id` entry — plus, since the two are two halves of
+    // one manifest reading, the envelope's peer dependencies.
+    // Leaves (no deps / optional deps / peers / peerDependenciesMeta)
+    // reuse the package id as their `NodeId`, collapsing every parent
+    // edge onto one tree node. Non-leaves still get a fresh per-
+    // occurrence id so the peer resolver can attach different peer
+    // suffixes per call site.
+    // The leaf flag is computed before the dedup insert so it can be
+    // persisted on [`ResolvedPackage::is_leaf`] for the lazy realisation
+    // path to read back.
+    // Workspace-link nodes get empty children (the linked project
+    // resolves its own deps as a separate importer), `depth = -1` flags
+    // the node for the peer-resolution short-circuit, and the
+    // [`ResolvedPackage`] carries no peer dependencies (peer matching is
+    // the linked importer's responsibility, not the parent's). The node
+    // id is collapsed to a leaf so every reference to the same workspace
+    // path shares one [`NodeId`].
+    let is_link = id.starts_with("link:");
+    let is_leaf = is_link || pkg_is_leaf(&result);
+    let node_id = if is_leaf { NodeId::leaf(&id) } else { NodeId::next() };
+
+    // The claim is taken while holding the `packages` lock so claims and
+    // envelope writes stay in the same order: an occurrence that lost the
+    // claim in between would otherwise leave its peer set on an envelope
+    // whose children the winning occurrence walked.
+    let mut packages = lock_recoverable(&ctx.workspace.packages);
+    let children_owner = claim_children_owner(
+        ctx,
+        &id,
+        depth,
+        ancestor_ids,
+        peer_shadowed_dependencies(
+            result.manifest.as_deref(),
+            parent_pkg_aliases,
+            ctx.workspace.auto_install_peers,
+        ),
+    );
+    let peer_dependencies = if is_link {
+        BTreeMap::new()
+    } else {
+        extract_peer_dependencies(&result, &children_owner.peer_shadowed)
+    };
+    let package_is_new = if let Some(existing) = packages.get_mut(&id) {
+        existing.optional = existing.optional && current_is_optional;
+        // A fresh claim can shadow a different set of dependencies than
+        // the occurrence that populated the envelope did, which moves
+        // names between the package's children and its peers.
+        if children_owner.owns_children && existing.peer_dependencies != peer_dependencies {
+            register_peer_dep_names(ctx, &peer_dependencies);
+            existing.peer_dependencies = peer_dependencies;
+        }
+        false
+    } else {
+        register_peer_dep_names(ctx, &peer_dependencies);
+        packages.insert(
+            id.clone(),
+            ResolvedPackage {
+                id: id.clone(),
+                result: Arc::clone(&result),
+                peer_dependencies,
+                optional: current_is_optional,
+                is_leaf,
+            },
+        );
+        true
+    };
+    drop(packages);
+
+    if package_is_new {
+        ctx.workspace.record_resolved_version(&result);
+        emit_deprecation_if_needed(ctx, &result, &id, depth);
+    }
+
+    let next_ancestors: Vec<String> =
+        ancestor_ids.iter().cloned().chain(std::iter::once(id.clone())).collect();
+    let next_ancestors = Arc::new(next_ancestors);
+
+    Ok(NodeSeed::Pending(Box::new(PendingNode {
+        result,
+        id,
+        alias,
+        node_id,
+        is_link,
+        parent_ancestors: Arc::clone(ancestor_ids),
+        next_ancestors,
+        children_owner,
+        depth,
+        current_is_optional,
+        prior_key,
+    })))
+}
+
+/// The name the edge installs under: the key its parent manifest
+/// declares, else the name the resolver resolved it under, else the
+/// resolved package's own name.
+// The manifest key has to come first: an importer entry is only
+// recorded for an alias the manifest declares, so an edge keyed by a
+// resolver alias that differs from its manifest key is dropped from the
+// lockfile importer.
+pub(super) fn node_alias(
+    wanted: &WantedDependency,
+    result: &pacquet_resolving_resolver_base::ResolveResult,
+    id: &str,
+) -> String {
+    wanted
+        .alias
+        .clone()
+        .filter(|alias| !alias.is_empty())
+        .or_else(|| result.alias.clone())
+        .or_else(|| result.name_ver.as_ref().map(|name_ver| name_ver.name.to_string()))
+        .unwrap_or_else(|| id.to_string())
+}
+
+/// Walk a seeded node's children. `children_overlay` is the preferred-versions
+/// overlay covering this node's own level (built by the caller from
+/// every sibling seed); the grandchildren's overlay layers this
+/// node's resolved children on top, extending the per-level fold.
+/// `children_pkg_aliases` follows the same shape: the caller folds this
+/// node's whole level into the scope, and the grandchildren's scope
+/// layers this node's resolved children on top.
+///
+/// Only the deterministic children owner walks this package's
+/// manifest children. Other occurrences stay lazy and expand from
+/// `children_by_id`, applying their own `parent_ids` cycle break.
+#[async_recursion]
+pub(super) async fn walk_node_children<Chain>(
+    ctx: &TreeCtx,
+    resolver: &Chain,
+    pending: PendingNode,
+    children_overlay: Option<Arc<PreferredVersionsOverlay>>,
+    children_pkg_aliases: Arc<ParentPkgAliases>,
+) -> Result<Option<DirectDep>, ResolveDependencyTreeError>
+where
+    Chain: Resolver + ?Sized,
+{
+    let PendingNode {
+        result,
+        id,
+        alias,
+        node_id,
+        is_link,
+        parent_ancestors,
+        next_ancestors,
+        children_owner,
+        depth,
+        current_is_optional,
+        prior_key,
+    } = pending;
+    let children = if is_link {
+        // Linked nodes don't walk their manifest's deps — see the
+        // `is_link` comment block above. They get an empty `Realized`
+        // map: a linked node has no children of its own here.
+        crate::resolved_tree::TreeChildren::Realized(BTreeMap::new())
+    } else if !children_owner.owns_children {
+        crate::resolved_tree::TreeChildren::Lazy {
+            parent_ids: AncestorIds::from(Arc::clone(&parent_ancestors)),
+        }
+    } else {
+        // Look up cached children specs first; only walk the manifest on
+        // a miss. The cache value is held by `Arc` so revisits clone the
+        // refcount instead of the inner `Vec<(String, String, bool)>`.
+        let child_specs = {
+            let cache = lock_recoverable(&ctx.workspace.children_specs_by_id);
+            cache.get(&id).map(Arc::clone)
+        };
+        let child_specs = if let Some(specs) = child_specs {
+            specs
+        } else {
+            let specs = Arc::new(extract_children(&result)?);
+            lock_recoverable(&ctx.workspace.children_specs_by_id)
+                .entry(id.clone())
+                .or_insert_with(|| Arc::clone(&specs));
+            specs
+        };
+        // The specs are cached unfiltered because which of them the
+        // package's own `peerDependencies` shadow is a property of the
+        // owner occurrence, not of the manifest.
+        let peer_shadowed = &children_owner.peer_shadowed;
+        let child_specs: Cow<'_, [ChildSpec]> = if peer_shadowed.is_empty() {
+            Cow::Borrowed(child_specs.as_slice())
+        } else {
+            Cow::Owned(
+                child_specs
+                    .iter()
+                    .filter(|(name, _, optional)| *optional || !peer_shadowed.contains(name))
+                    .cloned()
+                    .collect(),
+            )
+        };
+        let child_specs =
+            if result.resolved_via == "workspace" && result.id.as_str().starts_with("file:") {
+                Cow::Owned(
+                    child_specs
+                        .iter()
+                        .map(|(name, range, optional)| {
+                            resolve_catalog_specifier(name.clone(), range.clone(), &ctx.catalogs)
+                                .map(|(name, range)| (name, range, *optional))
+                        })
+                        .collect::<Result<Vec<_>, _>>()?,
+                )
+            } else {
+                child_specs
+            };
+        // An *updated* parent (one that landed on a different version
+        // than the lockfile recorded, or a new dep) discards its
+        // `resolvedDependencies` child refs, forcing its subtree to
+        // re-resolve. A parent that freshly resolved but landed back on
+        // its previously recorded version keeps the prior child refs
+        // alive (pnpm's non-`parentPkg.updated` arm), and each child
+        // edge re-enters the reuse gate with its recorded key — so a
+        // still-satisfied subtree is reused rather than re-resolved.
+        // Re-resolving those children would re-pick open ranges (`*`)
+        // at their newest versions and churn the lockfile.
+        let prior_children_snapshot = prior_key
+            .as_ref()
+            .filter(|key| landed_on_prior_entry(key, &id))
+            .and_then(|key| ctx.workspace.wanted_lockfile.as_ref()?.snapshots.as_ref()?.get(key));
+        // Phase 1: resolve every child package before any grandchild
+        // walk starts, so the level's resolved versions can feed the
+        // grandchildren's preferred-versions overlay — the
+        // postponed-resolution barrier.
+        // Snapshot this importer's direct-dep versions once for the whole
+        // child fanout instead of locking per edge.
+        let direct_versions = lock_recoverable(&ctx.workspace.direct_dep_versions)
+            .get(&ctx.importer_id)
+            .map(Arc::clone);
+        let declaring_dir = declaring_manifest_dir(ctx, &result);
+        let child_seeds = child_specs
+            .iter()
+            .map(|(child_name, child_range, child_optional)| {
+                let mut child_wanted = WantedDependency {
+                    alias: Some(child_name.clone()),
+                    bare_specifier: Some(child_range.clone()),
+                    optional: Some(*child_optional),
+                    ..WantedDependency::default()
+                };
+                let mut child_prior = prior_children_snapshot
+                    .and_then(|snapshot| prior_child_key(snapshot, child_name, child_range));
+                // Stale-pin refresh: force the edge onto a higher in-range
+                // direct-dep version instead of reusing the pin, so the
+                // pinned version is never resolved or fetched.
+                let forced_version = child_prior
+                    .as_ref()
+                    .and_then(|key| key.suffix.version_semver().cloned())
+                    .zip(child_range.parse::<node_semver::Range>().ok())
+                    .and_then(|(pinned, range)| {
+                        higher_direct_dep_version(
+                            direct_versions.as_deref(),
+                            child_name,
+                            &pinned,
+                            &range,
+                        )
+                    });
+                if let Some(higher) = forced_version {
+                    child_wanted.bare_specifier = Some(higher.to_string());
+                    child_prior = None;
+                }
+                let next_ancestors = Arc::clone(&next_ancestors);
+                let pick_overlay = children_overlay.clone();
+                let declaring_dir = declaring_dir.clone();
+                let parent_pkg_aliases = &children_pkg_aliases;
+                async move {
+                    let seed = resolve_node_seed(
+                        ctx,
+                        resolver,
+                        child_wanted,
+                        &next_ancestors,
+                        depth + 1,
+                        current_is_optional,
+                        ReuseSource::Transitive { key: child_prior },
+                        pick_overlay,
+                        declaring_dir.as_deref(),
+                        parent_pkg_aliases,
+                    )
+                    .await?;
+                    warm_children_resolutions(ctx, resolver, &seed).await;
+                    Ok::<NodeSeed, ResolveDependencyTreeError>(seed)
+                }
+            })
+            .pipe(future::try_join_all)
+            .await?;
+        let grandchild_overlay = PreferredVersionsOverlay::layer(
+            children_overlay.clone(),
+            level_versions(ctx, &child_seeds),
+        );
+        let grandchild_pkg_aliases = children_pkg_aliases.extend(level_aliases(&child_seeds));
+        // Phase 2: walk each child's own children with the extended
+        // overlay.
+        let child_results = child_seeds
+            .into_iter()
+            .map(|seed| {
+                let overlay = grandchild_overlay.clone();
+                let pkg_aliases = Arc::clone(&grandchild_pkg_aliases);
+                async move {
+                    match seed {
+                        NodeSeed::Done(dep) => Ok(dep),
+                        NodeSeed::Pending(pending) => {
+                            walk_node_children(ctx, resolver, *pending, overlay, pkg_aliases).await
+                        }
+                    }
+                }
+            })
+            .pipe(future::try_join_all)
+            .await?;
+        if is_current_children_owner(ctx, &id, &children_owner.owner) {
+            // Build the realized `(alias → NodeId)` map for THIS
+            // occurrence and the per-pkg `children_by_id` entry future
+            // revisits will reuse. `children_by_id` records the resolved
+            // child pkg ids (not NodeIds) plus the `optional` flag so
+            // lazy realisation can thread `current_is_optional` correctly.
+            let mut realized: BTreeMap<String, NodeId> = BTreeMap::new();
+            let mut by_id: Vec<crate::resolved_tree::ChildEdge> = Vec::new();
+            let optional_by_alias: HashMap<&str, bool> =
+                child_specs.iter().map(|(name, _, optional)| (name.as_str(), *optional)).collect();
+            for dep in child_results.into_iter().flatten() {
+                let optional = optional_by_alias.get(dep.alias.as_str()).copied().unwrap_or(false);
+                by_id.push(crate::resolved_tree::ChildEdge {
+                    alias: dep.alias.clone(),
+                    pkg_id: dep.id.clone(),
+                    optional,
+                });
+                realized.insert(dep.alias, dep.node_id);
+            }
+            lock_recoverable(&ctx.workspace.children_by_id).insert(id.clone(), Arc::new(by_id));
+            crate::resolved_tree::TreeChildren::Realized(realized)
+        } else {
+            crate::resolved_tree::TreeChildren::Lazy {
+                parent_ids: AncestorIds::from(Arc::clone(&parent_ancestors)),
+            }
+        }
+    };
+
+    // Repeat-visit leaves collapse onto one tree node; keep the
+    // shallowest depth seen so downstream consumers that read
+    // `tree_node.depth` (the peer pass folds it onto the graph node's
+    // `depth`) take the minimum across visits.
+    // Per-occurrence counter ids are unique by construction, so the
+    // `and_modify` arm is dead for non-leaves.
+    // Linked nodes carry `depth = -1` so the peer-resolution pass
+    // short-circuits them in `resolve_node`.
+    let node_depth = if is_link { -1 } else { depth };
+    remember_node_parent_ids(ctx, &node_id, parent_ancestors);
+    insert_tree_node(ctx, node_id.clone(), &id, children, node_depth);
+    if children_owner.owns_children
+        && !children_owner.children_context_unchanged
+        && is_current_children_owner(ctx, &id, &children_owner.owner)
+    {
+        make_non_owner_nodes_lazy(ctx, &id, &node_id);
+    }
+
+    Ok(Some(DirectDep { alias, node_id, id }))
+}
+
+/// Whether the `parent → child` edge closes a dependency cycle's
+/// *second* lap. The first re-entry of a cycle is kept (so the
+/// cycle-closing dependency edge appears in the tree and the lockfile
+/// snapshot); only the repeat of the full `parent … child` sequence is
+/// dropped.
+pub(crate) fn parent_ids_contain_sequence(
+    pkg_ids: &[String],
+    pkg_id1: &str,
+    pkg_id2: &str,
+) -> bool {
+    let Some(pkg1_index) = pkg_ids.iter().position(|id| id == pkg_id1) else {
+        return false;
+    };
+    if pkg1_index == pkg_ids.len() - 1 {
+        return false;
+    }
+    let Some(pkg2_index) = pkg_ids.iter().rposition(|id| id == pkg_id2) else {
+        return false;
+    };
+    pkg1_index < pkg2_index && pkg2_index != pkg_ids.len() - 1
+}
+
+/// Whether a freshly resolved node landed back on its previously
+/// recorded lockfile entry — the non-`updated` arm, which keeps the
+/// prior child refs alive.
+fn landed_on_prior_entry(prior_key: &PkgNameVerPeer, resolved_pkg_id: &str) -> bool {
+    prior_key.without_peer().to_string() == pacquet_deps_path::remove_suffix(resolved_pkg_id)
+}
+
+/// The package names the npm picker may consult the preferred-versions
+/// overlay under for one wanted edge: the alias itself, plus the real
+/// package name from [`real_package_name_of`] when it differs (the
+/// inner target of an `npm:` alias or the folded `@jsr/...` name of a
+/// `jsr:` specifier) — mirroring the name derivation in the npm
+/// resolver's `parse_bare_specifier`, which keys its overlay merge by
+/// the resolved `spec.name` rather than the outer alias.
+fn overlay_lookup_names(wanted: &WantedDependency) -> Vec<String> {
+    let mut names: Vec<String> = Vec::new();
+    if let Some(alias) = wanted.alias.as_deref()
+        && !alias.is_empty()
+    {
+        names.push(alias.to_string());
+    }
+    if let Some(real_name) = real_package_name_of(wanted)
+        && !names.iter().any(|name| name == real_name.as_ref())
+    {
+        names.push(real_name.into_owned());
+    }
+    names
+}
+
+/// Look the wanted edge up in the per-wanted dedup cache or run the
+/// resolver chain and the manifest-hook pipeline, caching the
+/// `Arc<ResolveResult>` under `cache_key`. Concurrent first-callers
+/// can both miss and resolve in parallel — the resolver's own
+/// per-cache-key fetch locker coalesces the network work, and the
+/// second `or_insert` loses the race harmlessly.
+async fn resolve_wanted_cached<Chain>(
+    ctx: &TreeCtx,
+    resolver: &Chain,
+    wanted: &WantedDependency,
+    opts: &ResolveOptions,
+    pick_overlay: Option<&Arc<PreferredVersionsOverlay>>,
+    cache_key: WantedKey,
+) -> Result<Arc<pacquet_resolving_resolver_base::ResolveResult>, ResolveDependencyTreeError>
+where
+    Chain: Resolver + ?Sized,
+{
+    let cached =
+        lock_recoverable(&ctx.workspace.resolved_by_wanted).get(&cache_key).map(Arc::clone);
+    if let Some(result) = cached {
+        return Ok(result);
+    }
+    // Combine two per-package opts adjustments into one clone. The
+    // `update_requested` flag is scoped per wanted-dependency — true only
+    // when the package's real name (parsed from `bare_specifier` for
+    // npm-aliases, folded from the jsr specifier for jsr deps) is in the
+    // update target list — so the picker's held-back-update warning fires
+    // only for the packages the user actually asked to update.
+    let needs_overlay = !cache_key.8.is_empty();
+    let update_target = cache_key.10;
+    let needs_update = update_target != opts.update_requested;
+    let owned_opts;
+    let opts = if needs_overlay || needs_update {
+        let mut owned = opts.clone();
+        if needs_overlay {
+            owned.preferred_versions_overlay = pick_overlay.map(Arc::clone);
+        }
+        if needs_update {
+            owned.update_requested = update_target;
+        }
+        owned_opts = owned;
+        &owned_opts
+    } else {
+        opts
+    };
+    let mut result = resolver.resolve(wanted, opts).await.map_err(map_resolve_error)?;
+    let Some(result_inner) = result.as_mut() else {
+        return Err(ResolveDependencyTreeError::SpecNotSupported {
+            specifier: render_specifier(wanted),
+        });
+    };
+    if result_inner.manifest.is_none() {
+        result_inner.manifest =
+            Some(Arc::new(fallback_manifest(wanted, opts.current_pkg.as_ref())));
+    }
+    // Apply the configured `readPackageHook` (today:
+    // `packageExtensions`) to the manifest fragment before
+    // anything downstream sees it. The hook clones the inner `Value`
+    // only when it modifies it, so unrelated manifests keep sharing the
+    // resolver's cached `Arc`.
+    if let Some(hook) = ctx.workspace.manifest_hook.as_ref()
+        && let Some(manifest) = result_inner.manifest.take()
+    {
+        result_inner.manifest = Some(hook(manifest));
+    }
+
+    if let Some(pnpmfile_hook) = ctx.workspace.pnpmfile_hook.as_ref()
+        && let Some(manifest) = result_inner.manifest.take()
+    {
+        let log = ctx.workspace.read_package_log.clone().unwrap_or_else(|| Arc::new(|_| {}));
+        // Directory resolutions carry their lockfile-root-relative dir so the
+        // hook can tell a workspace project's dependency instance apart from a
+        // registry manifest — see `HookContext::dir`.
+        let dir = match &result_inner.resolution {
+            pacquet_lockfile::LockfileResolution::Directory(directory_resolution) => {
+                Some(directory_resolution.directory.clone())
+            }
+            _ => None,
+        };
+        let hook_ctx = pacquet_hooks::HookContext { log, dir };
+
+        let updated = pnpmfile_hook
+            .read_package((*manifest).clone(), hook_ctx)
+            .await
+            .map_err(ResolveDependencyTreeError::PnpmfileHook)?;
+        result_inner.manifest = Some(updated);
+    }
+
+    // Overrides run last so a pnpmfile hook that replaced the manifest
+    // cannot erase them — see `WorkspaceTreeCtx::overrides_hook`.
+    if let Some(hook) = ctx.workspace.overrides_hook.as_ref()
+        && let Some(manifest) = result_inner.manifest.take()
+    {
+        result_inner.manifest = Some(hook(manifest));
+    }
+
+    let result = result.expect("Some-guarded above");
+    // Wrap in `Arc` once so the cache, the per-id
+    // `ResolvedPackage` envelope, and the later peer-resolved
+    // graph node share one heap-allocated `ResolveResult`
+    // instead of cloning every `String` field per occurrence.
+    let result = Arc::new(result);
+    lock_recoverable(&ctx.workspace.resolved_by_wanted)
+        .entry(cache_key)
+        .or_insert_with(|| Arc::clone(&result));
+    Ok(result)
+}
+
+/// Stand in for the manifest a resolver didn't supply (pnpm's
+/// `getManifestFromResponse`).
+///
+/// Every consumer downstream of the resolver chain reads the package's
+/// identity off its manifest — most of all
+/// [`build_pkg_id_with_patch_hash`], which has no name to prefix the dep
+/// path with when there is none, leaving a bare `file:<path>` / URL that
+/// keys no `packages:` row. A package with no `package.json` of its own
+/// (an archive that ships none, a directory pin restored from the
+/// lockfile) still has to install, so it borrows the identity the
+/// lockfile recorded for it, or failing that its alias and `0.0.0` — the
+/// version pnpm writes into `packages:` for such a package.
+fn fallback_manifest(
+    wanted: &WantedDependency,
+    current_pkg: Option<&CurrentPkg>,
+) -> pacquet_resolving_resolver_base::DependencyManifest {
+    if let Some(current) = current_pkg
+        && let Some(name) = current.name.as_deref().filter(|name| !name.is_empty())
+        && let Some(version) = current.version.as_deref().filter(|version| !version.is_empty())
+    {
+        return serde_json::json!({ "name": name, "version": version });
+    }
+    let name = match wanted.alias.as_deref().filter(|alias| !alias.is_empty()) {
+        Some(alias) => alias,
+        // A specifier's last path segment is the closest thing to a name
+        // an unaliased dep carries: `file:./no-manifest-1.0.0.tgz` and
+        // `https://host/no-manifest-1.0.0.tgz` both name the archive.
+        None => wanted
+            .bare_specifier
+            .as_deref()
+            .unwrap_or_default()
+            .rsplit('/')
+            .next()
+            .unwrap_or_default(),
+    };
+    serde_json::json!({ "name": name, "version": "0.0.0" })
+}
+
+/// Wrap a resolver-chain failure, keeping the pnpm error code of the ones
+/// that carry one. The chain hands back a type-erased
+/// [`ResolveError`], which drops the `miette::Diagnostic` facet, so the codes
+/// that are part of pnpm's public contract are recovered by downcast; every
+/// other failure keeps the generic envelope.
+fn map_resolve_error(err: ResolveError) -> ResolveDependencyTreeError {
+    let err = match err.downcast::<NoMatchingVersionError>() {
+        Ok(no_matching_version) => {
+            return ResolveDependencyTreeError::NoMatchingVersion(*no_matching_version);
+        }
+        Err(err) => err,
+    };
+    match err.downcast::<RegistryResponseError>() {
+        Ok(response) => ResolveDependencyTreeError::RegistryResponse(*response),
+        Err(err) => ResolveDependencyTreeError::Resolve(err.to_string()),
+    }
+}
+
+/// Speculatively warm a freshly-seeded node's children resolutions so
+/// their packuments download while the sibling level's barrier waits
+/// for its slowest member. Results are discarded — the real picks run
+/// in the walk phase with the level's preferred-versions overlay and
+/// hit the warm metadata caches — and errors are swallowed: a
+/// speculative fetch must never fail the install (the real resolve
+/// will surface it). Recovers the cross-level pipelining the
+/// postponed-resolution barrier otherwise serializes; pure overlap,
+/// no behavioral effect.
+pub(super) async fn warm_children_resolutions<Chain>(
+    ctx: &TreeCtx,
+    resolver: &Chain,
+    seed: &NodeSeed,
+) where
+    Chain: Resolver + ?Sized,
+{
+    // A configured pnpmfile hook is externally observable per call
+    // (`readPackage` IPC, `context.log`, custom resolvers), so
+    // speculative resolutions must not fire it; the pure in-memory
+    // manifest hook (packageExtensions / overrides) is idempotent and
+    // cache-deduped, indistinguishable from a first-caller win in the
+    // pre-existing concurrent-miss race.
+    if ctx.workspace.pnpmfile_hook.is_some() {
+        return;
+    }
+    let NodeSeed::Pending(pending) = seed else { return };
+    if pending.is_link || !pending.children_owner.owns_children {
+        return;
+    }
+    let Ok(specs) = extract_children(&pending.result) else { return };
+    let opts = ctx.opts_for_depth(pending.depth + 1);
+    let declaring_dir = declaring_manifest_dir(ctx, &pending.result);
+    specs
+        .iter()
+        .filter(|(name, _, optional)| {
+            *optional || !pending.children_owner.peer_shadowed.contains(name)
+        })
+        .map(|(name, range, optional)| {
+            let wanted = WantedDependency {
+                alias: Some(name.clone()),
+                bare_specifier: Some(range.clone()),
+                optional: Some(*optional),
+                ..WantedDependency::default()
+            };
+            let opts = opts_relative_to_declaring_manifest(opts, &wanted, declaring_dir.as_deref());
+            async move {
+                let opts = opts.as_ref();
+                // Warm through the same per-wanted dedup cache, under
+                // the empty-overlay-view key: when the real pick's
+                // view is empty too (the overwhelmingly common case)
+                // it reuses this entry outright; otherwise it misses
+                // into its own bucket and re-picks from the warm
+                // metadata caches.
+                let project_scope = project_relative_cache_scope(&wanted, opts);
+                let cache_key: WantedKey = (
+                    wanted.alias.clone(),
+                    wanted.bare_specifier.clone(),
+                    wanted.optional,
+                    wanted.injected,
+                    opts.pick_lowest_version,
+                    opts.published_by,
+                    project_scope,
+                    // No prior-lockfile key: a warm entry must only be
+                    // reused by edges that carry no currentPkg either.
+                    None,
+                    Vec::new(),
+                    ctx.update_cache_scope(),
+                    is_update_target(ctx.update_scope(), &wanted, pending.depth + 1),
+                );
+                let _ = resolve_wanted_cached(ctx, resolver, &wanted, opts, None, cache_key).await;
+            }
+        })
+        .pipe(future::join_all)
+        .await;
+}
+
+/// The install aliases one resolved level contributes to its
+/// children's [`ParentPkgAliases`] scope. An edge the walk dropped (a
+/// cycle re-entry, a skipped optional) contributes nothing, matching
+/// pnpm's fold over the level's resolved addresses.
+pub(super) fn level_aliases(seeds: &[NodeSeed]) -> HashSet<String> {
+    seeds
+        .iter()
+        .filter_map(|seed| match seed {
+            NodeSeed::Pending(pending) => Some(pending.alias.clone()),
+            NodeSeed::Done(Some(dep)) => Some(dep.alias.clone()),
+            NodeSeed::Done(None) => None,
+        })
+        .collect()
+}
+
+/// The `(name → versions)` additions one resolved level contributes
+/// to its children's preferred-versions overlay. Linked nodes carry no
+/// `name_ver` and contribute nothing — they're skipped in the fold.
+pub(super) fn level_versions(ctx: &TreeCtx, seeds: &[NodeSeed]) -> BTreeMap<String, Vec<String>> {
+    let packages = lock_recoverable(&ctx.workspace.packages);
+    let mut level: BTreeMap<String, Vec<String>> = BTreeMap::new();
+    for seed in seeds {
+        let name_ver = match seed {
+            NodeSeed::Pending(pending) => pending.result.name_ver.as_ref(),
+            NodeSeed::Done(Some(dep)) => {
+                packages.get(&dep.id).and_then(|pkg| pkg.result.name_ver.as_ref())
+            }
+            NodeSeed::Done(None) => None,
+        };
+        let Some(name_ver) = name_ver else { continue };
+        let versions = level.entry(name_ver.name.to_string()).or_default();
+        let version = name_ver.suffix.to_string();
+        if !versions.contains(&version) {
+            versions.push(version);
+        }
+    }
+    level
+}
+
+/// Map an ancestor-id chain to the `parents` payload of a
+/// skipped-optional-dependency notification, resolving each
+/// `pkgIdWithPatchHash` through the shared packages map (the
+/// counterpart of pnpm's `getPkgsInfoFromIds`).
+fn pkgs_info_from_ids(
+    ctx: &TreeCtx,
+    ancestor_ids: &[String],
+) -> Vec<SkippedOptionalDependencyParent> {
+    let packages = lock_recoverable(&ctx.workspace.packages);
+    ancestor_ids
+        .iter()
+        .map(|id| {
+            let name_ver = packages.get(id).and_then(|pkg| pkg.result.name_ver.as_ref());
+            SkippedOptionalDependencyParent {
+                id: id.clone(),
+                name: name_ver.map(|name_ver| name_ver.name.to_string()).unwrap_or_default(),
+                version: name_ver.map(|name_ver| name_ver.suffix.to_string()).unwrap_or_default(),
+            }
+        })
+        .collect()
+}
+
+/// Render `{alias}@{bare}` (either half dropped when absent) for the
+/// no-resolver error message.
+fn render_specifier(wanted: &WantedDependency) -> String {
+    let alias = wanted.alias.as_deref().unwrap_or("");
+    let bare = wanted.bare_specifier.as_deref().unwrap_or("");
+    match (alias.is_empty(), bare.is_empty()) {
+        (true, true) => String::new(),
+        (true, false) => bare.to_string(),
+        (false, true) => alias.to_string(),
+        (false, false) => format!("{alias}@{bare}"),
+    }
+}
+
+#[cfg(test)]
+mod tests;
